% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{addNoPixel2CohortData}
\alias{addNoPixel2CohortData}
\title{Add number of pixels per \code{pixelGroup} and add it has a new column to \code{cohortData}}
\usage{
addNoPixel2CohortData(cohortData, pixelGroupMap,
  doAssertion = getOption("LandR.assertions", TRUE))
}
\arguments{
\item{cohortData}{A \code{data.table} with columns:
\code{pixelGroup}, \code{ecoregionGroup}, \code{speciesCode}, \code{age},
\code{B}, \code{mortality}, \code{aNPPAct}, ond \code{sumB}.}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group number
that corresponds exactly to ]\code{pixelGroup} column in \code{cohortData}.}

\item{doAssertion}{Turns on/off assertion. Defaults to \code{getOption("LandR.assertions")}}
}
\value{
An \code{cohortData} \code{dat.table} with a new \code{noPixels}
column
}
\description{
Add number of pixels per \code{pixelGroup} and add it has a new column to \code{cohortData}
}
