% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{adjustAgeToLongevity2}
\alias{adjustAgeToLongevity2}
\title{Reduces the age of cohorts that exceed their longevity x adjustmentFactort}
\usage{
adjustAgeToLongevity2(pixelCohortData, longevity, adjustment)
}
\arguments{
\item{pixelCohortData}{A \code{cohortData} object}

\item{longevity}{A data.table with the longevity of each species.}

\item{adjustmentFactor}{A numeric controlling the proportion of species longevity
that cohort ages cannot exceed.}
}
\value{
A \code{cohortData} object with corrected ages.
}
\description{
Reduces the age of cohorts that exceed their longevity x adjustmentFactort
}
