% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{ageAdjust}
\alias{ageAdjust}
\title{Applies the smoothed age correction for a single species.}
\usage{
ageAdjust(age, adjustmentFactor, longevity, decayRange = 96)
}
\arguments{
\item{age}{input age values}

\item{adjustmentFactor}{A numeric controlling the proportion of species longevity
that cohort ages cannot exceed.}

\item{longevity}{A data.table with the longevity of each species.}

\item{decayRange}{range along the x-axis where \link{decayFunction} is applied}
}
\description{
Applies the smoothed age correction for a single species.
}
\keyword{internal}
