% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{aggregateRasByDT}
\alias{aggregateRasByDT}
\title{Aggregate a raster}
\usage{
aggregateRasByDT(ras, newRas, fn = sum)
}
\arguments{
\item{ras}{\code{RasterLayer} to aggregate}

\item{newRas}{\code{RasterLayer} to match}

\item{fn}{function to use to aggregate pixel values}
}
\value{
\code{RasterLayer}
}
\description{
Uses \pkg{data.table} to perform aggregation calculations, which is faster than
\code{raster::aggregate}.
}
