% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{asInteger}
\alias{asInteger}
\title{as.integer(round(x, 0))}
\usage{
asInteger(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer vector of length \code{x}, rounded to zero decimal places
  prior to \code{as.integer}
}
\description{
Simply a wrapper so that as.integer actually rounds, rather than trunc
}
