% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertCohortData}
\alias{assertCohortData}
\title{A test that \code{pixelGroupMap} and \code{cohortData} match}
\usage{
assertCohortData(cohortData, pixelGroupMap, sim,
  maxExpectedNumDiverge = 1, message = "",
  verbose = getOption("LandR.verbose", TRUE))
}
\arguments{
\item{cohortData}{a \code{data.table} with columns: \code{pixelGroup},
\code{ecoregionGroup}, \code{speciesCode}, \code{age}, \code{B},
\code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group
number that correspondsd exactly to \code{pixelGroup} column in
\code{cohortData}}

\item{sim}{If the simList is included, then the browser() call will be more useful}

\item{maxExpectedNumDiverge}{A numeric, length 1, indicating by how many they
can diverge. Default 1.}

\item{message}{An optional message to print. This may help identify where this function
was called.}

\item{verbose}{Integer, where increasing number is increasing verbosity. Currently,
only level 1 exists; but this may change.}
}
\description{
A test that \code{pixelGroupMap} and \code{cohortData} match
}
\note{
TODO
}
