% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertCohortDataAttr}
\alias{assertCohortDataAttr}
\title{Assert that \code{cohortData} has attribute \code{imputedPixID}}
\usage{
assertCohortDataAttr(
  cohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Assert that \code{cohortData} has attribute \code{imputedPixID}
}
