% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertPixelCohortData}
\alias{assertPixelCohortData}
\title{A test that \code{pixelGroupMap} and \code{pixelCohortData} match pixelIndex}
\usage{
assertPixelCohortData(pixelCohortData, pixelGroupMap)
}
\arguments{
\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group
number that correspondsd exactly to \code{pixelGroup} column in
\code{cohortData}}

\item{cohortData}{The full \code{cohortData} \code{data.table}}
}
\description{
This is the full pixelCohortData, not the collapsed one
}
