% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertRepsAllCohortData}
\alias{assertRepsAllCohortData}
\title{Assert that the \code{allCohortData} has the expected years and reps combinations}
\usage{
assertRepsAllCohortData(
  allCohortData,
  reps,
  years,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{allCohortData}{A \code{data.table} with all simulated \code{cohortData} to use for
validation}

\item{reps}{repetition ids}

\item{years}{years}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Assert that the \code{allCohortData} has the expected years and reps combinations
}
