% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertSppVectors}
\alias{assertSppVectors}
\title{Assert that vectors of species match}
\usage{
assertSppVectors(
  sppEquiv = NULL,
  sppNameVector = NULL,
  sppColorVect = NULL,
  sppEquivCol = NULL,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{sppEquiv}{an table with a column containing species names}

\item{sppNameVector}{A character vector of species to use. These species must all
be from one naming convention, i.e., from one column in the \code{sppEquiv}.}

\item{sppColorVect}{A named vector of colours to use for plotting.
The names must conform with species name convention used
(see LandR::sppEquivalencies_CA for possible naming
conventions) and should also contain a colour for 'Mixed',
when a mixed forest type is supported  (see vegLeadingProportion
parameter in LandR::vegTypeGenerator for details on mixed
fores types).}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
This assertion compares the species listed in the \code{sppEquiv}
table (under the \code{sppEquivCol} column), with species listed
in the \code{sppNameVector} and \code{sppColorVect}.
The comparison excludes the "Mixed" type, in case it exists.
}
