% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertStandAgeMapAttr}
\alias{assertStandAgeMapAttr}
\title{Assert that \code{standAgeMap} is a \code{RasterLayer} with attribute \code{imputedPixID}}
\usage{
assertStandAgeMapAttr(
  standAgeMap,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{standAgeMap}{A \code{rasterLayer} with attribute "imputedPixID"}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Assert that \code{standAgeMap} is a \code{RasterLayer} with attribute \code{imputedPixID}
}
