% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertStandCohortData}
\alias{assertStandCohortData}
\title{Assert that the \code{standCohortData} has no NAs}
\usage{
assertStandCohortData(
  standCohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{standCohortData}{A \code{data.table} with simulated and observed stand data for validation}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Assert that the \code{standCohortData} has no NAs
}
