% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{assignClimateEffect}
\alias{assignClimateEffect}
\title{assignClimateEffect}
\usage{
assignClimateEffect(...)
}
\arguments{
\item{...}{additional arguments that are passed to LandR.CS}
}
\value{
null estimate of growth due to climate
}
\description{
assignClimateEffect
}
