% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assignLightProb}
\alias{assignLightProb}
\title{Assign light probability}
\usage{
assignLightProb(sufficientLight, newCohortData)
}
\arguments{
\item{sufficientLight}{a data.table containing probability of establishment, given a site's light conditions (X0-X5) for
each level of a species shade tolerance (1-5)}

\item{newCohortData}{a modified version of /code{cohortData} that contains new cohorts.}
}
\value{
/code{newCohortData} with a lightProb column
}
\description{
Assign light probability
}
