% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.basename}
\alias{.basename}
\title{Basename}
\usage{
.basename(x)
}
\description{
Imports the non-exported function \code{SpaDES.core:::.basename}, which is a
simple wrapper around \code{\link{basename}} that deals with \code{NULL} values.
}
\keyword{internal}
