% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodWrappers.R
\name{biomodEnsembleFrcstWrapper}
\alias{biomodEnsembleFrcstWrapper}
\title{Simple wrapper around \code{\link[biomod2:BIOMOD_EnsembleForecasting]{biomod2::BIOMOD_EnsembleForecasting()}}}
\usage{
biomodEnsembleFrcstWrapper(
  bm.em,
  bm.proj = NULL,
  proj.name = NULL,
  new.env = NULL,
  new.env.xy = NULL,
  keep.in.memory = TRUE,
  ...
)
}
\arguments{
\item{bm.em}{output of \link[biomod2:BIOMOD_EnsembleModeling]{biomod2::BIOMOD_EnsembleModeling}}

\item{bm.proj}{output of \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}}

\item{proj.name}{passed to \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}}

\item{new.env}{passed to \link[biomod2:BIOMOD_EnsembleForecasting]{biomod2::BIOMOD_EnsembleForecasting}}

\item{new.env.xy}{passed to \link[biomod2:BIOMOD_EnsembleForecasting]{biomod2::BIOMOD_EnsembleForecasting}}

\item{keep.in.memory}{passed to \link[biomod2:BIOMOD_EnsembleForecasting]{biomod2::BIOMOD_EnsembleForecasting}}

\item{...}{passed to \link[reproducible:Cache]{reproducible::Cache}}
}
\description{
Simple wrapper around \code{\link[biomod2:BIOMOD_EnsembleForecasting]{biomod2::BIOMOD_EnsembleForecasting()}}
}
