% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodWrappers.R
\name{biomodEnsembleWrapper}
\alias{biomodEnsembleWrapper}
\title{BIOMOD_EnsembleModeling wrapper}
\usage{
biomodEnsembleWrapper(bm.mod, metric.select.thresh = NULL, ...)
}
\arguments{
\item{bm.mod}{output of \code{biomod2::BIOMOD_Modeling}}

\item{metric.select.thresh}{passed \code{biomod2::BIOMOD_EnsembleModeling}.
By default no thresholding is applied.}

\item{...}{passed to Cache}
}
\description{
BIOMOD_EnsembleModeling wrapper
}
