% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodWrappers.R
\name{biomodModelingWrapper}
\alias{biomodModelingWrapper}
\title{BIOMOD_Modeling wrapper}
\usage{
biomodModelingWrapper(
  sp,
  responseVar,
  responseVarData,
  predictorVars,
  predictorVarsData,
  dir.name,
  BIOMOD_ModelingArgs = list(models = c("GLM", "MARS"), bm.options = NULL, CV.k = 5,
    CV.perc = 100, metric.eval = c("TSS", "ROC"), modeling.id = "test"),
  ...
)
}
\arguments{
\item{sp}{character of species name to subset \code{responseVarData} table}

\item{responseVar}{character vector, or list of character vectors, of response
variable (column) to subset \code{responseVarData} table. If a list, it should
be named according to \code{sp}, for subsetting.}

\item{responseVarData}{a data.table or list of data.tables with environmental data.
If a list, it should be named according to \code{sp}, for subsetting.#'}

\item{predictorVars}{character vector, or list of character vectors, of environmental
variables (columns) to subset \code{predictorVarsData} table. If a list, it should
be named according to \code{sp}, for subsetting.}

\item{predictorVarsData}{environmental data.}

\item{dir.name}{passed to \link[biomod2:BIOMOD_FormatingData]{biomod2::BIOMOD_FormatingData}}

\item{BIOMOD_ModelingArgs}{a named list of arguments passed to \link[biomod2:BIOMOD_Modeling]{biomod2::BIOMOD_Modeling}}

\item{...}{further arguments passed to \link[reproducible:Cache]{reproducible::Cache}}
}
\description{
BIOMOD_Modeling wrapper
}
