% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosim.R
\name{BioSIM_extractPoints}
\alias{BioSIM_extractPoints}
\title{Extract point locations from DEM raster to pass to \code{BioSIM} functions}
\usage{
BioSIM_extractPoints(x)
}
\arguments{
\item{x}{A digital elevation model (DEM) \code{RasterLayer}.}
}
\value{
\code{data.table} with columns \code{Name}, \code{Long}, \code{Lat}, \code{Elev}.
}
\description{
Extract point locations from DEM raster to pass to \code{BioSIM} functions
}
\seealso{
prepInputsCanDEM
}
