% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosim.R
\name{BioSIM_getMPBSLR}
\alias{BioSIM_getMPBSLR}
\title{Get annual historic and projected MPB climate suitability maps from \code{BioSIM}}
\usage{
BioSIM_getMPBSLR(dem, years, SLR = "R", climModel = "GCM4", rcp = "RCP45")
}
\arguments{
\item{dem}{\code{RasterLayer} of elevation data (m).}

\item{years}{numeric vector corresponding to the years to retrieve.}

\item{SLR}{character. Specifies which climate suitability index to extract.
Currently, one of \code{"S"}, \code{"L"}, \code{"R"}, or \code{"G"},
corresponding to Safranyik-P3P4, Logan-2b, Régnière Cold Tolerance Survival, or
their Geometric product (\code{S*L*R}), respectively.}

\item{climModel}{climate model to use. one of \code{"GCM4"} or \code{"RCM4"}.}

\item{rcp}{RCP scenario to use. one of \code{"RCP45"} or \code{"RCP85"}.}
}
\value{
\code{RasterStack}
}
\description{
Raster stacks for all 9 MPB climate indices. See \code{BioSIM::getModelHelp("MPB_SLR")}.
}
\note{
Although the BioSIM MPB_SLR model provides several other indices
(see \code{BioSIM::getModelHelp("MPB_SLR")}), only 4 are currently used here.
}
