% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosim.R
\name{BioSIM_getWind}
\alias{BioSIM_getWind}
\title{Get annual historic and projected wind maps from \code{BioSIM}}
\usage{
BioSIM_getWind(dem, years, climModel = "GCM4", rcp = "RCP45")
}
\arguments{
\item{dem}{\code{RasterLayer} of elevation data (m).}

\item{years}{numeric vector corresponding to the years to retrieve.}

\item{climModel}{climate model to use. one of \code{"GCM4"} or \code{"RCM4"}.}

\item{rcp}{RCP scenario to use. one of \code{"RCP45"} or \code{"RCP85"}.}
}
\value{
\code{RasterStack}
}
\description{
Get annual historic and projected wind maps from \code{BioSIM}
}
