% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildVRT.R
\name{buildVRT}
\alias{buildVRT}
\title{Check if a SpatRaster should be converted to tiled virtual raster}
\usage{
buildVRT(
  ras,
  bitExponent = 30,
  destinationPath = NULL,
  writeTo = NULL,
  overwrite = NULL
)
}
\arguments{
\item{ras}{a file-backed raster}

\item{bitExponent}{2 exponent this number sets a threshold whereby rasters
containing more cells than the threshold are converted to virtual rasters via tiles}

\item{destinationPath}{the output directory for raster, if written}

\item{writeTo}{filename - extension will be modified to .vrt if tiled}

\item{overwrite}{overte files}
}
\value{
the raster as a VRT
}
\description{
Check if a SpatRaster should be converted to tiled virtual raster
}
