% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{calcSeverityB}
\alias{calcSeverityB}
\title{Calculate fire severity}
\usage{
calcSeverityB(cohortData, burnedPixelCohortData)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{burnedPixelCohortData}{An expanded \code{cohortData} \code{data.table} with pixel-level
cohort information on burnt pixels and columns:
\code{severity} - fire severity in that pixel calculated based on fire behaviour properties;
\code{firetolerance} - species-level fire tolerance;
\code{severityToleranceDif} - the difference between \code{severity} and \code{firetolerance}.}
}
\value{
\code{data.table} with columns \code{pixelIndex},
  \code{pixelGroup} and  \code{severityB}
}
\description{
Calculates fire severity as the loss of pre-fire to
  post-fire biomass.
}
