% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateGrowth}
\alias{calculateClimateGrowth}
\title{calculateClimateGrowth}
\usage{
calculateClimateGrowth(...)
}
\value{
null estimate of growth due to climate
}
\description{
calculateClimateGrowth
}
