% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateMortality}
\alias{calculateClimateMortality}
\title{calculateClimateMortality}
\usage{
calculateClimateMortality(...)
}
\value{
null estimate of mortality due to climate
}
\description{
calculateClimateMortality
}
