% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{CASFRItoSpRasts}
\alias{CASFRItoSpRasts}
\title{\code{CASFRItoSpRasts}}
\usage{
CASFRItoSpRasts(
  CASFRIRas,
  CASFRIattrLong,
  CASFRIdt,
  sppEquiv,
  sppEquivCol,
  destinationPath
)
}
\arguments{
\item{CASFRIRas}{\code{RasterLayer} object}

\item{CASFRIattrLong}{data.table of CASFRI attributes (long format) (i.e., from \link{loadCASFRI})}

\item{CASFRIdt}{data.table of CASFRI raster values (i.e., from \link{loadCASFRI})}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}
}
\value{
\code{RasterStack} (or equivalent)
}
\description{
Extract CASFRI data.table values and create species layer rasters
}
