% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{checkCloud}
\alias{checkCloud}
\title{Basic tool for using cloud-based caching}
\usage{
checkCloud(toDigest, checksumsFileID, writeChecksumsFolderID = NULL)
}
\arguments{
\item{toDigest}{The R object to consider, e.g., all the arguments to a function}

\item{checksumsFileID}{A google file ID where the checksums data.table is located,
provided as a character string}

\item{writeChecksumsFolderID}{The google folder ID where a new checksums file should
 be written. This will only be used if \code{checksumsFileID} is not provided
provided as a character string}
}
\description{
Very experimental
}
\examples{
\dontrun{

}
}
