% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\docType{data}
\name{columnsForPixelGroups}
\alias{columnsForPixelGroups}
\title{Default columns that define pixel groups}
\format{An object of class \code{character} of length 4.}
\usage{
columnsForPixelGroups
}
\description{
Default columns that define pixel groups
}
\keyword{datasets}
