% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnwantedLCC2.R
\name{convertUnwantedLCC2}
\alias{convertUnwantedLCC2}
\title{Convert Land Cover Classes (LCC) to another value in its neighbourhood}
\usage{
convertUnwantedLCC2(
  classesToReplace = 34:36,
  rstLCC,
  nIterations = 6,
  defaultNewValue = NA,
  invalidClasses = NA
)
}
\arguments{
\item{classesToReplace}{Integer vector of classes that are are to be replaced,
e.g., 34, 35, 36 on LCC2005, which are burned young, burned 10 year, and cities.}

\item{rstLCC}{LCC raster, e.g., LCC2005}

\item{nIterations}{the number of iterations to perform}

\item{defaultNewValue}{the value to assign a pixel in \code{classesToReplace} if no valid  pixel
is closer after \code{nIterations}}

\item{invalidClasses}{classes that are not valid options}
}
\value{
A \code{rasterLayer} with values in classesToReplace converted to adjacent values or NA.
}
\description{
This will search around the pixels on \code{rstLCC} that have
\code{classesToReplace}, and search in iteratively increasing
radii outwards for other Land Cover Classes than the those indicated in
\code{classesToReplace}.
.
}
\author{
Eliot McIntire Ian Eddy
}
