% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{createPrjFile}
\alias{createPrjFile}
\title{Create a \code{.prj} file}
\usage{
createPrjFile(shpFile,
  urlForProj = "http://spatialreference.org/ref/epsg/nad83-utm-zone-11n/prj/")
}
\arguments{
\item{shpFile}{The filename of a shapefile to add \code{.prj}}

\item{urlForProj}{The url from which to fetch the projection, e.g.,
\code{"http://spatialreference.org/ref/epsg/nad83-utm-zone-11n/prj/"}.}
}
\description{
In cases where a shapefile is missing its associated \code{.prj} file.
}
