% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{decayFunction}
\alias{decayFunction}
\title{Apply the decay function}
\usage{
decayFunction(x, Ymax, startAt = 120)
}
\arguments{
\item{x}{input value along the x-axis}

\item{Ymax}{maximum value on the y-axis, defining the asymptote}

\item{startAt}{value along the x-axis where the decay function is applied}
}
\description{
Exponential function starting at \code{startAt} and with an asymptote at \code{Ymax}
}
\keyword{internal}
