% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultData.R
\name{defaultClimateDataProj}
\alias{defaultClimateDataProj}
\title{Default future projections (forecasts) of climate variables}
\usage{
defaultClimateDataProj(
  vars = c("MAT", "PPT_wt", "PPT_sm", "CMI"),
  climateGCM = "ensemble_8GCMs",
  climateSSP = "585",
  periods = c("2001_2020", "2021_2040", "2041_2060", "2061_2080", "2081_2100"),
  studyArea,
  userTags,
  destinationPath,
  rasterToMatch
)
}
\arguments{
\item{vars}{character. Variables to download.}

\item{climateGCM}{character. The Global Circulation Model, GCM, (or ensemble) to use.}

\item{climateSSP}{character. The Socio-economic Shared Pathway, SSP (i.e. emissions scenario) to use.}

\item{periods}{character. The year periods to use.}

\item{studyArea}{a spatial polygon  (\code{sf}, \code{sp} or \code{SpatVector}) used to crop
and mask data layers. Passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{userTags}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{destinationPath}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{rasterToMatch}{a raster layer (\code{RasterLayer} or \code{SpatRaster}). Only used to \code{rasterize}
the default permafrost layer and fire perimeters.}
}
\value{
named list of \code{SpatRaster}s, with names following \code{vars}.
}
\description{
Downloads a set of default climate variables' forecasts as rasters -- used
to model \code{maxB}, SEP and thermokarst (among others).
}
\details{
By default this function downloads four bioclimatic variables (mean annual temperature
MAT, precipitation of warmest quarter, PPT_sm, precipitation of coldest quarter, PPT_wt, annual
climate moisture index, CMI) obtained from AdaptWest for 20 year periods between 2001 and 2100
(see https://adaptwest.databasin.org/pages/adaptwest-climatena/) - at 1 km resolution].
}
