% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultData.R
\name{defaultEnvirData}
\alias{defaultEnvirData}
\title{Default environmental variables}
\usage{
defaultEnvirData(
  vars = c("MAT", "PPT_wt", "PPT_sm", "CMI", "elevation", "timeSinceFire", "wetlands",
    "permafrost", "thermokarst", "thermXperm"),
  studyArea,
  userTags,
  destinationPath,
  rasterToMatch
)
}
\arguments{
\item{vars}{character. Variables to download.}

\item{studyArea}{a spatial polygon  (\code{sf}, \code{sp} or \code{SpatVector}) used to crop
and mask data layers. Passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{userTags}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{destinationPath}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{rasterToMatch}{a raster layer (\code{RasterLayer} or \code{SpatRaster}). Only used to \code{rasterize}
the default permafrost layer and fire perimeters.}
}
\value{
named list of \code{SpatRasters}, with names following \code{vars}.
}
\description{
Downloads a set of default environmental variable rasters to model
maxB, SEP and thermokarst (amongst others).
}
\details{
By default this function downloads four bioclimatic variables (mean annual temperature
MAT, precipitation of warmest quarter, PPT_sm, precipitation of coldest quarter, PPT_wt, annual
climate moisture index, CMI) obtained from AdaptWest for the reference period 1980-2010
(https://s3-us-west-2.amazonaws.com/www.cacpd.org/CMIP6/normals/Normal_1981_2010_bioclim.zip),
elevation (from https://adaptwest.databasin.org/pages/adaptwest-climatena/) - at 1 Km resolution -
a binary map of wetlands (from Wulder et al. 2018), permafrost peatland complex cover and amount
of thermokarst (from Gibson et al. 2021) -- which are combined to produce a compound variable called
'thermXperm' -- and a map of time since fire (considering fire perimeters from the Canadian Wildland
Fire Information System.
}
