% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{defineFlammable}
\alias{defineFlammable}
\title{Define flammability map}
\usage{
defineFlammable(LandCoverClassifiedMap = NULL, nonFlammClasses = c(36L,
  37L, 38L, 39L), mask = NULL, filename2 = NULL)
}
\arguments{
\item{LandCoverClassifiedMap}{A \code{Raster} that represents land cover
(e.g., Land Cover Classified map from 2005 or 2010 from the Canadian Forest Service).}

\item{nonFlammClasses}{numeric vector defining which classes in \code{LandCoverClassifiedMap}.}

\item{mask}{A raster to use as a mask (see \code{\link[raster]{mask}}).}

\item{filename2}{See \code{\link[reproducible]{postProcess}}. Default \code{NULL}.}
}
\description{
Define flammability map
}
