% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{describeCohortData}
\alias{describeCohortData}
\title{Summary for \code{cohortData}}
\usage{
describeCohortData(cohortData)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}
}
\description{
Summary for \code{cohortData}
}
