% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permafrost.R
\name{.assignPermafrostWrapper}
\alias{.assignPermafrostWrapper}
\title{Wrapper for \code{assignPermafrost}}
\usage{
.assignPermafrostWrapper(gridPoly, ras, id, saveDir, cores, outFilename, dPath)
}
\arguments{
\item{gridPoly}{passed to \code{assignPermafrost}}

\item{ras}{passed to \code{assignPermafrost}}

\item{id}{passed to \code{assignPermafrost}}

\item{saveDir}{passed to \code{assignPermafrost}}

\item{cores}{how many threads to use for parallelisation. If
\code{1}, no parallelisation occurs}

\item{outFilename}{name of output permafrost raster file}

\item{dPath}{directory where output permafrost raster file will be written.}
}
\description{
Wrapper for \code{assignPermafrost}
}
