% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanLandWebStudyArea.R
\name{.cleanLandWebStudyArea}
\alias{.cleanLandWebStudyArea}
\title{Clean up the LandWeb study area from David Andison}
\usage{
.cleanLandWebStudyArea(poly, minFRI = 40)
}
\arguments{
\item{poly}{A polygon or character string identifying the path to polygon}

\item{minFRI}{Numeric or integer, indicating the minimum fire return interval
that will be part of the cleanup of polygon. Anything below
this will be \code{NA}.}
}
\description{
Clean up the LandWeb study area from David Andison
}
