% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{.createCohortData}
\alias{.createCohortData}
\title{Internal function used by \code{\link[=makeAndCleanInitialCohortData]{makeAndCleanInitialCohortData()}}.}
\usage{
.createCohortData(
  inputDataTable,
  sppColumns,
  minCoverThreshold = 5,
  doAssertion = getOption("LandR.assertions", TRUE),
  rescale = TRUE,
  minSpeciesEcoregionShare = 0
)
}
\arguments{
\item{inputDataTable}{A \code{data.table} with columns described above.}

\item{sppColumns}{A vector of the names of the columns in \code{inputDataTable} that
represent percent cover by species, rescaled to sum up to 100\%\%.}

\item{minCoverThreshold}{minimum total cover percentage necessary to consider the pixel
vegetated, or a cohort present in a pixel.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{rescale}{Logical. If \code{TRUE}, the default, cover for each species will be rescaled
so all cover in \code{pixelGroup} or pixel sums to 100.}

\item{minSpeciesEcoregionShare}{See \code{\link[=makeAndCleanInitialCohortData]{makeAndCleanInitialCohortData()}}.}
}
\value{
\code{cohortData} (\code{data.table}) with attribute \code{"imputedPixID"}
}
\description{
Internal function used by \code{\link[=makeAndCleanInitialCohortData]{makeAndCleanInitialCohortData()}}.
}
\keyword{internal}
