% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMs.R
\name{.getMaxBCoefs}
\alias{.getMaxBCoefs}
\title{Get maximum biomass coefficient names}
\usage{
.getMaxBCoefs(mll, model = c("CR", "Logistic"))
}
\arguments{
\item{mll}{the output of an \code{bbmle::mle2} call (the fitted non-linear model),
from which coefficient values will be extracted}

\item{model}{character. Non-linear model form used to estimate average maximum
biomass. One of "CR" (Chapman-Richards) or "Logistic".}
}
\value{
a list of two vectors of parameter names one following
coefficient names in \code{mll} ('mllCoefNames'), the other using the original
names as in the data used for model fitting ('origCoefNames')
}
\description{
Extracts the names of linear coefficients
for the maximum-biomass-equivalent parameter in
the non-linear growth equations
}
