% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.interpolateLightProb}
\alias{.interpolateLightProb}
\title{Find interpolated value of light probability}
\usage{
.interpolateLightProb(x, x0, x1, y0, y1)
}
\arguments{
\item{x}{the species shade tolerance trait value for which we want to find
the interpolated probability.}

\item{x0}{the \code{floor} of \code{x} corresponding to a class of shade
tolerance in the \code{sufficientLight} table.}

\item{x1}{the \code{ceiling} of \code{x} corresponding to a class of shade
tolerance in the \code{sufficientLight} table.}

\item{y0}{the probability of germination in the \code{sufficientLight} table
corresponding to \code{x0}.}

\item{y1}{the probability of germination in the \code{sufficientLight} table
corresponding to \code{x1}.}
}
\value{
\code{vector} of the interpolated value
}
\description{
Find interpolated value of light probability
}
