% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.logFloor}
\alias{.logFloor}
\title{Log-transformed values, with a floor (> 0)}
\usage{
.logFloor(x, floor = 0.3)
}
\arguments{
\item{x}{Numeric.}

\item{floor}{Minimum age value boundary. Default \code{0.3}.}
}
\description{
Avoid \code{-Inf} problems when \code{x == 0} by setting a non-zero floor value for \code{x}.
This is preferred over using some \code{log(x + d)} transformation, as the choice of \code{d} is
arbitrary, and will affect model fit.
}
\examples{
\dontrun{
  x <- sample(0:400, 1e7, replace = TRUE)
  floor <- 0.3
  logFloor <- log(floor)
  microbenchmark::microbenchmark(
    log(pmax(floor, x)),
    pmax(log(floor), log(x)),
    {x[x < floor] <- floor; log(x)},
    {y <- log(x); y[is.infinite(y)] <- logFloor} ## fastest; TODO: implement? not a bottleneck
  )
}
}
\keyword{internal}
