% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD:man/dot-makePixelGroups.Rd
% Please edit documentation in R/LandR.R
\name{.makePixelGroups}
\alias{.makePixelGroups}
=======
% Please edit documentation in R/cohorts.R
\name{makePixelGroups}
\alias{makePixelGroups}
>>>>>>> development:man/makePixelGroups.Rd
\title{Create the correct string for pixelGroups}
\usage{
.makePixelGroups(maxPixelGroup, ecoregionGroup, speciesGroup,
  columns = c("ecoregionGroup", "speciesGroup", "age"))
}
\arguments{
\item{maxPixelGroup}{A length 1 numeric/integer indicating the current maximum pixelGroup value}

\item{ecoregionGroup}{A vector of ecoregionGroup strings}

\item{speciesGroup}{A vector of speciesGroup strings}

\item{columns}{A character vector of column names to use as part of the generation of unique
combinations of features. Default is \code{c("ecoregionGroup", "speciesCode", "age")}}
}
\value{
TODO: description needed
}
\description{
Create the correct string for pixelGroups
}
