% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.mergeRasterFiles}
\alias{.mergeRasterFiles}
\title{Load and merge a list of raster files into a single larger
raster}
\usage{
.mergeRasterFiles(rasFileLs)
}
\arguments{
\item{a}{\code{list} of raster file paths.}
}
\value{
A \code{SpatRaster}
}
\description{
Load and merge a list of raster files into a single larger
raster
}
