% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{.MLLMaxBPartialPlotData}
\alias{.MLLMaxBPartialPlotData}
\title{Prepare data for model plotting}
\usage{
.MLLMaxBPartialPlotData(
  mll,
  nonLinModelQuoted,
  linModelQuoted,
  targetCovar = "cover",
  fixMaxCover = TRUE,
  maxCover = 1,
  data,
  fun = "mean",
  plotCIs = TRUE
)
}
\arguments{
\item{mll}{outputs of an \code{bbmle::mle2} call (the fitted non-linear model),
from which coefficient values will be extracted.}

\item{nonLinModelQuoted}{The non-linear equation as a \code{call}
(quoted expression) passed to \code{mle2(minuslog1)}. See \code{?mle}.
Accepts equations with three parameters 'A', 'p' and 'k'.}

\item{linModelQuoted}{A list of linear equations/modes relating each
parameter ('A', 'p' and 'k') with a set of covariates. A \code{call}
(quoted expression) passed to \code{mle2(..., parameters)}. Note that for the
purpose of tree growth, the linear equation determining 'A' should include a
'cover' predictor indicating the tree cover or dominance in the stand. Should be
scaled between 0 and \code{maxCover}.}

\item{targetCovar}{the covariate for which variation in \code{maxB} values will be shown.
Defaults to showing how \code{maxB} values change with "cover".
All other covariates except "age" are averaged.
Age values are generated as \code{round(seq(min(age), max(age)*1.5, length.out = 100), 0)}.
When \code{targetCovar != "cover"}, "cover" will be fixed at \code{maxCover}. See \code{fixMaxCover}.}

\item{fixMaxCover}{logical. If \code{TRUE} and \code{targetCovar != "cover"}, cover is
not averaged and is fixed to \code{maxCover}.}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{data}{data for estimation of maximum biomass. Should contain at least
an 'age' column. Note that other covariates will be averaged and 'cover' values
will be replaced with the maximum cover value (\code{maxCover}).}

\item{fun}{The function to apply when summarizing other variables. By default,
the all other variables except age are  averaged (\code{"mean"}). Other options are:
\code{"median"}, \code{"min"}, \code{"max"}.}

\item{plotCIs}{should confidence intervals be calculated and plotted?}
}
\description{
Prepare data for model plotting
}
\seealso{
\code{\link[bbmle:mle2]{bbmle::mle2()}}
}
