% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{.MLLMaxBplotData}
\alias{.MLLMaxBplotData}
\title{Prepare data for model plotting}
\usage{
.MLLMaxBplotData(
  mll,
  nonLinModelQuoted,
  linModelQuoted,
  maxCover,
  data,
  averageCovariates = TRUE,
  observedAge = FALSE,
  plotCIs = TRUE
)
}
\arguments{
\item{mll}{outputs of an \code{bbmle::mle2} call (the fitted non-linear
model), from which coefficient values will be extracted.}

\item{nonLinModelQuoted}{The non-linear equation as a \code{call}
(quoted expression) passed to \code{mle2(minuslog1)}. See \code{?mle}.
Accepts equations with three parameters 'A', 'p' and 'k'.}

\item{linModelQuoted}{A list of linear equations/modes relating each
parameter ('A', 'p' and 'k') with a set of covariates. A \code{call}
(quoted expression) passed to \code{mle2(..., parameters)}. Note that for the
purpose of tree growth, the linear equation determining 'A' should include a
'cover' predictor indicating the tree cover or dominance in the stand. Should be
scaled between 0 and \code{maxCover}.}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{data}{data for estimation of maximum biomass. Should contain at least
an 'age' column. Note that other covariates will be averaged and 'cover' values
will be replaced with the maximum cover value (\code{maxCover}).}

\item{averageCovariates}{should covariates other than age/cover be averaged for
biomass predictions? If not, for each age (at maximum cover) there will be as
many predictions as other covariate values. If \code{observedAge == TRUE} and
\code{averageCovariates == FALSE} then the original data is used, with cover
changed to \code{maxCover}.}

\item{observedAge}{should observed age values be used, or should these be generated
as \code{round(seq(min(age), max(age)*1.5, length.out = 100), 0)}? If \code{observedAge == TRUE} and
\code{averageCovariates == FALSE} then the original data is used, with cover
changed to \code{maxCover}.}

\item{plotCIs}{should confidence intervals be calculated and plotted?}
}
\description{
Prepare data for model plotting
}
