% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.NAvalueFlag}
\alias{.NAvalueFlag}
\title{Set NA value in Raster/SpatRaster}
\usage{
.NAvalueFlag(ras, NAval)
}
\arguments{
\item{ras}{a Raster* , or SpatRaster object}

\item{NAval}{the value to use as \code{NA}}
}
\value{
a raster with attributed NA values
}
\description{
Set NA value in Raster/SpatRaster
}
