% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{.overlay}
\alias{.overlay}
\title{Overlaying function}
\usage{
.overlay(
  SPP,
  HQ,
  LQ,
  hqLarger,
  highQualityStack,
  lowQualityStack,
  outputFilenameSuffix = "overlay",
  destinationPath
)
}
\arguments{
\item{SPP}{\code{data.table} column of species layer name}

\item{HQ}{\code{data.table} column of whether \code{SPP} is present in HQ layers}

\item{LQ}{\code{data.table} column of whether \code{SPP} is present in LQ layers}

\item{highQualityStack}{high quality list/stack of rasters
(will be used preferentially)}

\item{lowQualityStack}{low quality list/stack of rasters
(will be used to fill \code{NA}s in \code{highQualityStack})}

\item{outputFilenameSuffix}{file suffix to save raster if there was overlaying.
Defaults to \code{"overlay"}.}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}
}
\description{
Used internally in \code{overlayStacks}. Function to be applied to each row
of a \code{data.table} containing information of whether the species layer
exists in the HQ and LQ data.
Only overlays if data exists in both layers, otherwise returns the layer with data.
}
\keyword{internal}
