% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.projectExtent}
\alias{.projectExtent}
\title{Project raster extent}
\usage{
.projectExtent(ras, crs)
}
\arguments{
\item{ras}{a Raster* or SpatRaster object}

\item{crs}{passed to \code{raster::projectRaster(..., crs = crs)}
and \code{terra::project(..., y = crs)}}
}
\value{
the projected extent
}
\description{
Project raster extent
}
