% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.stack}
\alias{.stack}
\title{Make stacked raster}
\usage{
.stack(rasList)
}
\arguments{
\item{rasList}{a list of Raster* or SpatRaster layers}
}
\value{
a stacked raster
}
\description{
Make stacked raster
}
