% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{downloadChecksums}
\alias{downloadChecksums}
\title{Download checksums file from Google Drive}
\usage{
downloadChecksums(checksumsFileID)
}
\arguments{
\item{checksumsFileID}{the Google Drive file id for the remote checksums file}
}
\description{
Very experimental
}
\keyword{internal}
