% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{extractMaxB}
\alias{extractMaxB}
\title{Maximum biomass estimator}
\usage{
extractMaxB(mll, newdata, average = FALSE, model = c("CR", "Logistic"))
}
\arguments{
\item{mll}{the output of an \code{bbmle::mle2} call (the fitted non-linear model),
from which coefficient values will be extracted}

\item{newdata}{data for estimation of 'A'}

\item{average}{should 'A' be estimated for average values of its predictors.}

\item{model}{character. Non-linear model form used to estimate average maximum
biomass. One of "CR" (Chapman-Richards) or "Logistic".}
}
\description{
Estimation of maximum biomass as the A parameter in the Chapman-Richards and
Logistic growth equations. Since A is modelled as linear term, it is a matrix
product of its linear coefficients. It is assumed that all coefficients are
related additively.
}
