% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{fitNLMModels}
\alias{fitNLMModels}
\title{FUNCTIONS TO FIT NON-LINEAR MODELS TO ESTIMATE MAXB
Wrapper function to fit non-linear growth model per species.}
\usage{
fitNLMModels(
  sp = NULL,
  predictorVarsData,
  sppVarsB,
  predictorVars,
  predictorVarsCombos = NULL,
  maxNoCoefs = 4,
  doFwdSelection = FALSE,
  sampleSize = 3000,
  Ntries = 2000,
  maxCover = 1L,
  models = c("CR", "Logistic"),
  modelOutputsPrev = NULL,
  randomStarts = FALSE,
  lowerBounds = TRUE,
  upperBounds = TRUE,
  nbWorkers = 1L
)
}
\arguments{
\item{sp}{species name -- only used for messaging.}

\item{predictorVarsData}{a \code{data.table} of predictor variables including
those in \code{predictorVars} and \code{age}, as well as \code{pixelIndex}.
Note that \code{age} should be in the original scale (e.g., not logged).}

\item{sppVarsB}{s \code{data.table} of species biomass (\code{B}) and \code{pixelIndex}.}

\item{predictorVars}{character vector of predictor variables to be included
in the linear component of the model affecting the asymptote (need to correspond to
\code{names(predictorVarsData[[sp]])}) the same predictors will be considered for all
species.}

\item{predictorVarsCombos}{a list of sets of covariates in \code{predictorVars} to add to
the fitted models. If this list has several entries with sets of covariates,
each will be fitted as part of the model selection process.}

\item{maxNoCoefs}{how many covariates from \code{predictorVars} should be added to
the linear component of the model affecting the asymptote? Note that the
more covariates are added the longer the model takes to fit, as all
combinations are attempted. For 2 or more covariates, only combinations with
"cover" are attempted.}

\item{doFwdSelection}{should covariates be added one at a time to the
linear component of the model? If \code{TRUE}, and \code{is.null(predictorVarsCombos)},
then each entry in \code{predictorVarsCombos} is used as the set of covariates to
test. Otherwise, \code{predictorVarsCombos} will be created from combinations of
\code{predictorVars}, with \code{maxNoCoefs} determining the maximum number of covariates
to add. If \code{FALSE} the full model is fitted.}

\item{sampleSize}{how many data points should be randomly sampled to fit the model?
If \code{NA} the full dataset will be used. Note that this may result in long
computation times. Biomass data will be binned into 10 regular bins before sampling points in number
equal to \code{sampleSize}.}

\item{Ntries}{how many times should the models be fit with new randomly
generated starting values? Only used if \code{randomStarts == TRUE}.}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{models}{character vector of models to fit. Only Chapman-Richards ('CR')
and 'Logistic' can be chosen at the moment.}

\item{modelOutputsPrev}{previous outputs of \code{fitNLMmodels}. The model will try
refitting and comparing AIC with the last results.}

\item{randomStarts}{logical. Should random starting values of A, k and p non-linear
parameters be picked from a range sensible values, or should all combinations
of values within this range be used? If FALSE, the default, the starting values
are spaced at regular intervals within an acceptable range for each parameter -- 20
values for A, 10 for k and p -- and all combinations are used (2000 starting
values in total). Parameter ranges are estimated from data following Fekedulegn et al. (1999)
as follows:
\itemize{
\item{range of \code{A} starting values (B0 parameter in Fekedulegn et al. 1999)
varies between \eqn{ObsMaxB \times 0.3} and \eqn{ObsMaxB \times 0.9}, where ObsMaxB is the
maximum observed B across the full dataset (not the sampled data for fitting)}
\item{\code{k} (CR model) and \code{p} (Logistic model; both are B2 parameter in
Fekedulegn et al. 1999) are estimated as a constant rate to get to ObsMaxB,
calculated as \eqn{\frac{\frac{Bobs2 - Bobs1}{age2 - age1}}{ObsMaxB}}, where \emph{B1/2} and
\emph{age1/2} are are observed values at two points in time. We draw 100 samples
of two \emph{age} values, and corresponding \emph{B}, to calculate a sample of rates.
After excluding rates <= 0, we take the minimum and maximum values as
the range of \code{k} (CR model) or \code{p} (Logistic model) parameters of the growth model}
\item{the \code{p} parameter (CR model) (related to B3 parameter in Fekedulegn
et al. 1999) should be > 1. Here we use a range of values between 1.1 and
80 which provided suitable fitting using data from the Northwest Territories,
Canada}
\item{the \code{k} parameter (Logistic model; B1 parameter in Fekedulegn et al. 1999)
is estimated as \eqn{B_0 = \frac{ObsMaxB}{1 + k}}, using a small positive
number for \eqn{B_0}, e.g. 2. Here, we estimate \code{k} values for \eqn{B_0}
values 1 to 5, and use the minimum and maximum to determine the range from
where to draw starting values.}
}}

\item{lowerBounds}{a named vector of lower parameter boundaries. If FALSE, no lower
boundaries are applied. If TRUE, coefficients of the linear model on the A
parameter (intercept, cover, k and p) are bound (intercept = observed maximum B * 0.5,
cover = 0, k = 0.05 and p = 1). Alternatively, pass a named vector of parameter
boundaries.}

\item{upperBounds}{a named vector of upper parameter boundaries. If FALSE, no lower
boundaries are applied. If TRUE, coefficient of the linear model on the A parameter
(intercept and k) are bound (intercept = observed maximum B * 1.5, k = 0.2). Alternatively, pass
a named vector of parameter boundaries.}

\item{nbWorkers}{integer. If > 1, the number of workers to use in \code{future.apply::future_apply}, otherwise
no parallelisation is done.}
}
\description{
The function will create sensible parameter ranges for A, k and p parameters
of the Chapman-Richards and Logistic growth curves and attempt to run a
forward step-AIC procedure to add covariates to the linear component of the
model (on the A parameter -- the asymptote). The maximum number of covariates
to add is determined by \code{maxNoCoefs}.
}
