% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gadm_canada}
\alias{gadm_canada}
\title{Canada administrative boundaries}
\usage{
gadm_canada(src = "stats_can", dst_path = tempdir())
}
\arguments{
\item{src}{Character. One of "stats_can" or (if \code{geodata} package is installed) "geodata".}

\item{dst_path}{Character specifying the path to download to.}
}
\description{
Canada administrative boundaries
}
