% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{generatePixelGroups}
\alias{generatePixelGroups}
\title{Add the correct \code{pixelGroups} to a \code{pixelDataTable} object}
\usage{
generatePixelGroups(
  pixelDataTable,
  maxPixelGroup,
  columns = c("ecoregionGroup", "speciesCode", "age", "B")
)
}
\arguments{
\item{pixelDataTable}{A \code{data.table} with column-based descriptions.
Must have a column called \code{pixelIndex}, which allows for multiple rows to be associated
with a single pixel.}

\item{maxPixelGroup}{A length 1 numeric indicating the current maximum \code{pixelGroup} value;
the \code{pixelGroup} numbers returned will start at \code{maxPixelGroup + 1}.}

\item{columns}{A character vector of column names to use as part of the generation of unique
combinations of features. Default is \code{c("ecoregionGroup", "speciesCode", "age", "B")}}
}
\value{
Returns a vector of \code{pixelGroup} in the original order of the input \code{pixelDataTable}.
This should likely be added to the \code{pixelDataTable} object immediately.
}
\description{
Generates unique groupings of a \code{data.table} object where one or more rows can
all belong to the same \code{pixelIndex}. Pixel groups will be identical pixels based
on unique combinations of \code{columns}.
}
