% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{genericExtract}
\alias{genericExtract}
\title{GENERIC EXTRACT POINTS}
\usage{
genericExtract(x, y, ...)
}
\arguments{
\item{y}{a points spatial object (\code{sp}, \code{sf}, or \code{terra})}

\item{...}{passed to \code{terra::extract}}

\item{ras}{a raster object (\code{raster} or \code{terra})}
}
\value{
a data.table with extracted values
}
\description{
Extracts points from raster layers using the
original raster layer projection.
}
