% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbances.R
\name{genPGsPostDisturbance}
\alias{genPGsPostDisturbance}
\title{Re-generate new \code{pixelGroup}s in partially disturbed pixels.}
\usage{
genPGsPostDisturbance(
  cohortData,
  pixelGroupMap,
  disturbedPixelTable,
  disturbedPixelCohortData,
  colsForPixelGroups = LandR::columnsForPixelGroups,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{\code{data.table}. The pre-disturbance \code{cohortData} table}

\item{pixelGroupMap}{\code{SpatRaster}. The pre-disturbance \code{pixelGroupMap}.}

\item{disturbedPixelTable}{\code{data.table}. A table with at least the \code{pixelIndex}
of all disturbed pixels. Additional columns are ignored.}

\item{disturbedPixelCohortData}{a \code{cohortData}-like table with information of dead,
and surviving, but \emph{NOT} regenerating cohorts (cohorts for whom regeneration via, e.g., serotiny
or resprouting was successfully activated), in \emph{disturbed pixels only}. Dead cohorts should
age B == 0, surviving cohorts B > 0.}

\item{colsForPixelGroups}{character. \code{cohortData} columns used to find identical pixels in terms of
cohort composition, and group them into \code{pixelGroups}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\value{
a named list with:
\itemize{
\item a \code{cohortData} table with the updated \code{pixelGroups}, as well as
survivor cohorts, but not dead cohorts.
\item a \code{pixelGroupMap} with the updated \code{pixelGroups} in disturbed pixels
}
}
\description{
Re-generate new \code{pixelGroup}s in partially disturbed pixels.
}
\details{
This function regenerates \code{pixelGroup}s in situations where
disturbances are not stand-replacing and create survivor and dead
cohorts in some, but potentially not all, pixels of a \code{pixelGroup}. This is
necessary to prevent reintroducing dead cohorts that were not affected
in the other pixels of the same, original, \code{pixelGroup}.
\strong{ATTENTION} This function alone will not generate final \code{pixelGroups},
and will likely need to be followed by an \code{updateCohortData} run.
The function:
\enumerate{
\item creates a table of undisturbed pixels, and disturbed pixels with dead and surviving cohorts
but not new cohorts (e.g. from serotiny/resprouting) -- these are added by \code{updateCohortData};
\item generates \code{pixelGroups} considering these dead and surviving cohorts -- this ensures that
we account for cohorts that died in some but not all pixels of a given \code{pixelGroup};
\item then removes dead cohorts (as they should not be in tables for downstream
functions like \code{updateCohortData}) and recalculates \code{pixelGroups} -- this ensures that
pixels that became similar/dissimilar after the death of some cohorts can form new \code{pixelGroups}.
}
}
