% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landr.R
\name{geom_landr}
\alias{geom_landr}
\title{This geom adds land polygons from OpenStreetMap to ggplot2 maps.}
\usage{
geom_landr(after = 0, simplified = FALSE, ...)
}
\arguments{
\item{after}{index of the layer after which the land polygons should be added. Set to 0 to add at the bottom.}
}
\description{
This geom adds land polygons from OpenStreetMap to ggplot2 maps.
}
