% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultData.R
\name{getPermafrostDataGibson}
\alias{getPermafrostDataGibson}
\title{Download permafrost/thermokarst layer}
\usage{
getPermafrostDataGibson(destinationPath, studyArea, userTags)
}
\arguments{
\item{destinationPath}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{studyArea}{a spatial polygon  (\code{sf}, \code{sp} or \code{SpatVector}) used to crop
and mask data layers. Passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{userTags}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}
}
\value{
Carolyn Gisbon's output polygon layer of \% PPC and thermokarst state
}
\description{
Downloads and crops/masks permafrost peatland complex cover and amount of thermokarst
spatial data from Gibson et al. (2021)
}
