% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{ggplotMLL_maxB}
\alias{ggplotMLL_maxB}
\title{Plot estimated maximum biomass by age}
\usage{
ggplotMLL_maxB(
  mll,
  data,
  maxCover = 1L,
  xCovar = "age",
  plotTitle = NULL,
  nonLinModelQuoted,
  linModelQuoted,
  averageCovariates = TRUE,
  observedAge = FALSE,
  plotCIs = TRUE
)
}
\arguments{
\item{mll}{a named list with outputs of an \code{bbmle::mle2} call (the fitted non-linear
model), from which coefficient values will be extracted. If several model outputs
are provided all fitted models will be plotted, with plot labels corresponding to list names.}

\item{data}{data for estimation of maximum biomass. Should contain at least
an 'age' column. Note that other covariates will be averaged and 'cover' values
will be replaced with the maximum cover value (\code{maxCover}). If \code{mll} is a list
data is assumed to be the same for the two models.}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{xCovar}{the variable shown in the x axis. Defaults to \code{age}.}

\item{plotTitle}{character. Passed to ggplot2::labs(title)}

\item{nonLinModelQuoted}{a named list of non-linear equations as a \code{call}
(quoted expression) passed to \code{mle2(minuslog1)}. See \code{?mle}.
Accepts equations with three parameters 'A', 'p' and 'k'. List names and
length must the same as in \code{mll}.}

\item{linModelQuoted}{A named list of lists of linear equations/modes relating each
parameter ('A', 'p' and 'k') with a set of covariates. A \code{call}
(quoted expression) passed to \code{mle2(..., parameters)}. Note that for the
purpose of tree growth, the linear equation determining 'A' should include a
'cover' predictor indicating the tree cover or dominance in the stand. Should be
scaled between 0 and \code{maxCover}. List names and length must the same as in \code{mll}.}

\item{averageCovariates}{should covariates other than age/cover be averaged for
biomass predictions? If not, for each age (at maximum cover) there will be as
many predictions as other covariate values. If \code{observedAge == TRUE} and
\code{averageCovariates == FALSE} then the original data is used, with cover
changed to \code{maxCover}.}

\item{observedAge}{should observed age values be used, or should these be generated
as \code{round(seq(min(age), max(age)*1.5, length.out = 100), 0)}? If \code{observedAge == TRUE} and
\code{averageCovariates == FALSE} then the original data is used, with cover
changed to \code{maxCover}.}

\item{plotCIs}{should confidence intervals be calculated and plotted?}
}
\description{
Plots a maximum biomass estimated at maximum 'cover'
(or dominance) levels as a function of age
}
