% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosim.R
\name{installBioSIM}
\alias{installBioSIM}
\title{Install \pkg{BioSIM} to retrieve climate and other projections using \code{BioSIM}}
\usage{
installBioSIM(lib)
}
\arguments{
\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }
}
\description{
Install \pkg{BioSIM} to retrieve climate and other projections using \code{BioSIM}
}
