% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\name{land}
\alias{land}
\title{Determine how far inland a point is located.}
\usage{
land(lon, lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\value{
a vector of distances in decimal degrees
}
\description{
Determine how far inland a point is located.
}

