% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\docType{data}
\name{landisIIrepo}
\alias{landisIIrepo}
\title{Default LANDIS-II project repo url}
\format{An object of class \code{character} of length 1.}
\usage{
landisIIrepo
}
\description{
Default LANDIS-II project repo url
}
\keyword{internal}
