% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landr.R
\docType{package}
\name{landr}
\alias{landr}
\title{Add land polygons to ggplot2 maps}
\description{
This package adds land polygons from OpenStreetMap to ggplot2 maps.
}
\author{
Pieter Provoost, \email{pieterprovoost@gmail.com}
}
