% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{loadCASFRI}
\alias{loadCASFRI}
\title{Load CASFRI data}
\usage{
loadCASFRI(
  CASFRIRas,
  attrFile,
  headerFile,
  sppEquiv,
  sppEquivCol,
  type = c("cover", "age")
)
}
\arguments{
\item{CASFRIRas}{\code{RasterLayer} object}

\item{attrFile}{character string specifying a CASFRI attribute filename}

\item{headerFile}{character string specifying a CASFRI header filename}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}

\item{type}{Character string. Either \code{"cover"} or \code{"age"}.}
}
\value{
list with named elements: \code{CASFRIattrLong} and \code{CASFRIdt} (both data.tables)
}
\description{
Loads CASFRIv4 data specifically supplied for the LandWeb project
}
