% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makeBiomassMap}
\alias{makeBiomassMap}
\title{Create \code{biomassMap}}
\usage{
makeBiomassMap(pixelCohortData, rasterToMatch)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}
}
\value{
The \code{biomassMap}, a raster of total stand biomass per pixel.
}
\description{
This is a function that creates the \code{biomassMap} raster used  for simulations in
\code{Biomass_core} module, using estimated data based on \code{rawBiomassMap} contained in
\code{pixelCohortData}.
}
