% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{makeCohortDataFiles}
\alias{makeCohortDataFiles}
\title{Make the \code{cohortData} table, while modifying the temporary
\code{pixelCohortData} that will be used to prepare other files.}
\usage{
makeCohortDataFiles(pixelCohortData, columnsForPixelGroups, speciesEcoregion)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{columnsForPixelGroups}{Default columns that define pixel groups}

\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values

#' @return
A list with a modified \code{pixelCohortData} and \code{cohortData} \code{data.table}s.}
}
\description{
Takes a \code{pixelCohortData} table (see \code{makeAndCleanInitialCohortData}),
  the \code{speciesEcoregion} list and returns a modified \code{pixelCohortData} and
  the \code{cohortData} tables to be used in the simulation.
  This function mainly removes unnecessary columns from \code{pixelCohortData},
  subsets pixels with \code{biomass > 0}, generates \code{pixelGroups},
  and adds \code{ecoregionGroup} and \code{totalBiomass} columns to \code{pixelCohortData}.
  \code{cohortData} is then created by subsetting unique combinations of \code{pixelGroup} and
  whatever columsn are listed in \code{columnsForPixelGroups}.
  The resulting \code{cohortData} table has the follwing columns:
\itemize{
  \item \code{speciesCode} (factor)
  \item \code{ecoregionGroup} (factor)
  \item \code{pixelGroup} (integer)
  \item \code{age} (integer)
  \item \code{B} (integer)
}
}
