% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyEcoregionFiles}
\alias{makeDummyEcoregionFiles}
\title{Makes dummy \code{ecoregionFiles} list}
\usage{
makeDummyEcoregionFiles(ecoregionMap, rstLCC, rasterToMatch)
}
\arguments{
\item{ecoregionMap}{a raster of all the unique groupings.}

\item{rstLCC}{a raster land-cover class per pixel,}

\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used
throughout the simulation)}
}
\value{
The \code{ecoregionFiles}, A list with two objects: the \code{ecoregionMap}
  and a table summarizing its information per pixelID. See \code{ecoregionProducer}
}
\description{
This is a function that makes a list of dummy \code{ecoregionFiles},
 using dummy versions of \code{ecoregionMap} and \code{rstLCC}
 (see \code{makeDummyEcoregionMap} and \code{makeDummyRstLCC})
 for test simulations in LBMR. It uses \code{ecoregionProducer} internally
}
