% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyEcoregionMap}
\alias{makeDummyEcoregionMap}
\title{Makes a dummy \code{ecoregionMap}}
\usage{
makeDummyEcoregionMap(rasterToMatch)
}
\arguments{
\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used
throughout the simulation)}
}
\value{
The \code{ecoregionMap}, a raster of all the unique groupings.
}
\description{
This is a function that makes a dummy \code{ecoregionMap} raster for
test simulations in LBMR.
}
