% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyRawBiomassMap}
\alias{makeDummyRawBiomassMap}
\title{Makes a dummy \code{rawBiomassMap}}
\usage{
makeDummyRawBiomassMap(rasterToMatch)
}
\arguments{
\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used
throughout the simulation)}
}
\value{
The \code{rawBiomassMap}, a raster of "raw" total stand biomass per pixel,
     with values between 100 and 20000 g/m2
}
\description{
This is a function that makes a dummy \code{rawBiomassMap} raster for
test simulations in LBMR.
}
