% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyRstLCC}
\alias{makeDummyRstLCC}
\title{Makes a dummy \code{rstLCC}}
\usage{
makeDummyRstLCC(rasterToMatch)
}
\arguments{
\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used
throughout the simulation)}
}
\value{
The \code{rstLCC}, a raster land-cover class per pixel,
     with values between 1 and 5 that have no
     correspondence to any real land-cover classes.
}
\description{
This is a function that makes a dummy \code{rstLCC} raster for
test simulations in LBMR.
}
