% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyStandAgeMap}
\alias{makeDummyStandAgeMap}
\title{Makes a dummy \code{standAgeMap}}
\usage{
makeDummyStandAgeMap(rawBiomassMap)
}
\arguments{
\item{rawBiomassMap}{a \code{rawBiomassMap} (e.g. the one used
throughout the simulation)}
}
\value{
The \code{standAgeMap}, a raster of stand age per pixel (where biomass exists)
     with values between 1 and 300 years
}
\description{
This is a function that makes a dummy \code{standAgeMap} raster for
test simulations in LBMR.
}
