% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{makeEcoregionDT}
\alias{makeEcoregionDT}
\title{Make the \code{ecoregion} table}
\usage{
makeEcoregionDT(pixelCohortData, speciesEcoregion)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-{ecoregion}-specific species
trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar
biophysical characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.}
}
\value{
A \code{data.table} with ecoregion codes and their active status per \code{pixelID}.
}
\description{
This function creates a table containing pixel-wise ecoregion codes and whether they
are "active" (have biomass > 0) or not for simulation. Unlike \code{ecoregionProducer},
this function creates the \code{ecoregion} table from pixel information contained in
\code{pixelCohortData}
}
