% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{makeForkClusterRandom}
\alias{makeForkClusterRandom}
\title{\code{makeForkCluster} with random seed set}
\usage{
makeForkClusterRandom(..., iseed = NULL)
}
\arguments{
\item{...}{passed to \code{makeForkCluster}, e.g.,}

\item{iseed}{passed to \code{clusterSetRNGStream}}
}
\description{
This will set different random seeds on the clusters (not the default)
with \code{makeForkCluster}.
It also defaults to creating a logfile with message of where it is.
}
