% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makeMinRelativeB}
\alias{makeMinRelativeB}
\title{Makes \code{minRelativeB}}
\usage{
makeMinRelativeB(pixelCohortData)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}
}
\value{
a data.frame of min relative biomass values per ecoregion group.
}
\description{
This is a function that creates the \code{minRelativeB} table used  for
  simulations in LBMR. It contains expert-based values for minimum relative
  biomass of each shade tolerance class (the minimum relative biomass a cohort
  with a given shade tolerance should have to be able to germinate),
 in each unique ecoregion group. In this function, all ecoregion groups have
 the same values.
}
