% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{makeOptimalCluster}
\alias{makeOptimalCluster}
\title{Create a parallel Fork cluster, if useful}
\usage{
makeOptimalCluster(useParallel = getOption("pemisc.useParallel", FALSE),
  MBper = 500, maxNumClusters = parallel::detectCores(), ...)
}
\arguments{
\item{useParallel}{Logical or numeric. If \code{FALSE}, returns NULL. If
\code{numeric}, then will return a cluster object with this
many cores, up to \code{maxNumClusters}}

\item{MBper}{Numeric. Passed to \code{memRequiredMB} in
\code{\link{optimalClusterNum}}}

\item{maxNumClusters}{Numeric or Integer. The theoretical upper limit
for number of clusters to create (e.g., because there are only
3 problems to solve, not \code{parallel::detectCores})}

\item{...}{Passed to \code{makeForkClusterRandom}.
Only relevant for \code{iseed}.}
}
\description{
Given the size of a problem, it may not be useful to create a cluster.
This will make a Fork cluster (so Linux only)
}
