% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permafrost.R
\name{makePermafrostRas}
\alias{makePermafrostRas}
\title{Source and prepare permafrost, wetland and land-cover
data and create a permafrost presence-absence raster}
\usage{
makePermafrostRas(
  cores = 1L,
  outPath = getOption("spades.outputPath"),
  cacheTags = NULL,
  destinationPath = getOption("reproducible.destinationPath"),
  studyArea = NULL,
  .studyAreaName = NULL,
  doMissing = FALSE,
  ...
)
}
\arguments{
\item{cores}{integer. Number of threads to use for parallelisation (during
permafrost raster creation). Defaults to no parallelisation (1 thread)}

\item{outPath}{character. A folder path where temporary permafrost rasters
will be saved. Note that these \emph{will not} be deleted at the end of this function}

\item{cacheTags}{character. Common tags to use across \code{reproducible::Cache} calls. These will
be appended to other tags distinguishing the nested \code{reproducible::Cache} calls.}

\item{destinationPath}{character. Passed to \code{reproducible::prepInputs} and the
location where the final permafrost layer will be saved.}

\item{studyArea}{character. Passed to \code{reproducible::postProcess}.}

\item{.studyAreaName}{character. A string to append to \code{filename2} in \code{postProcess}.}

\item{doMissing}{logical. If TRUE, the permafrost layer will only be computed in
polygon IDs missing from raster file names in \code{outPath}.}

\item{...}{further arguments passed to \code{assignPermafrost}}
}
\value{
a \code{SpatRaster} of permafrost presences (1) and absences (0) with the same
properties as the land-cover and wetland rasters, cropped and masked to \code{studyArea}.
}
\description{
Obtains permafrost peatland complex \% data from Gibson et al. (2021),
land-cover data for 2017 from Hermosilla et al. (2022) and wetland
presence/absence data from Wulder et al. (2018), which are crossed
to create a permafrost raster layer.
See \code{?makeSuitForPerm} and \code{?assignPermafrost} for details on how this
is achieved.
}
\references{
Gibson, C., Morse, P. D., Kelly, J. M., Turetsky, M. R., Baltzer, J. L.,
Gingras-Hill, T., & Kokelj, S. V. (2020). Thermokarst Mapping Collective:
Protocol for organic permafrost terrain and preliminary inventory from the Taiga
Plains test area, Northwest Territories (NWT Open Report 2020-010, p. 29).
Northwest Territories Geological Survey.

Hermosilla, T., Bastyr, A., Coops, N. C., White, J. C., & Wulder, M. A. (2022).
Mapping the presence and distribution of tree species in Canada's forested ecosystems.
Remote Sensing of Environment, 282, 113276. https://doi.org/10.1016/j.rse.2022.113276

Wulder, M., Li, Z., Campbell, E., White, J., Hobart, G., Hermosilla, T., &
Coops, N. (2018). A National Assessment of Wetland Status and Trends for Canada's
Forested Ecosystems Using 33 Years of Earth Observation Satellite Data.
Remote Sensing, 10(10), 1623. https://doi.org/10.3390/rs10101623
}
