% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.R
\name{makePixelGroups}
\alias{makePixelGroups}
\title{Create the correct string for pixelGroups}
\usage{
makePixelGroups(maxPixelGroup, ecoregionGroup, speciesGroup)
}
\arguments{
\item{maxPixelGroup}{A length 1 numeric/integer indicating the current maximum pixelGroup value}

\item{ecoregionGroup}{A vector of ecoregionGroup strings}

\item{speciesGroup}{A vector of speciesGroup strings}
}
\value{
TODO: description needed
}
\description{
Create the correct string for pixelGroups
}
