% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultData.R
\name{makeSuitForPerm}
\alias{makeSuitForPerm}
\title{Make a mask of areas that suitable/unsuitable for permafrost}
\usage{
makeSuitForPerm(
  rstLCC,
  wetlands,
  suitableCls = c(40, 50, 100, 210, 220, 230),
  unsuitableCls = c(0, 20, 31, 32, 33, 80, 81),
  studyArea,
  cacheTags
)
}
\arguments{
\item{rstLCC}{a land-cover raster. Should already have been cropped to
\code{studyArea}.}

\item{wetlands}{a wetlands raster (with \code{1L} and \code{NA}s only). Should already
have been cropped to \code{studyArea}.}

\item{suitableCls}{land-cover classes (in \code{rstLCC}) suitable
for permafrost presence}

\item{unsuitableCls}{land-cover classes (in \code{rstLCC}) unsuitable
for permafrost presence}

\item{studyArea}{a polygon of the study area to mask the
output raster to.}

\item{cacheTags}{passed to \code{reproducible::Cache}}
}
\value{
a raster layer with \code{1}s and \code{0}s for pixels with suitable
and unsuitable land cover for permafrost, respectively.
}
\description{
Uses a land-cover map and a wetlands maps to create a mask
of areas suitable for permafrost presence. By default, it
assumes suitable and unsuitable land-cover classification
follows \href{https://www.sciencedirect.com/science/article/pii/S0034425721005009?via\%3Dihub#bb0465}{Hermosilla et al. (2022)}.
See details below.
}
\details{
Suitable areas are classified as \code{1L} and unsuitable
areas as \code{0L} in the output raster.
Unsuitable land-cover classes are:
\itemize{
\item non-forested ecozone, which was not mapped for LC (0)
\item water (20)
\item snow/ice (31)
\item rock/rubble (32)
\item exposed/barren land (33)
\item wetland (80)
\item wetland-treed (81)
Wetland areas in \code{wetlands} (coded as 1L, i.e. wetland presence)
are also coded as unsuitable.
}
}
