% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{makeVegTypeMap}
\alias{makeVegTypeMap}
\title{Make a vegetation type map from a stack of species abundances}
\usage{
makeVegTypeMap(speciesStack, vegLeadingProportion, mixed = TRUE)
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of species abundances.
This must be one \code{RasterLayer} per species.}

\item{vegLeadingProportion}{The threshold as a proportion of the total abundance
that a species must have to be considered a "pure" stand of that type.
If no species reaches this proportion, then the pixel will be 'Mixed'.}

\item{mixedType}{An integer defining whether mixed stands are of any kind of species
admixture (1), or only when deciduous mixed with conifer (2);
or not to consider mixed stands at all (0). Defaults to 2.}
}
\value{
A factor raster
}
\description{
Make a vegetation type map from a stack of species abundances
}
