% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{Map2}
\alias{Map2}
\title{\code{Map} and \code{parallel::clusterMap} together}
\usage{
Map2(..., cl = NULL)
}
\arguments{
\item{...}{passed to \code{Map} or \code{clusterMap}}

\item{cl}{A cluster object, passed to \code{clusterMap}}
}
\description{
This will send to Map or clusterMap, depending on whether cl is provided.
}
