% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mergeSppRaster}
\alias{mergeSppRaster}
\title{Merge species percent-cover rasters}
\usage{
mergeSppRaster(sppMerge, speciesLayers, sppEquiv, column, suffix, dPath, ...)
}
\arguments{
\item{sppMerge}{TODO}

\item{speciesLayers}{A \code{RasterStack} or \code{RasterLayer} that
should contain species cover data in the study area}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{column}{TODO}

\item{suffix}{TODO}

\item{dPath}{destination path TODO}

\item{...}{Additional arguments TODO}
}
\description{
Used internally in \code{overlayStacks}.
}
\keyword{internal}
