% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{minRelativeBDefaults}
\alias{minRelativeBDefaults}
\title{minRelativeB defaults for Western Boreal Forest Canada}
\usage{
minRelativeBDefaults()
}
\description{
minRelativeB defaults for Western Boreal Forest Canada
}
