% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{modifySpeciesAndSpeciesEcoregionTable}
\alias{modifySpeciesAndSpeciesEcoregionTable}
\title{Update \code{species} and \code{speciesEcoregion} tables}
\usage{
modifySpeciesAndSpeciesEcoregionTable(speciesEcoregion, speciesTable)
}
\arguments{
\item{speciesEcoregion}{A \code{data.table} with \code{species}-\code{ecoregion}-specific species trait values.
Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical
characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.
May contain columns \code{inflationFactor}  (used to adjust \code{maxB}) and \code{mANPPproportion}
(used to calculate \code{maxANPP}).}

\item{speciesTable}{A species traits table, with \strong{at least} the following columns:
\itemize{
\item \code{speciesCode} an character representation of species;
\item \code{Area} a character of geographical area within a species range for which trait values are relevant;
\item \code{longevity} species longevity in years, \code{sexualmature} age in years at sexual maturity;
\item \code{shadetolerance} numeric value between 1-5 of relative shade tolerance (with respect to other species);
\item \code{seeddistance_eff} the "effective" seed dispersal distance;
\item \code{seeddistance_max} a numeric with the maximum seed dispersal distance;
\item \code{mortalityshape} and \code{growthcurve}: growth curve shape parameters.
Other columns (e.g. fire-related traits) can also be included depending on LandR modules in use.
Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015)
for further detail.
}}
}
\value{
a \code{list} with modified \code{speciesEcoregion} and \code{speciesTable},
under the names "newSpeciesEcoregion" and "newSpeciesTable", respectively.
}
\description{
Update \code{species} and \code{speciesEcoregion} tables
}
