% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{nonForestedPixels}
\alias{nonForestedPixels}
\title{Assess non-forested pixels based on species cover data and land-cover}
\usage{
nonForestedPixels(
  speciesLayers,
  omitNonTreedPixels,
  forestedLCCClasses,
  rstLCC
)
}
\arguments{
\item{speciesLayers}{A \code{RasterStack} or \code{RasterLayer} that
should contain species cover data in the study area}

\item{omitNonTreedPixels}{logical. Should pixels with classes in \code{forestedLCCClasses} be
included as non-forested?}

\item{forestedLCCClasses}{vector of forested land-cover classes in \code{rstLCC}}

\item{rstLCC}{raster of land-cover class, e.g., LCC2005}
}
\value{
a logical vector of length \code{ncell(rstLCC)}
where \code{TRUE} indicates non-forested pixels where there is no
species cover data, or a non-forested land-cover class
}
\description{
Assess non-forested pixels based on species cover data and land-cover
}
