% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapOrDoCall.R
\name{optimalClusterNum}
\alias{optimalClusterNum}
\title{Determine the number of nodes to use in a new cluster}
\usage{
optimalClusterNum(memRequiredMB = 500, maxNumClusters = 1)
}
\arguments{
\item{memRequiredMB}{The amount of memory needed in MB}

\item{maxNumClusters}{The maximum number of nodes to use}
}
\description{
TODO: DESCRIPTION NEEDED
}
