% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxBestimation.R
\name{partialggplotMLL_maxB}
\alias{partialggplotMLL_maxB}
\title{Partial effect plots of maximum biomass estimates by age}
\usage{
partialggplotMLL_maxB(
  mll,
  data,
  targetCovar = "cover",
  maxCover = 1,
  fixMaxCover = TRUE,
  xCovar = "age",
  showQuantiles = "allQuantiles",
  plotTitle = NULL,
  nonLinModelQuoted,
  linModelQuoted,
  fun = "mean",
  plotCIs = TRUE
)
}
\arguments{
\item{mll}{a named list with outputs of an \code{bbmle::mle2} call (the fitted non-linear
model), from which coefficient values will be extracted. If several model outputs
are provided all fitted models will be plotted, with plot labels corresponding to list names.}

\item{data}{data for estimation of maximum biomass. Should contain at least
an 'age' column. Note that other covariates will be averaged and 'cover' values
will be replaced with the maximum cover value (\code{maxCover}).
If \code{mll} is a list, data is assumed to be the same for the two models.}

\item{targetCovar}{the covariate for which variation in maxB values will be shown per
age value. Defaults to showing how maxB values change with "cover" at any given age.
Age values are generated as \code{round(seq(min(age), max(age)*1.5, length.out = 100), 0)}.
When \code{targetCovar != "cover"}, "cover" may be fixed at \code{maxCover}. See \code{fixMaxCover}.}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{fixMaxCover}{logical. If \code{TRUE} and \code{targetCovar != "cover"}, cover is
not averaged and is fixed to \code{maxCover}.}

\item{xCovar}{the variable shown in the x axis. Defaults to "age". When \code{xCovar == "age"}
the output plots are not true "Partial Effects" plots. Instead, they show the variation
is \code{B} values across values of \code{targetCovar} for each value of age.}

\item{showQuantiles}{controls whether quantile predictions will be shown. If
"allQuantiles", quantile values (5\%, 25\%, 50\%, 75\%, 95\%) of B will be
plotted as blue lines, with their respective asymptote values (quantile values
at maximum age) as dashed lines. If "maximum" the 100\% quantile will be plotted.
If "none", only the 50\% quantile line (average prediction) is plotted. Quantiles
are always calculated at \code{max(age)}.}

\item{plotTitle}{character. Passed to \code{title} in \code{\link[ggplot2:labs]{ggplot2::labs()}}}

\item{nonLinModelQuoted}{a named list of non-linear equations as a \code{call}
(quoted expression) passed to \code{mle2(minuslog1)}. See \code{?mle}.
Accepts equations with three parameters 'A', 'p' and 'k'. List names and
length must the same as in \code{mll}.}

\item{linModelQuoted}{A named list of lists of linear equations/modes relating each
parameter ('A', 'p' and 'k') with a set of covariates. A \code{call}
(quoted expression) passed to \code{mle2(..., parameters)}. Note that for the
purpose of tree growth, the linear equation determining 'A' should include a
'cover' predictor indicating the tree cover or dominance in the stand. Should be
scaled between 0 and \code{maxCover}. List names and length must the same as in \code{mll}.}

\item{fun}{passed to \code{.MLLMaxBPartialPlotData}.}

\item{plotCIs}{should confidence intervals be calculated and plotted?}
}
\description{
Plots the maximum biomass estimates along an age gradient as a function
of another target covariate, with all others held at average values.
}
\details{
Note that the original data, not the predicted values is shown.
}
\seealso{
\code{\link[bbmle:mle2]{bbmle::mle2()}}
}
