% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionBiomass.R
\name{partitionBiomass}
\alias{partitionBiomass}
\title{Partition biomass according to cover estimates}
\usage{
partitionBiomass(x = 1, pixelCohortData)
}
\arguments{
\item{x}{The ratio for deciduous cover:biomass, where conifer cover:biomass = 1}

\item{pixelCohortData}{A full \code{pixelCohortData} object (i.e., not \code{cohortData})}
}
\description{
This function will partition \code{totalBiomass} into each cohort.
It will discount deciduous cover, if \code{x < 1}.
}
