% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genericMap}
\alias{plot_genericMap}
\title{Plot function for generic SpatRaster objects with Plots}
\usage{
plot_genericMap(x, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{SpatRaster for current or cumulative harvest}

\item{title}{character, the plot title}

\item{subtitle}{character, the plot subtitle}
}
\value{
ggplot object
}
\description{
Plot function for generic SpatRaster objects with Plots
}
