% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeCanadianForestRasters.R
\name{plot_raster_statistics}
\alias{plot_raster_statistics}
\title{Plot Raster Statistics and Maps by Polygon}
\usage{
plot_raster_statistics(
  raster,
  polygons = NULL,
  polygon_id = NULL,
  region_type = "ecoregion",
  region_ids = NULL,
  counts_df = NULL,
  aggregate_factor = 1,
  remove_zeros = FALSE,
  raster_label = NULL,
  inset.Canada = TRUE,
  bin_width = 10,
  output_dir = "raster_figures",
  stats_csv = "raster_stats.csv"
)
}
\arguments{
\item{raster}{A \code{SpatRaster} object. The raster containing values to be summarized.}

\item{polygons}{Optional. A \code{SpatVector} or \code{sf} object of polygons to subset the raster. Defaults to Canadian Ecoregions.}

\item{polygon_id}{Name of the column in \code{polygons} to use as polygon ID.}

\item{region_type}{Optional. \code{"ecoregion"} or \code{"ecozone"} if \code{polygons} is not provided.}

\item{region_ids}{Optional. Vector of region IDs to filter.}

\item{counts_df}{Optional. A \code{data.frame} with precomputed counts (\code{RegionID}, \code{value}, \code{count}),
such as generated by \code{terra::extract()}. Polygon ID must match that from polygons.
If provided, the function will use these values and skip calculating counts.}

\item{aggregate_factor}{Optional. Integer factor to aggregate raster for plotting. Default is \code{1} (no aggregation).}

\item{remove_zeros}{Logical. Removes zeroes from raster and counts for plotting.
Useful to eliminate non-forested areas from forest rasters.}

\item{raster_label}{Optional. Character label for raster used in plots. Defaults to raster layer name.}

\item{inset.Canada}{Logical. Generates a small inset map of plotted polygons within Canada. Default to TRUE.}

\item{bin_width}{Numeric. Bin size for histogram plots. Defaults to 10.}

\item{output_dir}{Directory to save figures. Default is \code{"raster_figures"}.}

\item{stats_csv}{Path to save the statistics CSV. Default is \code{"raster_stats.csv"}.}
}
\value{
A \code{data.frame} of computed statistics.
}
\description{
This function calculates and returns statistics (min, max, median, mean,
25th percentile, 75th percentile, and proportion of zeroes) for a raster,
and creates and saves figures for each polygon in a vector layer, including
a histogram of raster values, a map of raster values within the polygon, and
a summary of statistics . Optionally, zero values can be removed from the
analysis, and the raster can be aggregated to a lower resolution to speed up plotting.
}
