% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-summaries.R
\name{plotLeadingSpecies}
\alias{plotLeadingSpecies}
\title{Leading species plots}
\usage{
plotLeadingSpecies(
  studyAreaName,
  climateScenario,
  Nreps,
  years,
  outputDir,
  treeSpecies,
  defineLeading = .defineLeading,
  leadingPercentage = 0.8,
  treeType = NULL,
  rasterToMatch,
  figurePath = file.path(outputDir, "figures")
)
}
\arguments{
\item{studyAreaName}{character string giving the study area name}

\item{climateScenario}{character string specifying the name of a CIMP6 climate scenario,
including SSP, formatted as in \code{ClimateNA}, using underscores as separator
(e.g., 'CanESM5_SSP370').}

\item{Nreps}{the number of simulation replicates/run used to produce summary figures.
NOTE: \code{mclapply} is used internally, so you should set \code{options(mc.cores = nReps)}
to take advantage of parallel processing.}

\item{years}{TODO}

\item{outputDir}{Path specifying the directory to which outputs figures/objects should be saved.}

\item{treeSpecies}{TODO}

\item{defineLeading}{TODO}

\item{leadingPercentage}{TODO}

\item{treeType}{TODO}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{figurePath}{Directory to which the \code{.png} figure is saved.
Defaults to \verb{figures/} under \code{outputDir}; a SpaDES module would pass
\code{SpaDES.core::figurePath(sim)}.}
}
\value{
list of file names corresponding to the figures and/or objects written to disk

list of filepaths corresponding to the images and/or objects written to disk
}
\description{
Plot effects on conifer-to-deciduous or deciduous-to-conifer conversions.
}
