% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{plotRast}
\alias{plotRast}
\title{PLOT RASTER FUNCTION}
\usage{
plotRast(ras, plotTitle)
}
\arguments{
\item{ras}{\code{SpatRaster} or \code{RasterLayer}}

\item{plotTitle}{character. A title for the plot passed to
\code{ggplot::labs(title = plotTitle)}}
}
\description{
A function that creates a \code{ggplot} of a raster
layer. Can be used with \code{SpaDES.core::Plots}
}
