% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{polygonClean}
\alias{polygonClean}
\title{Do an arbitrary set of operations on a polygon}
\usage{
polygonClean(poly, fn = NULL, type = NULL, ...)
}
\arguments{
\item{poly}{A polygon object, or a character string identifying the shapefile
path to load, and clean.}

\item{fn}{A function identifying the type of cleaning to do.}

\item{type}{If fn is not known, an character string can be specified to
identify which \code{fn} to use.
This MUST be a known type for this function.}

\item{...}{Passed to \code{fn}}
}
\description{
Do an arbitrary set of operations on a polygon
}
