% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepEcoregions.R
\name{prepEcoregions}
\alias{prepEcoregions}
\title{Prepare ecoregions objects}
\usage{
prepEcoregions(
  ecoregionRst = NULL,
  ecoregionLayer,
  ecoregionLayerField = NULL,
  rasterToMatchLarge,
  rstLCCAdj,
  pixelsToRm,
  cacheTags
)
}
\arguments{
\item{ecoregionRst}{an optional raster object that could be passed to \code{sim},
representing ecoregions}

\item{ecoregionLayer}{an \code{sf} polygons object representing ecoregions}

\item{ecoregionLayerField}{optional. The field in \code{ecoregionLayer} that represents ecoregions.}

\item{rasterToMatchLarge}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatchLarge} should correspond to \code{studyAreaLarge} if used.}

\item{rstLCCAdj}{\code{RasterLayer} representing land cover adjusted for non-forest classes}

\item{pixelsToRm}{a vector of pixels to remove}

\item{cacheTags}{\code{UserTags} to pass to cache}
}
\description{
Prepare ecoregions objects
}
