% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecozones.R
\name{prepEcozonesRst}
\alias{prepEcozonesRst}
\title{Prepare ecozones raster}
\usage{
prepEcozonesRst(url, destinationPath, studyArea = NULL, rasterToMatch = NULL)
}
\arguments{
\item{url}{character. ecozones shapefile url.
Default: \url{http://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip}.}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{studyArea}{A \code{SpatVector} or \code{sf} polygons object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}
}
\value{
\code{RasterLayer}
}
\description{
Download, rasterize, crop, mask, and reproject Canadian national ecozones shapefile.
}
