% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyArea.R
\name{prepGADM}
\alias{prepGADM}
\title{Get and prep GADM maps}
\usage{
prepGADM(country = "CAN", level = 1, prj, dPath)
}
\arguments{
\item{country}{Character string giving the country code (default \code{"CAN"}).}

\item{level}{Numeric giving the ADM level to use (default \code{1}).}

\item{prj}{Character string giving the projection to use.}

\item{dPath}{The destination path in which to save the data file.}
}
\description{
A wrapper around \code{\link[raster]{getData}} that does additional cleanup and reprojects.
}
