% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs_NTEMS.R
\name{prepInputs_NTEMS_DominantSpecies}
\alias{prepInputs_NTEMS_DominantSpecies}
\title{Obtain an Dominant species layer for a given year from NTEMS}
\usage{
prepInputs_NTEMS_DominantSpecies(
  year = 2011,
  destinationPath,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR",
  ...
)
}
\arguments{
\item{year}{stack of species layers rasters}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}

\item{...}{passed to \code{prepInputs}}
}
\value{
a \code{SpatRaster} with dominant species
}
\description{
Obtain an Dominant species layer for a given year from NTEMS
}
