% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs_NTEMS.R
\name{prepInputs_NTEMS_Nonforest}
\alias{prepInputs_NTEMS_Nonforest}
\title{reclassify non-flammable pixels that become flammable - herbaceous or shrubby - vegetation}
\usage{
prepInputs_NTEMS_Nonforest(rstLCC, endYear = 2018, lccToFix = 33)
}
\arguments{
\item{endYear}{NTEMS LCC year to use for correcting transition from bare to non-forest}

\item{...}{passed to prepInputs}
}
\value{
A \code{SpatRaster} with non-flammable pixels corrected if they become flammable non-forest
}
\description{
reclassify non-flammable pixels that become flammable - herbaceous or shrubby - vegetation
}
