% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{prepInputsCanDEM}
\alias{prepInputsCanDEM}
\title{Get digital elevation map of Canada}
\usage{
prepInputsCanDEM(studyArea, rasterToMatch, destinationPath)
}
\arguments{
\item{studyArea}{A \code{SpatVector} or \code{sf} polygons object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}
}
\value{
\code{SpatRaster}
}
\description{
Defaults to using 3 arcmin DEM of Canada in lonlat.
}
