% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyArea.R
\name{randomStudyArea}
\alias{randomStudyArea}
\title{Create default study areas for use with LandR modules}
\usage{
randomStudyArea(center = NULL, size = 10000, seed = NULL)
}
\description{
This simply re-exports \code{SpaDES.tools::randomStudyArea}
}
