% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{rasterRead}
\alias{rasterRead}
\title{Method to read raster}
\usage{
rasterRead(...)
}
\arguments{
\item{...}{passed to \code{terra::rast} or
\code{raster::raster}}
}
\value{
the function
}
\description{
TODO: Move to \code{reproducible}
}
