% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{methods}
\name{rasterToMatch}
\alias{rasterToMatch}
\alias{rasterToMatch,Raster-method}
\alias{rasterToMatch,SpatialPolygonsDataFrame-method}
\title{Extract or create a raster to match}
\usage{
rasterToMatch(x, ...)

\S4method{rasterToMatch}{Raster}(x, studyArea, ...)

\S4method{rasterToMatch}{SpatialPolygonsDataFrame}(x, studyArea,
  rasterToMatch, ...)
}
\arguments{
\item{x}{A Raster Layer with correct resolution and origin.}

\item{...}{Additional arguments}

\item{studyArea}{A SpatialPolygon* object that will be sent to \code{postProcess}.}

\item{rasterToMatch}{The raster to match in a \code{fasterize} call.}
}
\value{
A \code{RasterLayer} object.
}
\description{
This extracts or creates a new raster layer, whose intention is to be used as
the \code{rasterToMatch} argument in further \code{prepInputs} calls.
}
