% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.resample}
\alias{.resample}
\title{Resample}
\usage{
.resample(x, ...)
}
\description{
Imports the non-exported function \code{SpaDES.tools:::resample}.
}
\seealso{
\code{\link[SpaDES.tools]{resample}}
}
\keyword{internal}
