% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{rescale}
\alias{rescale}
\title{Rescale function (as in \code{scales::rescale})}
\usage{
rescale(x, to)
}
\arguments{
\item{x}{a \code{numeric} vector}

\item{to}{a \code{numeric} vector of length 2. The new range of values.}
}
\description{
This is a simple function copied from the scales package (almost the same). Too heavy
to use one simple function
}
