% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scheduleDisturbance}
\alias{scheduleDisturbance}
\title{Test whether disturbance should be scheduled}
\usage{
scheduleDisturbance(disturbanceLayer, currentYear)
}
\arguments{
\item{disturbanceLayer}{a \code{RasterLayer} object}

\item{currentYear}{time of simulation}
}
\value{
Logical indicating whether to schedule a disturbance event
}
\description{
Test whether disturbance should be scheduled
}
\examples{
\dontrun{
  doDisturbance <- scheduleDisturbance(sim$rstCurrentBurn, time(sim))
}
}
