% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{speciesEcoregionLatestYear}
\alias{speciesEcoregionLatestYear}
\title{Pull out the values from \code{speciesEcoregion} table for current time}
\usage{
speciesEcoregionLatestYear(speciesEcoregion, currentTime)
}
\arguments{
\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values.}

\item{currentTime}{The current simulation time e.g., \code{time(sim)}.}
}
\value{
The \code{speciesEcoregion} input object, but with data from only one year, the year
that is less than or equal to the \code{currentTime}
}
\description{
Pull out the values from \code{speciesEcoregion} table for current time
}
