% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{speciesEcoregionStack}
\alias{speciesEcoregionStack}
\title{Create Stacks of the speciesEcoregion content}
\usage{
speciesEcoregionStack(
  ecoregionMap,
  speciesEcoregion,
  columns = c("establishprob", "maxB", "maxANPP"),
  stackFilenames = NULL
)
}
\arguments{
\item{ecoregionMap}{The Factor RasterLayer of the ecoregionMap}

\item{speciesEcoregion}{The data.table with the speciesEcoregion information}
}
\description{
Create Stacks of the speciesEcoregion content
}
