% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPresentFromKNN.R
\name{speciesPresentFromSCANFI}
\alias{speciesPresentFromSCANFI}
\title{Make a species factor raster based on SCANFI Data}
\usage{
speciesPresentFromSCANFI(
  year = 2020,
  dPath = asPath("."),
  res = 2400,
  minPctCover = 10
)
}
\arguments{
\item{year}{One of 2000, 2010, or 2020. Default is 2020.}

\item{dPath}{A character string indicating where to download all the SCANFI layers}

\item{res}{The resolution (one dimension, in m) for the resulting raster}

\item{minPctCover}{An integer indicating what percent cover a species must have
in a pixel to be considered present in that pixel.}
}
\value{
A \code{SpatRaster} object with 2 layers: \code{"speciesPresent"} is a factor, with
a legend (i.e., it is numbers on a map, that correspond to a legend) and
\code{"numberSpecies"} which represents the number of species in each pixel.
}
\description{
This will download SCANFI species layer for forests of Canada, and make
a factor raster at resolution provided by \code{res} (larger is faster).
}
\examples{
\dontrun{
if (requireNamespace("googledrive", quietly = TRUE)) {
  # Make the dataset
  speciesPresent <- speciesPresentFromSCANFI(dPath = "~/data/SCANFI")

  # To upload this:
  speciesPresentRas <- speciesPresent[[1]]
  fn <- "SpeciesPresentInCanadianForests_SCANFI.tif"
  writeRaster(speciesPresentRas, file = fn)
  zipFn <- gsub(".tif", ".zip", fn)
  zip(files = dir(pattern = fn), zipFn)
  out <- googledrive::drive_put(zipFn)

  ## Get species list
  sa <- LandR::randomStudyArea(size = 1e11)
  species <- LandR::speciesInStudyArea(sa, dataSource = "SCANFI")
}
}

}
