% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\name{speciesTableUpdate}
\alias{speciesTableUpdate}
\title{Change species table of parameters/traits}
\usage{
speciesTableUpdate(species, speciesTable, sppEquiv = NULL, sppEquivCol = NULL)
}
\arguments{
\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.
Must have column \code{speciesCode}, with species names/IDs. The following is a
list of default trait columns:
\itemize{
\item "species" same as "speciesCode" -- species ID name
\item "speciesCode"
\item "Area" -- inherited from LANDIS-II default table, the Canadian ecoregion
from which traits where derived. Not used during the simulation
\item "firetolerance" -- module\emph{relative} (to other species) fire tolerance
\item "growthcurve" and "mortalityshape" -- growth curve shape parameters.
\item "longevity" -- maximum species age
\item "postfireregen" -- post-fire regeneration strategy ("serotiny", "resprout" or "none")
\item "resproutprob" -- probability of resprouting
\item "resproutage_min" -- minimum age at which species is capable of resprouting
\item "resproutage_max" -- maximum age at which species is capable of resprouting
\item "seeddistance_eff" -- effective dispersal distance
\item "seeddistance_max" -- maximum dispersal distance
\item "shadetolerance" -- \emph{relative} (to other species) shade tolerance
\item "sexualmature" -- age at sexual maturity
Known optional parameters added/needed by some modules (the user may add others for their own modules):
\item "inflationFactor" -- \code{Biomass_speciesParameters} module: inflation factor for \code{maxB}
\item "growthCurveSource" -- \code{Biomass_speciesParameters} module: how "growthcurve" was estimated
\item "mANPPproportion" -- \code{Biomass_speciesParameters} module: multiplication factor to calculate \code{maxANPP} from \code{maxB}
\item "thermokarsttol" -- \code{Biomass_disturbances} module: proportion of biomass surviving after thermokarst (i.e. permafrost thaw). Applied equally across cohorts.
Parameters inherited from LANDIS-II default table, but not used in LandR at the moment:
\item "leaflongevity"
\item "wooddecayrate"
\item "leafLignin"
\item "hardsoft"
Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015)
for further detail.
}}

\item{speciesTable}{A species traits table, with \strong{at least} the following columns:
\itemize{
\item \code{speciesCode} an character representation of species;
\item \code{Area} a character of geographical area within a species range for which trait values are relevant;
\item \code{longevity} species longevity in years, \code{sexualmature} age in years at sexual maturity;
\item \code{shadetolerance} numeric value between 1-5 of relative shade tolerance (with respect to other species);
\item \code{seeddistance_eff} the "effective" seed dispersal distance;
\item \code{seeddistance_max} a numeric with the maximum seed dispersal distance;
\item \code{mortalityshape} and \code{growthcurve}: growth curve shape parameters.
Other columns (e.g. fire-related traits) can also be included depending on LandR modules in use.
Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015)
for further detail.
}}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}
}
\value{
An updated species \code{data.table}
}
\description{
Changes longevity and shade tolerance values in the species table.
Longevity values are changed to follow Burton & Cumming (1995) for the following species:
\itemize{
\item \emph{Abies balsamea};
\item \emph{Abies lasiocarpa};
\item \emph{Betula papyrifera};
\item \emph{Larix laricina};
\item \emph{Larix occidentalis};
\item \emph{Picea engelmannii};
\item \emph{Picea glauca};
\item \emph{Picea mariana};
\item \emph{Pinus banksiana};
\item \emph{Pinus contorta};
\item \emph{Pinus resinosa};
\item \emph{Pinus strobus};
\item \emph{Populus balsamifera v. balsamifera};
\item \emph{Populus tremuloides};
\item \emph{Pseudotsuga menziesii var. glauca};
\item \emph{Pseudotsuga menziesii};
\item \emph{Thuja plicata};
\item \emph{Tsuga heterophylla};
\item \emph{Tsuga mertensiana x heterophylla};
}
}
\details{
and only for the following \code{speciesTable$Area}s:
\itemize{
\item Boreal Shield West (BSW);
\item Boreal Plains (BP);
\item Montane Cordillera (MC).
}

Note that BSW and BP areas correspond more closely to the region considered in Table 2 of
Burton & Cumming (1995), while MC will correspond to both tables.

Of the above species, shade tolerance values are changed for \emph{Abies spp}, \emph{Picea spp},
and \emph{Tsuga spp} to better reflect the western boreal.

When different longevity/shade tolerance trait values exist for a given species, the minimum
value across \code{Area}'s (BSW, BP, MC) is kept.

ATTENTION: if none of species in \code{species} are from BSW, BP or MC area this function will not
change any values.

All other species/Area trait values follow Dominic Cyr and Yan Boulanger's trait values
(\url{https://raw.githubusercontent.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv}).
}
