% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sppColors}
\alias{sppColors}
\alias{sppColours}
\title{Create species colour vector from a \code{sppEquiv} table}
\usage{
sppColors(sppEquiv, sppEquivCol, newVals = NULL, palette = "Accent")
}
\arguments{
\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}

\item{newVals}{An optional character vector of extra names to use, e.g., "Mixed".}

\item{palette}{An \pkg{RColorBrewer} palette, e.g., "Accent".
Can get \pkg{RColorBrewer} palette names from \code{rownames(RColorBrewer::brewer.pal.info)}.}
}
\value{
A named vector of colour codes, where the names are the species names
plus any extra names passed with \code{newVals}.
}
\description{
Create species colour vector from a \code{sppEquiv} table
}
