% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{standAgeMapGenerator}
\alias{standAgeMapGenerator}
\title{Produce stand age map based on \code{cohortData}}
\usage{
standAgeMapGenerator(
  cohortData,
  pixelGroupMap,
  weight = "biomass",
  doAssertion = getOption("LandR.assertions", FALSE)
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{weight}{variable by which to weight \code{cohortData} ages. one of \code{"biomass"} or \code{NA}.
\code{NA} means use max unweighted age.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\value{
raster of the same type as \code{pixelGroupMap}.
}
\description{
Produce stand age map based on \code{cohortData}
}
