% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyArea.R
\name{studyAreaEco}
\alias{studyAreaEco}
\title{Ecological boundaries within a study area}
\usage{
studyAreaEco(
  studyArea = NULL,
  destinationPath = tempdir(),
  type = c("ECODISTRICT", "ECOREGION", "ECOPROVINCE", "ECOZONE")
)
}
\arguments{
\item{studyArea}{A \code{SpatVector} or \code{sf} polygons object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{type}{character. The polygon type to use:
one of "ECODISTRICT", "ECOREGION", "ECOPROVINCE", or "ECOZONE".}
}
\value{
spatial polygons object of the same class as \code{studyArea}
}
\description{
Extend a study area boundary to include the full ecological boundary polygons intersecting it.
}
\examples{
## using SpatVector objects
sa <- randomStudyArea(size = 1e9)
sa_eco <- studyAreaEco(studyArea = sa)

## using sf objects
sa_sf <- sf::st_as_sf(sa)
sa_eco_sf <- studyAreaEco(studyArea = sa_sf)

if (interactive()) {
  ggplot() +
    geom_sf(data = sa_eco_sf, fill = "gray") +
    geom_sf(data = sa_sf, fill = "violet", alpha = 0.3)
}

}
