% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{subsetDT}
\alias{subsetDT}
\title{Subset a \code{data.table} with random subsampling within \code{by} groups}
\usage{
subsetDT(DT, by, doSubset = TRUE)
}
\arguments{
\item{DT}{A \code{data.table}}

\item{by}{Character vector of column names to use for groups}

\item{doSubset}{Logical or numeric indicating the number of subsamples to use}
}
\description{
Subset a \code{data.table} with random subsampling within \code{by} groups
}
