% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeCanadianForestRasters.R
\name{SummarizeCanadianRasterCounts}
\alias{SummarizeCanadianRasterCounts}
\title{Extract Raster Value Counts by Polygon using Ecoregions or Ecozones.}
\usage{
SummarizeCanadianRasterCounts(raster, polygon = NULL, use_ecozones = FALSE)
}
\arguments{
\item{raster}{SpatRaster. The raster to extract values from.}

\item{polygon}{Optional. A custom polygon to use for extraction. If NULL, defaults to ER.}

\item{use_ecozones}{Logical. If TRUE, aggregates results by ecozone (\code{ZONE_NAME}) using ER.}
}
\value{
A data.frame with columns for polygon ID, raster value, and count.
}
\description{
This function extracts counts of raster values within polygons using \code{terra::extract()}.
It defaults to using ecoregions (\code{ER}) but can also aggregate results by ecozones (\code{EZ})
or use a custom polygon. Due to ecozone size, if ecozones are selected extraction is performed over ecoregions
and then aggregated by Ecozone (\code{ZONE_NAME}). The raster is cropped to the Canada census boundaries before extraction.
}
