% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{sumRastersBySpecies}
\alias{sumRastersBySpecies}
\title{Function to sum rasters of species layers}
\usage{
sumRastersBySpecies(speciesLayers, layersToSum, filenameToSave, newLayerName)
}
\arguments{
\item{speciesLayers}{stack of species layers rasters}

\item{layersToSum}{names/indices of layers to be summed - optional}

\item{filenameToSave}{file path to save output raster}

\item{newLayerName}{name of the output raster layer}
}
\description{
Function to sum rasters of species layers
}
