% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{testEquivalentMetadata}
\alias{testEquivalentMetadata}
\title{Test that metadata of 2 or more objects is the same}
\usage{
testEquivalentMetadata(...)
}
\arguments{
\item{...}{2 or more of the same type of object to test
for equivalent metadata}
}
\description{
Currently, only Raster class has a useful method. Defaults to
\code{all(sapply(list(...)[-1], function(x) identical(list(...)[1], x)))}
}
