% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toSentenceCase}
\alias{toSentenceCase}
\title{Convert a character string to sentence case}
\usage{
toSentenceCase(x)
}
\arguments{
\item{x}{A character string.}
}
\description{
Convert a character string to sentence case
}
