% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTable.R
\name{updateSpeciesTable}
\alias{updateSpeciesTable}
\title{Customize species trait table values}
\usage{
updateSpeciesTable(speciesTable, params)
}
\arguments{
\item{speciesTable}{The species traits table}

\item{params}{A named list (of parameters) of named lists (by species), with species
traits overrides (e.g., \code{list(seeddistance_eff = list(Abie_sp = 25))}).}
}
\description{
Customize species trait table values
}
\author{
Alex Chubaty and Ceres Barros
}
