% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{user}
\alias{user}
\title{Determine name of current user}
\usage{
user(username = NULL)
}
\arguments{
\item{username}{Character.}
}
\value{
Logical indicating whether the current user matches \code{username};
or, if \code{username} is NULL, a character string indicating the current user.
}
\description{
Determine name of current user
}
