% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedDispersalLANDIS.R
\docType{methods}
\name{WardFast}
\alias{WardFast}
\alias{WardVec}
\title{WardFast Seed Dispersal kernel}
\format{
An object of class \code{expression} of length 1.
}
\usage{
WardFast

WardVec(dist, cellSize, effDist, maxDist, k, b, algo = 2)
}
\arguments{
\item{algo}{Either 1 or 2. 2 is faster and is default. 1 is "simpler code" as it
uses `ifelse`}
}
\description{
A probability distribution used in LANDIS-II.
}
\author{
Eliot McIntire

Eliot McIntire
}
\keyword{datasets}
