% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedDispersalLANDIS.R
\docType{methods}
\name{WardKernel}
\alias{WardKernel}
\alias{Ward}
\title{Ward Dispersal Kernel -- vectorized, optimized for speed}
\usage{
Ward(dist, cellSize, effDist, maxDist, k, b, algo = 2)
}
\arguments{
\item{dist}{A vector of distances to evaluate kernel against}

\item{cellSize}{A numeric, length 1, of the cell resolution (e.g., res(raster))}

\item{effDist}{A vector of effective distance (parameter in kernel),
with same length as \code{dist}}

\item{maxDist}{A vector of maximum distance (parameter in kernel),
with same length as \code{dist}}

\item{k}{A parameter in the kernel}

\item{b}{A parameter in the kernel}

\item{algo}{Either 1 or 2. 2 is faster and is default. 1 is "simpler code" as it
uses `ifelse`}
}
\description{
A probability distribution used in LANDIS-II.
}
\author{
Eliot McIntire
}
