% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{writeCloud}
\alias{writeCloud}
\title{Basic tool for using cloud-based caching}
\usage{
writeCloud(object, digest, cloudFolderID = NULL, checksums)
}
\arguments{
\item{object}{The R object to write to cloud}

\item{digest}{The hash of the input arguments, outputted from \code{checkCloud}}

\item{cloudFolderID}{The google folder ID where a new object should
be written}

\item{checksums}{A \code{data.table} that is outputted from \code{checkCloud} that
is the the checksums file}
}
\description{
Very experimental
}
\seealso{
\code{\link{checkCloud}}
}
