% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{adjustAgeToLongevity}
\alias{adjustAgeToLongevity}
\title{Reduces the age of cohorts that exceed their \verb{longevity x adjustmentFactor}}
\usage{
adjustAgeToLongevity(pixelCohortData, longevity, adjustmentFactor)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{longevity}{A data.table with the longevity of each species.}

\item{adjustmentFactor}{A numeric controlling the proportion of species longevity
that cohort ages cannot exceed.}
}
\value{
A \code{cohortData} data.table with corrected ages.
}
\description{
Reduces the age of cohorts that exceed their \verb{longevity x adjustmentFactor}
}
