% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{asInteger}
\alias{asInteger}
\title{Convert numeric values to rounded integers}
\usage{
asInteger(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer vector of length \code{x}, rounded to zero decimal places
  prior to \code{as.integer}
}
\description{
Simple wrapper around \code{as.integer} to round, rather than truncate, values.
}
