% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{asInteger}
\alias{asInteger}
\title{Convert numeric values to rounded integers}
\usage{
asInteger(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer vector of length \code{x}, rounded to zero decimal places
  prior to \code{as.integer}
}
\description{
Essentially a wrapper around round, rather than truncate which is what \code{as.integer}
does. Internally, this is simply \code{as.integer(floor(x + 0.5))}.
}
\note{
Values ending in .5 will be rounded up, whether positive or negative, "round up". This
is different than \code{round}.
}
\examples{
x <- seq(-2, 2, 0.25)
data.frame(dbl = x, int = asInteger(x))
}
