% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert1}
\alias{assert1}
\alias{assert2}
\alias{assertSppMaxBMaxANPP}
\alias{assertUniqueCohortData}
\alias{assertERGs}
\alias{assertColumns}
\alias{assertCohortData}
\alias{assertCohortDataERG}
\alias{assertPixelCohortData}
\alias{assertSpeciesPlotLabels}
\alias{assertFireToleranceDif}
\alias{assertSpeciesLayers}
\alias{assertRstLCChange}
\alias{assertSpeciesEcoregionCohortDataMatch}
\alias{assertPixelCohortDataValid}
\alias{assertRepsAllCohortData}
\alias{assertStandAgeMapAttr}
\alias{assertCohortDataAttr}
\alias{assertSppVectors}
\alias{assertSpeciesTableRaw}
\alias{assertSpeciesTable}
\title{Assertions}
\usage{
assert1(
  cohortData34to36,
  cohortData,
  rmZeroBiomassQuote,
  classesToReplace = 34:36,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assert2(
  cohortDataNo34to36,
  classesToReplace = 34:36,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSppMaxBMaxANPP(
  speciesEcoregion,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertUniqueCohortData(
  cohortData,
  columns,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertERGs(
  ecoregionMap,
  cohortData,
  speciesEcoregion,
  minRelativeB,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertColumns(
  obj,
  colClasses,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertCohortData(
  cohortData,
  pixelGroupMap,
  maxExpectedNumDiverge = 1,
  message = "",
  doAssertion = getOption("LandR.assertions", TRUE),
  verbose = getOption("LandR.verbose", TRUE),
  cohortDefinitionCols = LandR::cohortDefinitionCols()
)

assertCohortDataERG(
  cohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertPixelCohortData(
  pixelCohortData,
  pixelGroupMap,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSpeciesPlotLabels(
  speciesNames,
  sppEquiv,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertFireToleranceDif(
  burnedPixelCohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSpeciesLayers(
  speciesLayers,
  thresh,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertRstLCChange(
  rstLCChange,
  rasterToMatch,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSpeciesEcoregionCohortDataMatch(
  cohortData,
  speciesEcoregion,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertPixelCohortDataValid(
  standCohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertRepsAllCohortData(
  allCohortData,
  reps,
  years,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertStandAgeMapAttr(
  standAgeMap,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertCohortDataAttr(
  cohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSppVectors(
  sppEquiv = NULL,
  sppNameVector = NULL,
  sppColorVect = NULL,
  sppEquivCol = NULL,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSpeciesTableRaw(
  speciesTableRaw,
  doAssertion = getOption("LandR.assertions", TRUE)
)

assertSpeciesTable(
  speciesTable,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData34to36}{A \code{cohortData} \code{data.table} with only the
pixels what were LCC 34:36}

\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{rmZeroBiomassQuote}{An expression to evaluate, in the form of \code{quote(B>0)},
used to select cohorts with biomass.}

\item{classesToReplace}{Integer vector of classes that are are to be replaced,
e.g., 34, 35, 36 on LCC2005, which are burned young, burned 10 year, and cities.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{cohortDataNo34to36}{A \code{cohortData} \code{data.table} with only the
pixels what were LCC 34:36}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-\code{ecoregion}-specific species trait values.
Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical
characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.
May contain columns \code{inflationFactor}  (used to adjust \code{maxB}) and \code{mANPPproportion}
(used to calculate \code{maxANPP}).}

\item{columns}{Vector of column names on which to test for unique \code{cohortData}}

\item{ecoregionMap}{The \code{ecoregionMap}, a raster of all the unique ecoregion groupings,
which group together pixels of similar biophysical characteristics.}

\item{minRelativeB}{a \code{data.frame} with  the cut points to classify stand shadiness.}

\item{obj}{A \code{data.frame}- or \code{data.table}-like object}

\item{colClasses}{A named vector of column classes, where the names are the column names}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{maxExpectedNumDiverge}{A numeric, length 1, indicating by how many they
can diverge. Default 1.}

\item{message}{An optional message to print. This may help identify where this function
was called.}

\item{verbose}{Controls message output. Defaults to \code{getOption("LandR.verbose")}}

\item{cohortDefinitionCols}{the columns in \code{cohortData} that define unique cohorts
(default: \code{c("pixelGroup", "speciesCode", "age")}).}

\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{speciesNames}{A vector of species names for which the labels will be checked}

\item{sppEquiv}{an table with a column containing species names}

\item{burnedPixelCohortData}{An expanded \code{cohortData} \code{data.table} with pixel-level
cohort information on burnt pixels and the following (optional) columns:
\code{severity} - fire severity in that pixel calculated based on fire behaviour properties;
\code{firetolerance} - species-level fire tolerance;
\code{severityToleranceDif} - the difference between \code{severity} and \code{firetolerance}.}

\item{speciesLayers}{A \code{RasterStack} or \code{RasterLayer} that
should contain species cover data in the study area}

\item{thresh}{the minimum number of pixels where the species must have
\code{biomass > 0} to be considered present in the study area.
Defaults to 1.}

\item{rstLCChange}{a raster layer indicating pixels were land-use change occurred as 1s}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{standCohortData}{A \code{data.table} with simulated and observed stand data for validation}

\item{allCohortData}{A \code{data.table} with all simulated \code{cohortData} to use for
validation}

\item{reps}{repetition ids}

\item{years}{years}

\item{standAgeMap}{A \code{rasterLayer} of stand ages with attribute \code{"imputedPixID"}}

\item{sppNameVector}{A character vector of species to use. These species must all
be from one naming convention, i.e., from one column in the \code{sppEquiv}.}

\item{sppColorVect}{A named vector of colours to use for plotting.
The names must conform with species name convention used
(see \link{sppEquivalencies_CA} for possible naming
conventions) and should also contain a colour for 'Mixed',
when a mixed forest type is supported  (see \code{vegLeadingProportion}
parameter in \code{\link[=vegTypeGenerator]{vegTypeGenerator()}} for details on mixed
forest types).}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{speciesTableRaw}{raw species traits \code{data.table} (see \code{\link[=getSpeciesTable]{getSpeciesTable()}})}

\item{speciesTable}{species traits \code{data.table} (see \code{\link[=prepSpeciesTable]{prepSpeciesTable()}})}
}
\description{
\itemize{
\item \code{assert1}: Assert that \code{ecoregionCodes} that were replaced, were correctly identified;
}

This is the full \code{pixelCohortData}, not the collapsed one.
}
