% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert1}
\alias{assert1}
\alias{assertUniqueCohortData}
\alias{assertERGs}
\alias{assertColumns}
\alias{assertCohortData}
\alias{assertPixelCohortData}
\alias{assertSpeciesPlotLabels}
\title{Assertions}
\usage{
assert1(cohortData34to36, cohortData,
  doAssertion = getOption("LandR.assertions", TRUE))

assertUniqueCohortData(cohortData, columns,
  doAssertion = getOption("LandR.assertions", TRUE))

assertERGs(ecoregionMap, cohortData, speciesEcoregion, minRelativeB,
  doAssertion = getOption("LandR.assertions", TRUE))

assertColumns(obj, colClasses,
  doAssertion = getOption("LandR.assertions", TRUE))

assertCohortData(cohortData, pixelGroupMap, sim,
  maxExpectedNumDiverge = 1, message = "",
  doAssertion = getOption("LandR.assertions", TRUE),
  verbose = getOption("LandR.verbose", TRUE))

assertPixelCohortData(pixelCohortData, pixelGroupMap,
  doAssertion = getOption("LandR.assertions", TRUE))

assertSpeciesPlotLabels(speciesNames, sppEquiv,
  doAssertion = getOption("LandR.assertions", TRUE))
}
\arguments{
\item{cohortData34to36}{A \code{cohortData} \code{data.table} with only the
pixels what were lcc 34:36}

\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{columns}{Vector of column names on which to test for unique cohortData}

\item{ecoregionMap}{The \code{ecoregionMap}, a raster of all the unique groupings}

\item{speciesEcoregion}{The \code{speciesEcoregion} \code{data.table}}

\item{minRelativeB}{TODO: add description}

\item{obj}{A data.frame or data.table-like object}

\item{colClasses}{A named vector of column classes, where the names are the column names}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group
number that correspondsd exactly to \code{pixelGroup} column in
\code{cohortData}}

\item{sim}{If the simList is included, then the browser() call will be more useful}

\item{maxExpectedNumDiverge}{A numeric, length 1, indicating by how many they
can diverge. Default 1.}

\item{message}{An optional message to print. This may help identify where this function
was called.}

\item{verbose}{Controls message output. Defaults to \code{getOption("LandR.verbose")}}

\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{speciesNames}{A vector of species names for which the labels will be checked}

\item{sppEquiv}{table with species name equivalencies between the
kNN format and the final naming format.
See \code{data("sppEquivalencies_CA", "LandR")}.}
}
\description{
This is the full pixelCohortData, not the collapsed one
}
\note{
TODO
}
