% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertPostPartialDist}
\alias{assertPostPartialDist}
\title{Assert post-fire disturbance mortality and regeneration}
\usage{
assertPostPartialDist(
  cohortDataOrig,
  pixelGroupMapOrig,
  cohortDataNew,
  pixelGroupMapNew,
  postDistPixelCohortData,
  distrbdPixelCohortData,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortDataOrig}{object used in as \code{updateCohortData(cohortData = .)}}

\item{pixelGroupMapOrig}{object used in as \code{updateCohortData(pixelGroupMap = .)}}

\item{cohortDataNew}{\code{cohortData} output from \code{updateCohortData}}

\item{pixelGroupMapNew}{\code{pixelGroupMap} output from \code{updateCohortData}}

\item{postDistPixelCohortData}{object used in as \code{updateCohortData(newPixelCohortData = .)}}

\item{distrbdPixelCohortData}{\code{cohortData}-like object containing all dead, surviving and new cohorts
(i.e. activated by serotiny/resprouting)}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Assert post-fire disturbance mortality and regeneration
}
