% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertSpeciesEcoregionCohortDataMatch}
\alias{assertSpeciesEcoregionCohortDataMatch}
\title{Assert that the cohortData speciesEcoregion have matching clases}
\usage{
assertSpeciesEcoregionCohortDataMatch(
  cohortData,
  speciesEcoregion,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{A cohortData object}

\item{speciesEcoregion}{A speciesEcoregion object}
}
\description{
Specifically, whether all combinations of ecoregionGroup and speciesCode are in both objects, no more
no less.
}
