% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertSpeciesEcoregionCohortDataMatch}
\alias{assertSpeciesEcoregionCohortDataMatch}
\title{Assert that the \code{cohortData} \code{speciesEcoregion} have matching classes}
\usage{
assertSpeciesEcoregionCohortDataMatch(
  cohortData,
  speciesEcoregion,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{A \code{cohortData} object}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-{ecoregion}-specific species
trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar
biophysical characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\description{
Specifically, whether all combinations of \code{ecoregionGroup} and \code{speciesCode} are in
both objects, no more no less.
}
