% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{assignClimateEffect}
\alias{assignClimateEffect}
\title{\code{assignClimateEffect}}
\usage{
assignClimateEffect(...)
}
\arguments{
\item{...}{additional arguments that are passed to LandR.CS}
}
\value{
null (zero) estimate of growth due to climate
}
\description{
\code{assignClimateEffect}
}
