% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.R
\name{assignLightProb}
\alias{assignLightProb}
\title{Assign light probability}
\usage{
assignLightProb(sufficientLight, newCohortData)
}
\arguments{
\item{sufficientLight}{TODO: description needed}

\item{newCohortData}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
Assign light probability
}
