% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permafrost.R
\name{assignPermafrost}
\alias{assignPermafrost}
\title{Create permafrost P/A layer}
\usage{
assignPermafrost(
  gridPoly,
  ras,
  saveOut = TRUE,
  saveDir = NULL,
  id = NULL,
  IDcol = "OBJECTID",
  rasClass = 1L
)
}
\arguments{
\item{gridPoly}{a \code{SpatVector}, \code{PackedSpatVector} or \code{character}.
The full polygon layer containing information about \% permafrost
(as a \code{SpatVector} or \code{PackedSpatVector}) or the file name to that
layer.}

\item{ras}{a \code{SpatRaster}, \code{PackedSpatRaster} or \code{character}.
The land-cover suitability for permafrost raster layer (as a
\code{SpatRaster} or \code{PackedSpatRaster}) or the file name to that layer.
Suitable cells are coded as \code{rasClass}.}

\item{saveOut}{logical. Should the processed rasters for the focal
polygon be saved (as the file name returned) or directly returned?
If parallelising, choose to save, as \code{SpatRasters} cannot be serialized.}

\item{saveDir}{character. The directiory to save output rasters.}

\item{id}{character or numeric. The polygon ID in \code{gridPoly[IDcol]}
to process (the focal polygon).}

\item{IDcol}{character. Column name in \code{gridPoly} containing polygon IDs.}

\item{rasClass}{Class}
}
\value{
a file name or the permafrost presence/absence raster for the focal
polygon.
}
\description{
Creates a permafrost presence/absence raster layer based
on information about land-cover suitability for permafrost
(a raster layer at high resolution) and \% of permafrost
present (a polygon layer that groups several cells of the
land-cover layer, i.e. at a larger spatial scale). The function
has been designed to be looped over polygons, and so only processes
only polygon at a time.
}
