% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomodWrappers.R
\name{biomodProjWrapper}
\alias{biomodProjWrapper}
\title{BIOMOD_Projection wrapper}
\usage{
biomodProjWrapper(
  bm.mod,
  proj.name = "testProj",
  new.env = NULL,
  new.env.xy = NULL,
  ...
)
}
\arguments{
\item{bm.mod}{passed to \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}. If not supplied
the data used to fit the model will be used.}

\item{proj.name}{passed to \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}}

\item{new.env}{passed to \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}}

\item{new.env.xy}{passed to \link[biomod2:BIOMOD_Projection]{biomod2::BIOMOD_Projection}}

\item{...}{passed to \link[reproducible:Cache]{reproducible::Cache}}
}
\description{
BIOMOD_Projection wrapper
}
