% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireSeverity.R, R/maps.R
\name{calcSeverityB}
\alias{calcSeverityB}
\title{Calculate fire severity}
\usage{
calcSeverityB(cohortData, burnedPixelCohortData)

calcSeverityB(cohortData, burnedPixelCohortData)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{burnedPixelCohortData}{An expanded \code{cohortData} \code{data.table} with pixel-level
cohort information on burnt pixels and the following (optional) columns:
\code{severity} - fire severity in that pixel calculated based on fire behaviour properties;
\code{firetolerance} - species-level fire tolerance;
\code{severityToleranceDif} - the difference between \code{severity} and \code{firetolerance}.}
}
\value{
\code{data.table} with columns \code{pixelIndex},
  \code{pixelGroup} and  \code{severityB}

\code{data.table} with columns \code{pixelIndex},
  \code{pixelGroup} and  \code{severityB}
}
\description{
Calculates fire severity as the loss of pre-fire to
  post-fire biomass.

Calculates fire severity as the loss of pre-fire to
  post-fire biomass.
}
\note{
if \code{burnedPixelCohortData} does not have a \code{B}
   column, the fire is assumed to be stand replacing (i.e.
   we assume B to be 0 across all pixels/cohorts in
   \code{burnedPixelCohortData})
}
