% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light-shade.R
\name{calcSiteShade}
\alias{calcSiteShade}
\title{Calculate site shade}
\usage{
calcSiteShade(currentTime, cohortData, speciesEcoregion, minRelativeB)
}
\arguments{
\item{currentTime}{The current simulation time e.g., \code{time(sim)}.}

\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-{ecoregion}-specific species
trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar
biophysical characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.}

\item{minRelativeB}{a \code{data.frame} with  the cut points to classify stand shadiness.}
}
\value{
\code{cohortData} table with a \code{siteShade} column
}
\description{
Calculate site shade
}
\seealso{
\code{\link[=makeMinRelativeB]{makeMinRelativeB()}}
}
