% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{Predict biomass change with climate variables
 @return NULL in place of a model object}
\usage{
calculateClimateEffect(...)
}
\description{
Predict biomass change with climate variables
 @return NULL in place of a model object
}
