% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{Null climate effect}
\usage{
calculateClimateEffect(cohortData, ...)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{...}{additional arguments that are passed to LandR.CS}
}
\value{
\code{data.table} with \code{pixelGroup}, \code{age} and \code{speciesCode}, as well as
\code{mortPred} and \code{growthPred} columns with \\% reduction/increase of mortality and
growth biomasses resulting from a climate effect.
These percentages are later multiplied by the by baseline biomasses of mortality and growth
(e.g. 0\\% meaning total mortality or growth reduction, and 100\\% meaning no reduction).
This default, no climate effect, function outputs 100\\% for all cohorts for both
\code{mortPred} and \code{growthPred}.
}
\description{
Default climate effects function in the case where no climate effect is simulated
}
\details{
the \code{cohortData} object is used to calculate the \\%
reduction/increase of mortality and growth biomasses per cohort.
}
