% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{calculateClimateEffect}
\usage{
calculateClimateEffect(...)
}
\value{
NULL in place of a model object
}
\description{
calculateClimateEffect
}
