% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullClimate.R
\name{calculateClimateEffect}
\alias{calculateClimateEffect}
\title{\code{calculateClimateEffect}}
\usage{
calculateClimateEffect(...)
}
\arguments{
\item{...}{additional arguments that are passed to LandR.CS}
}
\value{
NULL in place of a model object
}
\description{
\code{calculateClimateEffect}
}
