% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{CASFRItoSpRasts}
\alias{CASFRItoSpRasts}
\title{CASFRItoSpRasts}
\usage{
CASFRItoSpRasts(CASFRIRas, CASFRIattrLong, CASFRIdt, sppEquiv, sppEquivCol,
  destinationPath)
}
\arguments{
\item{CASFRIRas}{TODO: description needed}

\item{CASFRIattrLong}{TODO: description needed}

\item{CASFRIdt}{TODO: description needed}

\item{sppEquiv}{TODO: description needed}

\item{sppEquivCol}{TODO: description needed}

\item{destinationPath}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
TODO: description and title needed
}
