% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{checkSpeciesTraits}
\alias{checkSpeciesTraits}
\title{Check if all species in have trait values}
\usage{
checkSpeciesTraits(speciesLayers, species, sppColorVect)
}
\arguments{
\item{speciesLayers}{A \code{RasterStack} or \code{RasterLayer} that
should contain species cover data in the study area}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.}

\item{sppColorVect}{A named vector of colours to use for plotting.
The names must conform with species name convention used
(see \link{sppEquivalencies_CA} for possible naming
conventions) and should also contain a colour for 'Mixed',
when a mixed forest type is supported  (see \code{vegLeadingProportion}
parameter in \code{\link[=vegTypeGenerator]{vegTypeGenerator()}} for details on mixed
forest types).}
}
\value{
A \code{list} with the \code{speciesLayers} and \code{sppColorVect}
containing only the species that have trait values in \code{species}
}
\description{
Check if all species in have trait values
}
