% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\docType{data}
\name{cohortDefinitionCols}
\alias{cohortDefinitionCols}
\title{Default columns that define cohorts}
\format{
An object of class \code{character} of length 3.
}
\usage{
cohortDefinitionCols
}
\description{
Default columns that define cohorts
}
\keyword{datasets}
