% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Colors}
\alias{Colors}
\title{Helper for setting Raster or SpatRaster colors}
\usage{
Colors(ras, cols, n = NULL)
}
\arguments{
\item{ras}{A \verb{Raster*} or \code{SpatRaster} class object.}

\item{cols}{a character vector of colours. See examples. Can also be a \code{data.frame},
see \code{terra::coltab}}

\item{n}{A numeric scalar giving the number of colours to create. Passed to
\verb{quickPlot::setColors(ras, n = n) <- }. If missing, then \code{n} will be \code{length(cols)}}
}
\description{
This is a wrapper to help with migration to terra. Currently can only be used
for a single layer \code{SpatRaster} or a \code{RasterLayer}
}
\examples{
\donttest{
cols <- colorRampPalette(c("blue", "red"))(12)
ras <- terra::rast(matrix(1:100, 10, 10))
ras <- Colors(ras, cols)
terra::plot(ras)

ras <- raster::raster(matrix(1:100, 10, 10))
ras <- Colors(ras, cols)
raster::plot(ras)
}
}
