% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{columnsForPixelGroups}
\alias{columnsForPixelGroups}
\title{Default columns that define pixel groups}
\usage{
columnsForPixelGroups()
}
\description{
Default columns that define pixel groups
}
\note{
because the name \code{columnsForPixelGroups} is also used as a function argument,
be sure to use \code{LandR::columnsForPixelGroups()} in those functions or you'll get a
"promise already under evaluation" error.
}
