% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.compareRas}
\alias{.compareRas}
\alias{.compareCRS}
\title{Compare raster properties}
\usage{
.compareRas(ras1, ...)

.compareCRS(ras1, ras2)
}
\arguments{
\item{ras1}{a Raster* or SpatRaster object}

\item{...}{additional Raster* or SpatRaster objects, and arguments
passed to \code{terra::compareGeom}}

\item{ras2}{a Raster* or SpatRaster object}
}
\value{
the projected extent
}
\description{
TODO: Move to \code{reproducible}

TODO: Move to \code{reproducible}
TODO: expand to multiple objects
}
\details{
Note: this function internally converts Raster* to
\code{SpatRaster} to allow using \code{compareGeom} and benefit
from its complexity
}
