% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.compareRas}
\alias{.compareRas}
\alias{.compareCRS}
\title{Compare two raster's properties}
\usage{
.compareRas(ras1, ras2, ...)

.compareCRS(ras1, ras2, ...)
}
\arguments{
\item{ras2}{a Raster* or SpatRaster object}

\item{...}{passed to \code{terra::compareGeom}}

\item{ras}{a Raster* or SpatRaster object}
}
\value{
the projected extent
}
\description{
TODO: Move to \code{reproducible}
TODO: should be extended to many rasters
}
\details{
Note: this function internally converts Raster* to
\code{SpatRaster} to allow using \code{compareGeom} and benefit
from its complexity
}
