% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.compareRas}
\alias{.compareRas}
\alias{.compareCRS}
\title{Compare raster properties}
\usage{
.compareRas(x, ...)

.compareCRS(x, y)
}
\arguments{
\item{x, y}{a \code{Raster}, \code{SpatRaster}, \code{sf}, \code{SpatVector}, or \code{Spatial} object}

\item{...}{additional \code{Raster} or \code{SpatRaster} objects, and arguments
passed to \code{terra::compareGeom()}.}
}
\value{
the projected extent
}
\description{
TODO: Move to \code{reproducible}

TODO: Move to \code{reproducible}
TODO: expand to multiple objects
}
\note{
this function internally converts \code{Raster} to \code{SpatRaster} to allow using \code{compareGeom()},
and benefit from its complexity.
}
