% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{convert_SCANFI_LCC_codes}
\alias{convert_SCANFI_LCC_codes}
\title{Convert SCANFI Landcover layers from 1-8 codes to typical Canada LCC codes (0-230)}
\usage{
convert_SCANFI_LCC_codes(year = 2000, dataVersion = "V2", ...)
}
\arguments{
\item{year}{data year for SCANFI landcover data. 2000, 2010, and 2020 possible for V1.
1985, 1990, 1995, 2000, 2005, 2010, 2015, 2020, 2025 possible for V2.}

\item{dataVersion}{Character. SCANFI product version for data. Default is currently V2. V1 also available.}

\item{...}{additional args (not used)}
}
\value{
a \code{SpatRaster} with corrected classification codes
}
\description{
This function has been run previously to create layers with the correct codes which
are available on Google Drive. These corrected layers are available here:
\url{https://drive.google.com/drive/folders/1zLYV-wcDjJfSflH1VkXG6sosqZZF4SYc}
In the respective folder for each year, under the file name
\verb{SCANFI_att_nfiLandCover_CanadaLCCclassCodes_S_<YEAR>_v1_1.tif}
}
