% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{convert_SCANFI_LCC_codes}
\alias{convert_SCANFI_LCC_codes}
\title{Convert SCANFI Landcover layers from 1-8 codes to typical Canada LCC codes (0-230)}
\usage{
convert_SCANFI_LCC_codes(year = 2000, ...)
}
\arguments{
\item{year}{data year for SCANFI landcover data. 2000, 2010, and 2020 possible.}
}
\value{
a \code{SpatRaster} with corrected classification codes
}
\description{
Convert SCANFI Landcover layers from 1-8 codes to typical Canada LCC codes (0-230)
}
