% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultData.R
\name{defaultEnvirData}
\alias{defaultEnvirData}
\title{Default environmental variables}
\usage{
defaultEnvirData(
  vars = c("MAT", "PPT_wt", "PPT_sm", "CMI", "elevation", "timeSinceFire", "wetlands",
    "permafrost", "thermokarst"),
  studyArea,
  userTags,
  destinationPath,
  rasterToMatch
)
}
\arguments{
\item{vars}{character. Variables to download.}

\item{studyArea}{a spatial polygon  (\code{sf}, \code{sp} or \code{SpatVector}) used to crop
and mask data layers. Passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{userTags}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{destinationPath}{passed to \code{reproducible::Cache} and \code{reproducible::prepInputs}}

\item{rasterToMatch}{a raster layer (\code{RasterLayer} or \code{SpatRaster}). Only used to \code{rasterize}
the default permafrost layer and fire perimeters.}
}
\value{
named list of \code{SpatRaster}s (and a \code{SpatVector}), with names following \code{vars},
except for "permafrost", "thermokarst", which are both in the \code{"permafrostPoly"} \code{SpatVector}.
}
\description{
Downloads a set of default environmental variable rasters to model
\code{maxB}, SEP and thermokarst (among others).
}
\details{
By default this function downloads four bioclimatic variables (mean annual temperature
MAT, precipitation of warmest quarter, PPT_sm, precipitation of coldest quarter, PPT_wt, annual
climate moisture index, CMI) obtained from AdaptWest for the reference period 1980-2010
(https://s3-us-west-2.amazonaws.com/www.cacpd.org/CMIP6/normals/Normal_1981_2010_bioclim.zip),
elevation (from https://adaptwest.databasin.org/pages/adaptwest-climatena/) - at 1 Km resolution -
a binary map of wetlands (from Wulder et al. 2018), permafrost peatland complex cover and level
of thermokarst (from Gibson et al. 2021), and a map of time since fire (considering fire perimeters from the Canadian Wildland
Fire Information System. All layers are \code{SpatRaster}s except for \code{"permafrostPoly"}, which is a
polygon layer containing permafrost peatland complex cover ("Permafrost" column)
and level of thermokarst ("Degree_Of_" column), among other information.
}
\references{
Gibson, C., Morse, P. D., Kelly, J. M., Turetsky, M. R., Baltzer, J. L.,
Gingras-Hill, T., & Kokelj, S. V. (2020). Thermokarst Mapping Collective: Protocol
for organic permafrost terrain and preliminary inventory from the Taiga Plains
test area, Northwest Territories (NWT Open Report 2020-010, p. 29).
Northwest Territories Geological Survey.

Wulder, M., Li, Z., Campbell, E., White, J., Hobart, G., Hermosilla, T., &
Coops, N. (2018). A National Assessment of Wetland Status and Trends for Canada's
Forested Ecosystems Using 33 Years of Earth Observation Satellite Data.
Remote Sensing, 10(10), 1623. https://doi.org/10.3390/rs10101623
}
