% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{describeCohortData}
\alias{describeCohortData}
\title{Summary for cohortData}
\usage{
describeCohortData(cohortData)
}
\arguments{
\item{cohortData}{A cohortData object}
}
\description{
Summary for cohortData
}
