% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbances.R
\name{FireDisturbance}
\alias{FireDisturbance}
\alias{FireDisturbancePM}
\alias{PeatlandThermokarst}
\title{Disturbance functions}
\usage{
FireDisturbance(
  cohortData = copy(sim$cohortData),
  cohortDefinitionCols = c("pixelGroup", "age", "speciesCode"),
  calibrate = FALSE,
  postFireRegenSummary = copy(sim$postFireRegenSummary),
  treedFirePixelTableSinceLastDisp = sim$treedFirePixelTableSinceLastDisp,
  rstCurrentBurn = sim$rstCurrentBurn,
  inactivePixelIndex = sim$inactivePixelIndex,
  pixelGroupMap = sim$pixelGroupMap,
  currentTime = NULL,
  rasterToMatch = sim$rasterToMatch,
  species = copy(sim$species),
  sufficientLight = copy(sim$sufficientLight),
  speciesEcoregion = copy(sim$speciesEcoregion),
  initialB = 10,
  successionTimestep = 10L,
  doAssertion = getOption("LandR.assertions", TRUE),
  verbose = getOption("LandR.verbose", TRUE)
)

FireDisturbancePM(
  cohortData = copy(sim$cohortData),
  cohortDefinitionCols = c("pixelGroup", "age", "speciesCode"),
  colsForPixelGroups = LandR::columnsForPixelGroups(),
  calibrate = FALSE,
  LANDISPM = TRUE,
  postFireRegenSummary = copy(sim$postFireRegenSummary),
  treedFirePixelTableSinceLastDisp = copy(sim$treedFirePixelTableSinceLastDisp),
  rstCurrentBurn = sim$rstCurrentBurn,
  inactivePixelIndex = sim$inactivePixelIndex,
  pixelGroupMap = sim$pixelGroupMap,
  currentTime = NULL,
  rasterToMatch = sim$rasterToMatch,
  fireDamageTable = copy(sim$fireDamageTable),
  fireRSORas = sim$fireRSORas,
  fireROSRas = sim$fireROSRas,
  fireCFBRas = sim$fireCFBRas,
  species = copy(sim$species),
  sufficientLight = copy(sim$sufficientLight),
  speciesEcoregion = copy(sim$speciesEcoregion),
  initialB = 10,
  minRelativeB = copy(sim$minRelativeB),
  successionTimestep = 10L,
  doAssertion = getOption("LandR.assertions", TRUE),
  verbose = getOption("LandR.verbose", TRUE)
)

PeatlandThermokarst(
  thawedPixIDs = copy(sim$thawedPixIDs),
  treedThawedPixelTableSinceLastDisp = copy(sim$treedThawedPixelTableSinceLastDisp),
  wetlands = sim$wetlands,
  cohortData = copy(sim$cohortData),
  pixelGroupMap = sim$pixelGroupMap,
  rasterToMatch = sim$rasterToMatch,
  species = copy(sim$species),
  speciesEcoregion = copy(sim$speciesEcoregion),
  cohortDefinitionCols = LandR::cohortDefinitionCols(),
  initialB = 10L,
  inactivePixelIndex = sim$inactivePixelIndex,
  currentTime = NULL,
  successionTimestep = 10L
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{cohortDefinitionCols}{the columns in \code{cohortData} that define unique cohorts
(default: \code{c("pixelGroup", "speciesCode", "age")}).}

\item{calibrate}{logical. Determines whether to output \code{postFirePixelCohortData}.
Defaults to FALSE}

\item{postFireRegenSummary}{a data.table summarizing for which species serotiny/resprouting were
activated and in how many pixels, for each year. Only necessary if \code{calibrate = TRUE}.}

\item{treedFirePixelTableSinceLastDisp}{data.table with 3 columns: \code{pixelIndex},
\code{pixelGroup}, and \code{burnTime}. Each row represents a forested pixel that was
burned up to and including this year, since last dispersal event, with its
corresponding \code{pixelGroup} and time it occurred Pixel group IDs correspond to
the last year's \code{pixelGroupMap} and not necessarily the pixelGroupMap of
the \code{burnTime} year.}

\item{rstCurrentBurn}{raster layer with burnt pixels in \code{currentTime}.
Usually generated by a fire spread module (or collection of modules) like
\code{fireSense}. Values in the raster should contain 0's for unburnt pixels and
values > 0 for burn pixels. NA's are assumed to be outside of the study area.}

\item{inactivePixelIndex}{an internal object to \code{Biomass_core} that track of
inactive pixels.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{currentTime}{numeric. Current simulation time step (as in \code{time(sim)})}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.
Must have column \code{speciesCode}, with species names/IDs. The following is a
list of default trait columns:
\itemize{
\item "species" same as "speciesCode" -- species ID name
\item "speciesCode"
\item "Area" -- inherited from LANDIS-II default table, the Canadian ecoregion
from which traits where derived. Not used during the simulation
\item "firetolerance" -- module\emph{relative} (to other species) fire tolerance
\item "growthcurve" and "mortalityshape" -- growth curve shape parameters.
\item "longevity" -- maximum species age
\item "postfireregen" -- post-fire regeneration strategy ("serotiny", "resprout" or "none")
\item "resproutprob" -- probability of resprouting
\item "resproutage_min" -- minimum age at which species is capable of resprouting
\item "resproutage_max" -- maximum age at which species is capable of resprouting
\item "seeddistance_eff" -- effective dispersal distance
\item "seeddistance_max" -- maximum dispersal distance
\item "shadetolerance" -- \emph{relative} (to other species) shade tolerance
\item "sexualmature" -- age at sexual maturity
Known optional parameters added/needed by some modules (the user may add others for their own modules):
\item "inflationFactor" -- \code{Biomass_speciesParameters} module: inflation factor for \code{maxB}
\item "growthCurveSource" -- \code{Biomass_speciesParameters} module: how "growthcurve" was estimated
\item "mANPPproportion" -- \code{Biomass_speciesParameters} module: multiplication factor to calculate \code{maxANPP} from \code{maxB}
\item "thermokarsttol" -- \code{Biomass_disturbances} module: proportion of biomass surviving after thermokarst (i.e. permafrost thaw). Applied equally across cohorts.
Parameters inherited from LANDIS-II default table, but not used in LandR at the moment:
\item "leaflongevity"
\item "wooddecayrate"
\item "leafLignin"
\item "hardsoft"
Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015)
for further detail.
}}

\item{sufficientLight}{a \code{data.frame} (not \code{data.table}!) containing probability
of establishment, given a site's light conditions (\code{X0}-\code{X5}) for each level
of a species shade tolerance (\code{1}-\code{5}).}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-\code{ecoregion}-specific species trait values.
Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical
characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.
May contain columns \code{inflationFactor}  (used to adjust \code{maxB}) and \code{mANPPproportion}
(used to calculate \code{maxANPP}).}

\item{initialB}{the initial biomass of new cohorts. Defaults to 10.
If \code{NULL} or \code{NA}, initial cohort biomass is calculated as in LANDIS-II Biomass Succession
Extension v3.2.1 (Scheller & Miranda, 2015):

\if{html}{\out{<div class="sourceCode r">}}\preformatted{initialB = asInteger(pmin(maxANPP, asInteger(pmax(1, maxANPP * exp(-1.6 * sumB / maxB_eco)))))
}\if{html}{\out{</div>}}

where \code{maxANPP} and \code{maxB_eco} are the maximum ANPP and B parameters of the species
in question within the pixel's \code{ecolocation}, and \code{sumB} is the total stand biomass
excluding cohorts with ages less than \code{successionTimestep}.}

\item{successionTimestep}{The time between successive seed dispersal events.
In LANDIS-II, this is called "Succession Timestep".}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{verbose}{Controls message output. Defaults to \code{getOption("LandR.verbose")}}

\item{colsForPixelGroups}{character. \code{cohortData} columns used to find identical pixels in terms
of cohort composition, and group them into \code{pixelGroups}
(default: \code{c("ecoregionGroup", "speciesCode", "age", "B")}).}

\item{LANDISPM}{logical. Should partial mortality be calculated as in LANDIS-II Dynamic Fire System v3.0.
Must be \code{TRUE} for the time being.}

\item{fireDamageTable}{\code{data.table} defining upper age limit of cohorts killed by fire.
See LANDIS-II Dynamic Fire System v3.0 Manual for an example.}

\item{fireRSORas}{raster layer of critical spread rate for crowning (m/min) for each burned pixel.
If not provided dummy values are used.}

\item{fireROSRas}{raster layer of equilibrium rate of spread (m/min) for each burned pixel.
If not provided dummy values are used.}

\item{fireCFBRas}{raster layer of crown fraction burnt for each burned pixel.
If not provided dummy values are used.}

\item{minRelativeB}{a \code{data.frame} with  the cut points to classify stand shadiness.}

\item{thawedPixIDs}{integer. Vector of pixel IDs than underwent thermokarst
(and were converted to wetlands) in the current year (not the last thermokarst event).}

\item{treedThawedPixelTableSinceLastDisp}{data.table with 3 columns: \code{pixelIndex},
\code{pixelGroup}, and \code{burnTime}. Each row represents a forested pixel that was
burned up to and including this year, since last dispersal event, with its
corresponding \code{pixelGroup} and time it occurred Pixel group IDs correspond to
the last year's \code{pixelGroupMap} and not necessarily the pixelGroupMap of
the \code{burnTime} year.}

\item{wetlands}{binary SpatRaster with current wetland pixels.}
}
\value{
a list of objects to be exported to sim:
\itemize{
\item cohortData
\item pixelGroupMap
\item lastFireYear
\item treedFirePixelTableSinceLastDisp
\item serotinyResproutSuccessPixels
\item severityBMap
\item severityData
\item postFireRegenSummary (if \code{calibrate == TRUE} and \code{!is.null(postFireRegenSummary)})
}
}
\description{
Disturbance functions
}
\section{Stand-replacing fire disturbances}{

\code{FireDisturbance()} simulates post-fire mortality, serotiny and regeneration sequentially occurring after a fire.
Post-fire mortality is assumed to be 100\% (stand-replacement). The serotiny and regeneration
algorithms are based on those in LANDIS-II Biomass Succession extension, v3.2.1, with modifications
Requires the following objects in \code{sim} (passed as \verb{sim$*}):
@template FireDisturbance.
For any given burnt pixel, the function begins by killing all cohorts
(i.e. removing them from \code{cohortData}). Then it activates serotiny for
serotinous species that had been present pre-fire, and reprouting for reprouter
species. Whether a species successfully regenerates via serotiny or resprouting
depends on its traits (whether it is serotinous, resprouter, both or none;
\code{species} table), pre-fire cohort age (in \code{cohortData} table), its light/shade
requirements (\code{species} table) and its local \code{maxB} (\code{speciesEcoregion})
table (see Scheller & Miranda, 2015 and Scheller & Mladenoff, 2004 for details).
Unlike in LANDIS-II Biomass Succession extension, v3.2.1, here we allow serotiny
and resprouting to occur in the same pixel to reflect the competitive advantage
of reprouters. However, for a given species only serotiny (takes precedence)
or resprouting can be activated.
The \code{species} table must contain the columns:
\itemize{
\item \code{sexualmature} -- age at sexual maturity
\item \code{postfireregen} -- post-fire regeneration strategy ("serotiny", "resprout" or "none")
\item \code{shadetolerance} -- shade tolerance value relative to \emph{other} species.
\item \code{resproutage_min}, \code{resproutage_max} -- minimum and maximum age at which species can repsrout
\item \code{resproutprob} -- probability of resporuting success (before light/shade suitability is assessed)
}
}

\section{Partial severity (i.e. mortality) fire disturbances}{

\code{FireDisturbancePM()} simulates partial post-fire mortality, serotiny and regeneration
sequentially after a fire. The level of mortality depends of fire severity,
and, by default, follows the mechanisms in LANDIS-II Dynamic Fire System v3.0.
Serotiny and regeneration algorithms algorithms are based on those in
LANDIS-II Biomass Succession extension, v3.2.1, with modifications (see \code{FireDisturbance()}).
Requires the following objects in \code{sim} (and passed as \verb{sim$*}):
@template FireDisturbance
\itemize{
\item \code{fireDamageTable}
\item \code{fireRSORas} (Critical spread rate for crowning)
\item \code{fireROSRas} (Equilibrium Head Fire Rate of Spread)
\item \code{fireCFBRas} (Crown Fraction Burnt)
\item \code{minRelativeB}
Rasters of fire behaviour properties (\code{fireRSORas}, \code{fireROSRas} and \code{fireCFBRas})
can be calculated using the \code{cffdrs} package.
The \code{species} table must contain the columns:
\item \code{firetolerance} -- fire tolerance value relative to \emph{other} species.
\item \code{longevity} -- maximum species age
\item \code{sexualmature} -- age at sexual maturity
\item \code{postfireregen} -- post-fire regeneration strategy ("serotiny", "resprout" or "none")
\item \code{shadetolerance} -- shade tolerance value relative to \emph{other} species.
\item \code{resproutage_min}, \code{resproutage_max} -- minimum and maximum age at which species can repsrout
\item \code{resproutprob} -- probability of resporuting success (before light/shade suitability is assessed)
}
}

\section{Peatland permafrost degradation (thermokarst) disturbances}{

\code{PeatlandThermokarst()} simulates tree cohort survival/mortality after
peatland permafrost thermokarst. The level of mortality depends on species
tolerance to thermokarst, determined by the \code{thermokarsttol} trait column
in the \code{species} traits table. At the moment, this level of tolerance is used
as the proportion of a cohort biomass that survives (is kept) when a pixel undergoes
thermokarst. This is similar to the partial disturbance effects used in LANDIS-II Biomass
Harvest v4.0. Requires the following objects in \code{sim} (and passed as \verb{sim$*}):
\itemize{
\item \code{thawedPixIDs}
\item \code{treedThawedPixelTableSinceLastDisp}
\item \code{wetlands}
\item \code{cohortData}
\item \code{pixelGroupMap}
\item \code{rasterToMatch}
\item \code{species}
\item \code{speciesEcoregion}
\item \code{inactivePixelIndex}
The \code{species} table must contain the columns:
\item \code{thermokarsttol} -- proportion of cohort biomass (B) that survives thermokarst
}
}

\references{
Scheller, R.M. & Miranda, B.R. (2015). LANDIS-II Biomass Succession v3.2 Extension  – User Guide.

Scheller, R.M. & Mladenoff, D.J. (2004). A forest growth and biomass module for a landscape simulation model, LANDIS: design, validation, and application. Ecological Modelling, 180, 211–229.

Sturtevant, B.R., Miranda, B.R., Scheller, R.M. & Shinneman, D. (2018). LANDIS-II Dynamic Fire System Extension v3.0  – User Guide.

Scheller, R.M. & Domingo, J.B. (2021). LANDIS-II Biomass Harvest v4.4 Extension – User Guide
}
