% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbances.R
\name{FireDisturbance}
\alias{FireDisturbance}
\alias{FireDisturbancePM}
\title{Disturbance functions}
\usage{
FireDisturbance(
  cohortData = sim$cohortData,
  cohortDefinitionCols = SpaDES.core::P(sim)$cohortDefinitionCols,
  calibrate = SpaDES.core::P(sim)$calibrate,
  postFireRegenSummary = sim$postFireRegenSummary,
  rstCurrentBurn = sim$rstCurrentBurn,
  inactivePixelIndex = sim$inactivePixelIndex,
  pixelGroupMap = sim$pixelGroupMap,
  currentTime = SpaDES.core::time(sim),
  rasterToMatch = sim$rasterToMatch,
  species = sim$species,
  sufficientLight = sim$sufficientLight,
  speciesEcoregion = sim$speciesEcoregion,
  initialB = SpaDES.core::P(sim)$initialB,
  successionTimestep = SpaDES.core::P(sim)$successionTimestep,
  verbose = getOption("LandR.verbose", TRUE)
)

FireDisturbancePM(
  cohortData = sim$cohortData,
  cohortDefinitionCols = SpaDES.core::P(sim)$cohortDefinitionCols,
  calibrate = SpaDES.core::P(sim)$calibrate,
  LANDISPM = TRUE,
  postFireRegenSummary = sim$postFireRegenSummary,
  rstCurrentBurn = sim$rstCurrentBurn,
  inactivePixelIndex = sim$inactivePixelIndex,
  pixelGroupMap = sim$pixelGroupMap,
  currentTime = SpaDES.core::time(sim),
  rasterToMatch = sim$rasterToMatch,
  fireDamageTable = sim$fireDamageTable,
  fireRSORas = sim$fireRSORas,
  fireROSRas = sim$fireROSRas,
  fireCFBRas = sim$fireCFBRas,
  species = sim$species,
  sufficientLight = sim$sufficientLight,
  speciesEcoregion = sim$speciesEcoregion,
  initialB = SpaDES.core::P(sim)$initialB,
  minRelativeB = sim$minRelativeB,
  successionTimestep = SpaDES.core::P(sim)$successionTimestep,
  verbose = getOption("LandR.verbose", TRUE)
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{cohortDefinitionCols}{the columns in \code{cohortData} that define unique cohorts}

\item{calibrate}{logical. Determines whether to output \code{postFirePixelCohortData}.
Defaults to FALSE}

\item{postFireRegenSummary}{a data.table summarizing for which species serotiny/resprouting were
activated and in how many pixels, for each year. Only necessary if \code{calibrate = TRUE}.}

\item{rstCurrentBurn}{raster layer with burnt pixels in \code{currentTime}.
Usually generated by a fire spread module (or collection of modules) like
\code{fireSense}. Values in the raster should contain 0's for unburnt pixels and
values > 0 for burn pixels. NA's are assumed to be outside of the study area.}

\item{inactivePixelIndex}{an internal object to \code{Biomass_core} that track of
inactive pixels.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{currentTime}{numeric. Current simulation time step (as in \code{time(sim)})}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.
Must have column \code{speciesCode}, with species names/IDs}

\item{sufficientLight}{a \code{data.frame} (not \code{data.table}!) containing probability
of establishment, given a site's light conditions (\code{X0}-\code{X5}) for each level
of a species shade tolerance (\code{1}-\code{5}).}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-\code{ecoregion}-specific species trait values.
Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical
characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.
May contain columns \code{inflationFactor}  (used to adjust \code{maxB}) and \code{mANPPproportion}
(used to calculate \code{maxANPP}).}

\item{initialB}{the initial biomass of new cohorts. Defaults to 10.
If \code{NULL} or \code{NA}, initial cohort biomass is calculated as in LANDIS-II Biomass Succession
Extension v3.2.1 (Scheller & Miranda, 2015):

\if{html}{\out{<div class="sourceCode r">}}\preformatted{initialB = asInteger(pmin(maxANPP, asInteger(pmax(1, maxANPP \\* exp(-1.6 \\* sumB / maxB_eco)))))
}\if{html}{\out{</div>}}

where \code{maxANPP} and \code{maxB_eco} are the maximum ANPP and B parameters of the species
in question within the pixel's \code{ecolocation}, and \code{sumB} is the total stand biomass
excluding cohorts with ages less than \code{successionTimestep}.}

\item{successionTimestep}{The time between successive seed dispersal events.
In LANDIS-II, this is called "Succession Timestep".}

\item{verbose}{Controls message output. Defaults to \code{getOption("LandR.verbose")}}

\item{LANDISPM}{logical. Should partial mortality be calculated as in LANDIS-II Dynamic Fire System v3.0.
Must be \code{TRUE} for the time being.}

\item{fireDamageTable}{\code{data.table} defining upper age limit of cohorts killed by fire.
See LANDIS-II Dynamic Fire System v3.0 Manual for an example.}

\item{fireRSORas}{raster layer of critical spread rate for crowning (m/min) for each burned pixel.}

\item{fireROSRas}{raster layer of equilibrium rate of spread (m/min) for each burned pixel.}

\item{fireCFBRas}{raster layer of crown fraction burnt for each burned pixel.}

\item{minRelativeB}{a \code{data.frame} with  the cut points to classify stand shadiness.}
}
\value{
a list of objects to be exported to sim:
\itemize{
\item cohortData
\item pixelGroupMap
\item lastFireYear
\item treedFirePixelTableSinceLastDisp
\item serotinyResproutSuccessPixels
\item severityBMap
\item severityData
\item postFireRegenSummary (if \code{calibrate == TRUE} and \code{!is.null(postFireRegenSummary)})
}
}
\description{
Disturbance functions
}
\details{
For any given burnt pixel, the function begins by killing all cohorts
(i.e. removing them from \code{cohortData}). Then it activates serotiny for
serotinous species that had been present pre-fire, and reprouting for reprouter
species. Whether a species successfully regenerates via serotiny or resprouting
depends on its traits (whether it is serotinous, resprouter, both or none;
\code{species} table), pre-fire cohort age (in \code{cohortData} table), its light/shade
requirements (\code{species} table) and its local \code{maxB} (\code{speciesEcoregion})
table (see Scheller & Miranda, 2015 and Scheller & Mladenoff, 2004 for details).
Unlike in LANDIS-II Biomass Succession extension, v3.2.1, here we allow serotiny
and resprouting to occur in the same pixel to reflect the competitive advantage
of reprouters. However, for a given species only serotiny (takes precedence)
or resprouting can be activated.
}
\section{Stand-replacing fire disturbances}{

\code{FireDisturbance()} simulates post-fire mortality, serotiny and regeneration sequentially occurring after a fire.
Post-fire mortality is assumed to be 100\% (stand-replacement). The serotiny and regeneration
algorithms are based on those in LANDIS-II Biomass Succession extension, v3.2.1, with modifications
see \code{details} below. Requires the following objects in \code{sim}:
@template FireDisturbance
}

\section{Partial severity (i.e. mortality) fire disturbances}{

\code{FireDisturbancePM()} simulates partial post-fire mortality, serotiny and regeneration
sequentially after a fire. The level of mortality depends of fire severity,
and, by default, follows the mechanisms in LANDIS-II Dynamic Fire System v3.0. Serotiny and regeneration
algorithms algorithms are based on those in LANDIS-II Biomass Succession extension, v3.2.1,
with modifications (see \code{FireDisturbance()}). Requires the following objects in \code{sim}:
@template FireDisturbance
\itemize{
\item \code{fireDamageTable}
\item \code{fireRSORas} (optional)
\item \code{fireROSRas}
\item \code{fireCFBRas}
\item \code{minRelativeB}
Rasters of fire behaviour properties (\code{fireRSORas}, \code{fireROSRas} and \code{fireCFBRas})
can be calculated using the \code{cffdrs} package.
}
}

\references{
Scheller, R.M. & Miranda, B.R. (2015). LANDIS-II Biomass Succession v3.2 Extension  – User Guide.

Scheller, R.M. & Mladenoff, D.J. (2004). A forest growth and biomass module for a landscape simulation model, LANDIS: design, validation, and application. Ecological Modelling, 180, 211–229.

Sturtevant, B.R., Miranda, B.R., Scheller, R.M. & Shinneman, D. (2018). LANDIS-II Dynamic Fire System Extension v3.0  – User Guide.
}
