% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serotiny-resprouting.R
\name{doSerotiny}
\alias{doSerotiny}
\title{Activate serotiny after a (fire) disturbance}
\usage{
doSerotiny(
  burnedPixelCohortData,
  postFirePixelCohortData,
  postFireRegenSummary = NULL,
  species,
  sufficientLight,
  speciesEcoregion,
  simuTime,
  treedFirePixelTableSinceLastDisp,
  calibrate = FALSE
)
}
\arguments{
\item{burnedPixelCohortData}{An expanded \code{cohortData} \code{data.table} with pixel-level
cohort information on burnt pixels and columns:
\code{severity} - fire severity in that pixel calculated based on fire behaviour properties;
\code{firetolerance} - species-level fire tolerance;
\code{severityToleranceDif} - the difference between \code{severity} and \code{firetolerance}.}

\item{postFirePixelCohortData}{an empty \code{cohortData}-like table with columns:
\code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB} removed;
and \code{pixelIndex} added.}

\item{postFireRegenSummary}{a \code{data.table} summarizing for which species
serotiny/resprouting were activated and in how many pixels, for each year.
Only necessary if \code{calibrate = TRUE}.}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.}

\item{sufficientLight}{a \code{data.table} containing probability of establishment,
given a site's light conditions (\code{X0}-\code{X5}) for each level of a species shade
tolerance (\code{1}-\code{5}).}

\item{speciesEcoregion}{a \code{data.table} with \code{speciesEcoregion} values}

\item{simuTime}{integer. The current simulation time obtained with \code{time(sim)}}

\item{treedFirePixelTableSinceLastDisp}{a vector of pixels that burnt and were forested
in the previous time step.}

\item{calibrate}{logical. Determines whether to output \code{postFirePixelCohortData}.
Defaults to \code{FALSE}.}
}
\value{
A list of objects:
    \code{postFirePixelCohortData}, a \code{data.table} with the cohorts that undergo serotiny;
    \code{serotinyPixel}, a vector of pixels where serotiny was activated;
    \code{postFireRegenSummary}, the updated \code{postFireRegenSummary}, if \code{calibrate = TRUE}.
}
\description{
Activate serotiny after a (fire) disturbance
}
