% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permafrost.R
\name{.assignPermafrostWrapper}
\alias{.assignPermafrostWrapper}
\title{Wrapper for \code{assignPermafrost}}
\usage{
.assignPermafrostWrapper(id, cores, dPath, ...)
}
\arguments{
\item{id}{passed to \code{assignPermafrost}}

\item{cores}{how many threads to use for parallelisation. If
\code{1}, no parallelisation occurs}

\item{dPath}{directory where output permafrost raster file will be written.}

\item{...}{further arguments passed to \code{assignPermafrost}, like \code{gridPoly}
\code{ras}, etc.}
}
\description{
Wrapper for \code{assignPermafrost}
}
