% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{.createCohortData}
\alias{.createCohortData}
\title{Generate template \code{cohortData} table}
\usage{
.createCohortData(
  inputDataTable,
  pixelGroupBiomassClass,
  doAssertion = getOption("LandR.assertions", TRUE),
  rescale = TRUE
)
}
\arguments{
\item{inputDataTable}{A \code{data.table} with columns described above.}

\item{pixelGroupBiomassClass}{Round B to the nearest \code{pixelGroupBiomassClass}
to establish unique \code{pixelGroups}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{rescale}{Logical. If \code{TRUE}, the default, cover for each species will be rescaled
so all cover in \code{pixelGroup} or pixel sums to 100.}
}
\description{
Internal function used by \code{\link{makeAndCleanInitialCohortData}}.
}
\keyword{internal}
