% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLMs.R
\name{.fitNLMwCovariates}
\alias{.fitNLMwCovariates}
\title{Fit non-linear growth model under various starting conditions}
\usage{
.fitNLMwCovariates(
  data,
  nonLinModelQuoted,
  linModelQuoted,
  mllsOuterPrev,
  model = c("CR", "Logistic"),
  maxCover = 1L,
  cores = 1L,
  starts = NULL,
  lower = NULL,
  upper = NULL,
  nbWorkers = 1L
)
}
\arguments{
\item{data}{a \code{data.table} or \code{data.frame} with all covariates
and the response variable. Note that incomplete lines are removed.}

\item{nonLinModelQuoted}{The non-linear equation as a \code{call}
(quoted expression) passed to \code{mle2(minuslog1)}. See \code{?mle}.
Accepts equations with three parameters 'A', 'p' and 'k'.}

\item{linModelQuoted}{A list of linear equations/modes relating each
parameter ('A', 'p' and 'k') with a set of covariates. A \code{call}
(quoted expression) passed to \code{mle2(..., parameters)}. Note that for the
purpose of tree growth, the linear equation determining 'A' should include a
'cover' predictor indicating the tree cover or dominance in the stand. Should be
scaled between 0 and \code{maxCover}.}

\item{mllsOuterPrev}{the output of a previous \code{fitNLMwCovariates} run which
is used to extract last best AIC and maximum biomass estimate and judge if new
iterations are better.}

\item{model}{character. Non-linear model form used to estimate average maximum
biomass. One of "CR" (Chapman-Richards) or "Logistic". In both cases, maximum biomass
is equivalent to the 'A' asymptote parameter, which is estimated using observed mean
values of predictors entering its linear equation and \code{cover == maxCover}, if this
predictor is included (as it should). Passed to \code{extractMaxB}}

\item{maxCover}{numeric. Value indicating maximum cover/dominance.}

\item{starts}{\code{data.table} or \code{data.frame} of parameter starting values. Will be coerced to named list
with names being parameter names.}

\item{lower}{passed to \code{mle2}}

\item{upper}{passed to \code{mle2}}
}
\value{
a \code{list} with entries 'mll' (the maximum likelihood-estimated
coefficients) and 'AICbest' (the AIC of the best models generating these
coefficients)
}
\description{
Uses likelihood parameter estimation to fit non linear models
while attempting several starting values.
}
