% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\docType{data}
\name{.speciesTableColNames}
\alias{.speciesTableColNames}
\title{Species Table Column Names}
\format{
An object of class \code{character} of length 18.
}
\usage{
.speciesTableColNames
}
\description{
Species Table Column Names
}
\keyword{internal}
