% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{dropTerm}
\alias{dropTerm}
\title{Drop factor term including interactions from a model formula}
\usage{
dropTerm(form, term)
}
\arguments{
\item{form}{A model formula.}

\item{term}{Character vector giving the name of the term to drop.}
}
\value{
An updated model formula.
}
\description{
Based on \url{https://stackoverflow.com/a/23382097/1380598}.
}
