% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummyInputs.R
\name{makeDummyEcoregionMap}
\alias{makeDummyEcoregionMap}
\alias{makeDummyRawBiomassMap}
\alias{makeDummyStandAgeMap}
\alias{makeDummyRstLCC}
\alias{makeDummyEcoregionFiles}
\title{Create dummy inputs for test simulations}
\usage{
makeDummyEcoregionMap(rasterToMatch)

makeDummyRawBiomassMap(rasterToMatch)

makeDummyStandAgeMap(rawBiomassMap)

makeDummyRstLCC(rasterToMatch)

makeDummyEcoregionFiles(ecoregionMap, rstLCC, rasterToMatch)
}
\arguments{
\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{rawBiomassMap}{a \code{rawBiomassMap} (e.g. the one used
throughout the simulation)}

\item{ecoregionMap}{The \code{ecoregionMap}, a raster of all the unique ecoregion groupings,
which group together pixels of similar biophysical characteristics.}

\item{rstLCC}{a raster land-cover class per pixel,}
}
\value{
a \code{RasterLayer} object or, in the case of \code{makeDummyEcoregionFiles}, a list.
}
\description{
\code{ecoregionMap}is a raster of all the unique groupings.
}
\details{
\code{rawBiomassMap} is a raster of "raw" total stand biomass per pixel,
with values between 100 and 20000 g/m^2.

\code{standAgeMap} is a raster of stand age per pixel (where biomass exists)
with values between 1 and 300 years.

\code{rstLCC} is a raster land-cover class per pixel, with values between 1 and 5 that have no
correspondence to any real land-cover classes.

\code{ecoregionFiles} uses dummy versions of \code{ecoregionMap} and \code{rstLCC}
to create a list with two objects: the \code{ecoregionMap} and a table summarizing its
information per \code{pixelID}.
See \code{ecoregionProducer} (it uses \code{ecoregionProducer} internally).
}
