% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{ecoregionProducer}
\alias{ecoregionProducer}
\title{Make \code{ecoregionMap} and \code{ecoregion} table}
\usage{
ecoregionProducer(
  ecoregionMaps,
  ecoregionName = NULL,
  rasterToMatch,
  ecoregionTable
)
}
\arguments{
\item{ecoregionMaps}{a \code{list} with two rasters, one with ecoregions (e.g. ecodistricts)
and another with land cover (e.g. LCC).}

\item{ecoregionName}{the name describing the type of ecoregions in first map
(e.g. \code{"ecoDistrict"}) if passing a polygon file.}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{ecoregionTable}{A data.table that has 2 columns, \code{ecoregionName} (a factor)
and \code{ID} a factor of \code{paddedFloatToChar(1:length(unique(ecoregionName)),  padL = max(nchar(length(unique(ecoregionName)))))}. This
represents all the possible values that are available; this will be joined
to \code{ecoregionMaps[[1]]} values}
}
\value{
A list with two objects: the \code{ecoregionMap} and a table summarizing
its information per \code{pixelID}
}
\description{
This function combines an ecoregion map and a land cover map (e.g. ecodistricts and LCC)
and creates a map and table of containing their combined values and pixel IDs.
Used internally in LandR modules to prepare maps for to make \code{cohortData}.
}
