% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{ecoregionProducer}
\alias{ecoregionProducer}
\title{Make \code{ecoregionMap} and \code{ecoregion} table}
\usage{
ecoregionProducer(
  ecoregionMaps,
  ecoregionName,
  ecoregionActiveStatus,
  rasterToMatch
)
}
\arguments{
\item{ecoregionMaps}{a \code{list} with two rasters, one with ecoregions (e.g. ecodistricts)
and another with land cover (e.g. LCC)}

\item{ecoregionName}{the name describing the type of ecoregions in first map
(e.g. \code{"ecoDistrict"})}

\item{ecoregionActiveStatus}{A two column \code{data.table} detailing with ecoregions
are to be considered active for the simulations. Columns should be named 'active'
(with 'yes' or 'no' values) and 'ecoregion'.}

\item{rasterToMatch}{A \code{RasterLayer} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}
}
\value{
A list with two objects: the \code{ecoregionMap} and a table summarizing
its information per \code{pixelID}
}
\description{
This function combines an ecoregion map and a land cover map (e.g. ecodistricts and LCC)
and creates a map and table of containing their combined values and pixel IDs.
Used internally in LandR modules to prepare maps for to make \code{cohortData}.
}
