% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregionProducer.R
\name{ecoregionProducer}
\alias{ecoregionProducer}
\title{Make \code{ecoregionMap} and \code{ecoregion} table}
\usage{
ecoregionProducer(ecoregionMaps, ecoregionName, ecoregionActiveStatus,
  rasterToMatch)
}
\arguments{
\item{ecoregionMaps}{a \code{list} with two rasters, one with eco-regions (e.g. eco-districts) and
another with land cover (e.g. LCC)}

\item{ecoregionName}{the name describing the type of eco-regions in first map (e.g. "ecoDistrict")}

\item{ecoregionActiveStatus}{A two column \code{data.table} detailing with eco-regions are to be considered
active for the simulations. Columns should be named 'active' (with 'yes' or 'no' values) and 'ecoregion'.}

\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used throughout the simulation)}
}
\value{
A list with two objects: the \code{ecoregionMap} and a table summarizing
it's information per pixelID
}
\description{
This function combines an eco-region map and a land cover map (e.g. eco-districts and LCC)
and creates a map and table of containing their combined values and pixel IDs.
Used internally in LandR modules to prepare maps for to make \code{cohortData}.
}
