% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseName.R
\name{equivalentName}
\alias{equivalentName}
\title{Return equivalent name from a \code{data.frame} of equivalencies}
\usage{
equivalentName(value, df, column, multi = FALSE, searchColumn = NULL)
}
\arguments{
\item{value}{Vector of values to match in \code{df}.}

\item{df}{A \code{data.frame} where every row is a set of equivalent names.}

\item{column}{A character string or numeric of length 1, indicating the column
in \code{df} to return names from.}

\item{multi}{Logical. If \code{TRUE}, then all matches will be returned. Default
is \code{FALSE} for backwards compatibility. This may result in more
elements than were in \code{value}, as each \code{value} may be matched
by more than one entry, returning more than one result each from \code{column}.
If a species is found in the \code{df}, but there is no corresponding
value in \code{column}, it will return \code{NA}}

\item{searchColumn}{Optionally, provide the name of a column in \code{df} that results
must be found in. The return value will still be from \code{column}}
}
\description{
This is simply a wrapper around \code{match} or \code{\%in\%} for
a specific \code{data.frame} of values.
}
