% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{fasterizeFromSp}
\alias{fasterizeFromSp}
\title{Rasterize polygons using \code{fasterize}}
\usage{
fasterizeFromSp(sp, raster, fieldName)
}
\arguments{
\item{sp}{a shapefile to rasterize}

\item{raster}{the template raster to use}

\item{fieldName}{the field to use (will be ignored if the shapefile has no fields)}
}
\value{
TODO: is it a \code{RasterLayer}?
}
\description{
Rasterize polygons using \code{fasterize}
}
