% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{generatePixelGroups}
\alias{generatePixelGroups}
\title{Add the correct \code{pixelGroups} to a \code{pixelCohortData} object}
\usage{
generatePixelGroups(pixelCohortData, maxPixelGroup,
  columns = c("ecoregionGroup", "speciesCode", "age", "biomass"))
}
\arguments{
\item{pixelCohortData}{# pixel groups are groups of identical pixels based
on \code{speciesGroup} x \code{Age} and \code{ecoregionGroup}.}

\item{maxPixelGroup}{A length 1 numeric/integer indicating the current maximum pixelGroup value}

\item{columns}{A character vector of column names to use as part of the generation of unique
combinations of features. Default is \code{c("ecoregionGroup", "speciesCode", "age", "biomass")}}
}
\value{
Returns a vector of pixelGroup in the original order of the input \code{pixelCohortData}.
This should likely be added to the \code{pixelCohortData} object immediately.
}
\description{
Add the correct \code{pixelGroups} to a \code{pixelCohortData} object
}
