% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{genericExtract}
\alias{genericExtract}
\title{GENERIC EXTRACT POINTS}
\usage{
genericExtract(x, y, field = NULL, ...)
}
\arguments{
\item{x}{a raster or polygon object (\code{sp}, \code{raster} or \code{terra})}

\item{y}{a points spatial object (\code{sp}, \code{sf}, or \code{terra})}

\item{field}{character. The field(s) to extract when x is a polygon.
If \code{NULL}, all fields are extracted and returned. IDs of y are always
returned (\code{ID} column) .}

\item{...}{passed to \code{terra::extract}}
}
\value{
a data.table with extracted values and an \code{ID} column of
y point IDs
}
\description{
Extracts points from raster layers using the
original raster layer projection.
}
