% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfire.R
\name{getWildfire_NFI}
\alias{getWildfire_NFI}
\title{Download and prepare raster fire data from NFI}
\usage{
getWildfire_NFI(dPath, rasterToMatch, url = NULL)
}
\arguments{
\item{dPath}{destination path for archive}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{url}{location from which to download the wildfire raster layer(s)}
}
\value{
a raster with values representing fire year 1985-2015
}
\description{
Download and prepare raster fire data from NFI
}
