% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R, R/prepSpeciesLayers.R
\name{makeVegTypeMap}
\alias{makeVegTypeMap}
\alias{loadkNNSpeciesLayersValidation}
\alias{prepSpeciesLayers_KNN2011}
\title{Make a vegetation type map from a stack of species abundances}
\usage{
makeVegTypeMap(speciesStack, vegLeadingProportion, mixed, ...)

loadkNNSpeciesLayersValidation(
  dPath,
  rasterToMatch,
  studyArea,
  sppEquiv,
  knnNamesCol = "KNN",
  sppEquivCol,
  thresh = 1,
  url,
  ...
)

prepSpeciesLayers_KNN2011(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  thresh = 10,
  ...
)
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of species abundances.
This must be one \code{RasterLayer} per species.}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{mixed}{Deprecated. See \code{mixedType} argument to \link{vegTypeMapGenerator}.}

\item{...}{Additional arguments passed to \code{\link[reproducible:Cache]{reproducible::Cache()}}
and \code{\link[=equivalentName]{equivalentName()}}.}

\item{dPath}{path to the data directory}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{studyArea}{A \code{SpatVector} or \code{sf} polygons object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{sppEquiv}{table with species name equivalencies between naming conventions
(see \link{sppEquivalencies_CA}, also \code{?sppEquiv}).
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{LandR::sppEquivalencies_CA}.}

\item{knnNamesCol}{character string indicating the column in \code{sppEquiv}
containing kNN species names.
Default \code{"KNN"} for when \code{sppEquivalencies_CA} is used.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv} (see \link{sppEquivalencies_CA}).
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If \code{sppEquiv} is not provided and \code{mixedType == 2}, will use the column of
\code{LandR::sppEquivalencies_CA} that best matches the species names.}

\item{thresh}{the minimum number of pixels where the species must have
\code{biomass > 0} to be considered present in the study area.
Defaults to 1.}

\item{url}{the source url for the data, passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{outputPath}{path to output directory}
}
\value{
A factor raster

A raster stack of percent cover layers by species.
}
\description{
\code{makeVegTypeMap} is a wrapper around \code{vegTypeMapGenerator}
that works from a species stack of percent cover.
These do not have to sum to 100\%\%

Downloads the 2011 kNN species cover layers from the Canadian Forestry Service,
National Inventory System, for validation.
}
