% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{makeVegTypeMap}
\alias{makeVegTypeMap}
\title{Make a vegetation type map from a stack of species abundances}
\usage{
makeVegTypeMap(speciesStack, vegLeadingProportion, mixed, ...)
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of species abundances.
This must be one \code{RasterLayer} per species.}

\item{vegLeadingProportion}{See vegTypeMapGenerator}

\item{mixed}{Deprecated. See vegTypeMapGenerator and \code{mixedType}}

\item{...}{Other arguments passed to vegTypeMapGenerator, i.e.,
\code{vegLeadingProportion}, \code{mixedType}, \code{sppEquiv},
\code{sppEquivCol}, \code{colors}, \code{pixelGroupColName}, and \code{doAssertion}}
}
\value{
A factor raster
}
\description{
\code{makeVegTypeMap} is a wrapper around \code{vegTypeMapGenerator}
that works from a species stack of percent cover. These do not have
to sum to 100%
}
