% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R, R/prepSpeciesLayers.R
\name{makeVegTypeMap}
\alias{makeVegTypeMap}
\alias{loadkNNSpeciesLayersValidation}
\alias{prepSpeciesLayers_KNN2011}
\title{Make a vegetation type map from a stack of species abundances}
\usage{
makeVegTypeMap(speciesStack, vegLeadingProportion, mixed, ...)

loadkNNSpeciesLayersValidation(
  dPath,
  rasterToMatch,
  studyArea,
  sppEquiv,
  knnNamesCol = "KNN",
  sppEquivCol,
  thresh = 1,
  url,
  ...
)

prepSpeciesLayers_KNN2011(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  thresh = 10,
  ...
)
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of species abundances.
This must be one \code{RasterLayer} per species.}

\item{vegLeadingProportion}{See \code{vegTypeMapGenerator}.}

\item{mixed}{Deprecated. See \code{mixedType} argument to \code{vegTypeMapGenerator}.}

\item{...}{Additional arguments passed to \code{\link[reproducible]{Cache}}
and \code{\link{equivalentName}}.}

\item{dPath}{path to the data directory}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{studyArea}{A \code{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible]{prepInputs}}.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{knnNamesCol}{character string indicating the column in \code{sppEquiv}
containing kNN species names.
Default \code{"KNN"} for when \code{sppEquivalencies_CA} is used.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{thresh}{the minimum number of pixels where the species must have
\code{biomass > 0} to be considered present in the study area.
Defaults to 1.}

\item{url}{the source url for the data, passed to \code{\link[reproducible]{prepInputs}}}

\item{destinationPath}{path to data directory}

\item{outputPath}{path to output directory}
}
\value{
A factor raster

A raster stack of percent cover layers by species.
}
\description{
\code{makeVegTypeMap} is a wrapper around \code{vegTypeMapGenerator}
that works from a species stack of percent cover. These do not have
to sum to 100%

Downloads the 2011 kNN species cover layers from the Canadian Forestry Service,
National Inventory System, for validation.
}
