\docType{package}
\name{landr-package}
\alias{landr}
\alias{landr-package}
\alias{landrLibrary}
\title{Landings data from disparate source as R-binaries}
\description{
  More about what it does (maybe more than one line)
}
\details{
  Version 0.1.0: 7. March 2014
}
\author{
  Various folks at MRI
}
\seealso{
  link to other documents
}

