% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR-package.R
\name{LandR-package}
\alias{LandR-package}
\title{\code{LandR} package}
\description{
Utilities for 'LandR' suite of landscape simulation models.
These models simulate forest vegetation dynamics based on LANDIS-II, and
incorporate fire and insect disturbance, as well as other important ecological
processes. Models are implemented as 'SpaDES' modules.
}
