% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR-package.R
\docType{package}
\name{LandR-package}
\alias{LandR}
\alias{LandR-package}
\title{\code{LandR} package}
\description{
Utilities for 'LandR' suite of landscape simulation models.
These models simulate forest vegetation dynamics based on LANDIS-II, and
incorporate fire and insect disturbance, as well as other important ecological
processes. Models are implemented as 'SpaDES' modules.
}
\details{
------------------------------------------------------------------------------------------
}
\section{Package options}{


\code{LandR} packages use the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{LandR.assertions}: If \code{TRUE}, additional code checks are run
        during function calls. Default \code{FALSE}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://landr.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/LandR}
  \item Report bugs at \url{https://github.com/PredictiveEcology/LandR/issues}
}

}
\author{
\strong{Maintainer}: Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})

Authors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
