% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccMapGenerator.R
\name{lccMapGenerator}
\alias{lccMapGenerator}
\title{Calculate landcover classes based on species cohorts}
\usage{
lccMapGenerator(
  cohortData,
  pixelGroupMap,
  lccTable = NULL,
  deciduousCoverDiscount = 0.8418911,
  vegLeadingProportion = 0.75,
  decidousSpp = NULL,
  rstLCC
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{lccTable}{\code{data.table} (with columns \code{leading} and \code{LCC}) defining the mappings
between leading species and landcover classes.}

\item{deciduousCoverDiscount}{numeric between 0 and 1 that translates \%\% cover to \%\% biomass.
It assumes all hardwoods are equivalent; all softwoods are equivalent;
and that \%\% cover of hardwoods will be an overesimate of the \%\%\% biomass of hardwoods.
Hardwoods in Canada have a much wider canopy than softwoods.
E.g., 30\%\% cover of hardwoods might translate to 20\%\% biomass of hardwoods.
\strong{NB:} the default \code{deciduousCoverDiscount} value was estimated from NWT data in March 2020,
and may not be useful for other study areas.}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{decidousSpp}{character vector of deciduous species names. If not supplied, will
attempt to match deciduous species names in \link{sppEquivalencies_CA} to those in \code{cohortData}.}

\item{rstLCC}{landcover raster}
}
\value{
raster of the same type as \code{pixelGroupMap}, containing LCC values
}
\description{
Convert \code{cohortData} into landcover class based on leading species and density.
\enumerate{
\item Classify each pixel as "broadleaf", "mixedwood", or "conifer";
\item Define the openness of each pixel as "dense", "open", or "sparse";
\item Combine these into cover classes for forested cells.
}
}
\details{
By default, landcover classes are coded to match the Land Cover Map of Canada 2005 at 250m
(Latifovic & Pouliot, 2005).

Code modified from
\url{https://github.com/tati-micheletti/caribouRSF_NT/blob/main/R/makeLCCfromCohortData.R}.
}
\note{
Sparse classes:
Even though the default LCC2005 input does \emph{not} have "mixed sparse" or "broadleaf sparse",
these are likely to be produced in simulations, starting from "conifer sparse" sites
(i.e. broadleaf trees might grow on these sites, and eventually convert these into mixed or
broadleaf sparse sites).
}
\references{
Latifovic, R. & Pouliot, D. (2005). Multitemporal land cover mapping for Canada:
methodology and products. Canadian Journal of Remote Sensing, 31, 347–363.
}
\author{
Tati Micheletti and Alex Chubaty
}
