% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{loadCASFRI}
\alias{loadCASFRI}
\title{Load CASFRI data}
\usage{
loadCASFRI(CASFRIRas, attrFile, headerFile, sppEquiv, sppEquivCol,
  type = c("cover", "age"))
}
\arguments{
\item{CASFRIRas}{TODO: description needed}

\item{attrFile}{TODO: description needed}

\item{headerFile}{TODO: description needed}

\item{sppEquiv}{table with species name equivalencies between the
kNN format and the final naming format.
See \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{type}{Character string. Either \code{"cover"} or \code{"age"}.}
}
\value{
TODO: description needed
}
\description{
TODO: description needed
}
