% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{loadkNNSpeciesLayers}
\alias{loadkNNSpeciesLayers}
\title{Load kNN species layers from online data repository}
\usage{
loadkNNSpeciesLayers(
  dPath = tempdir(),
  rasterToMatch,
  studyArea,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppNameVector = NULL,
  knnNamesCol = "KNN",
  sppEquivCol = "Boreal",
  thresh = 1,
 
    url = "http://ftp.maps.canada.ca/pub/nrcan_rncan/Forests_Foret/canada-forests-attributes_attributs-forests-canada/2001-attributes_attributs-2001/",
  ...
)
}
\arguments{
\item{dPath}{path to the data directory}

\item{rasterToMatch}{Defaults to a random place in central Alberta, Canada with
metadata same as KNN layers (res 250m, proj)}

\item{studyArea}{A \code{SpatialPolygons} object. Defaults to a random area in Central Alberta}

\item{sppEquiv}{Default is LandR::sppEquivalencies_CA}

\item{sppNameVector}{An option character vector of species names
(as provided in one column of \code{sppEquiv}) that
can be a subset of all the species in the \code{sppEquivCol} (or any one column
in \code{sppEquiv})}

\item{knnNamesCol}{character string indicating the column in \code{sppEquiv}
containing kNN species names.
Default \code{"KNN"} for when \code{sppEquivalencies_CA} is used.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
If not supplied, it will attempt to use "Boreal", if it exists,
otherwise it will use the first column in \code{sppEquiv}}

\item{thresh}{the minimum number of pixels where the species must have
\code{biomass > 0} to be considered present in the study area.
Defaults to 1.}

\item{url}{the source url for the data, passed to \code{\link[reproducible]{prepInputs}}}

\item{...}{Additional arguments passed to \code{\link[reproducible]{Cache}}
and \code{\link{equivalentName}}.}
}
\value{
A raster stack of percent cover layers by species.
}
\description{
TODO: description needed
}
