% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{loadSCANFISpeciesLayers}
\alias{loadSCANFISpeciesLayers}
\title{Load SCANFI species layers from online data repository}
\usage{
loadSCANFISpeciesLayers(
  dPath,
  rasterToMatch = NULL,
  studyArea = NULL,
  sppEquiv,
  year = 2020,
  SCANFINamesCol = "SCANFI",
  sppEquivCol = "SCANFI",
  thresh = 10,
  url = NULL,
  ...
)
}
\arguments{
\item{dPath}{path to the data directory}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{studyArea}{A \verb{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{year}{which year's layers should be retrieved? One of 2000, 2010, or 2020 (default).}

\item{SCANFINamesCol}{character string indicating the column in \code{sppEquiv}
containing kNN species names.
Default \code{"NFI"} for when \code{sppEquivalencies_CA} is used.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{thresh}{the minimum percent cover a species must have (per pixel)
to be considered present in the study area.
Defaults to 10.}

\item{url}{the source url for the data, default is KNN 2011 dataset
(\url{https://ftp.maps.canada.ca/pub/nrcan_rncan/Forests_Foret/canada-forests-attributes_attributs-forests-canada/2011-attributes_attributs-2011/})}

\item{...}{Additional arguments passed to \code{\link[reproducible:Cache]{reproducible::Cache()}}
and \code{\link[=equivalentName]{equivalentName()}}. Also valid: \code{outputPath}, and \code{studyAreaName}.}
}
\value{
A raster stack of percent cover layers by species.
}
\description{
Load SCANFI species layers from online data repository
}
