% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{makeAndCleanInitialCohortData}
\alias{makeAndCleanInitialCohortData}
\title{Generate initial \code{cohortData} table}
\usage{
makeAndCleanInitialCohortData(inputDataTable, sppColumns,
  pixelGroupBiomassClass)
}
\arguments{
\item{inputDataTable}{A \code{data.table} with columns described above.}

\item{sppColumns}{A vector of the names of the columns in \code{inputDataTable} that
represent percent cover by species}

\item{pixelGroupBiomassClass}{Round B to the nearest \code{pixelGroupBiomassClass}
to establish unique pixelGroups}
}
\description{
Takes a single data.table input, which has the following columns:
c("age", "logAge", "initialEcoregionCode", "totalBiomass", "Abie_sp",
"Pice_gla", "Pice_mar", "Pinu_sp", "Popu_sp", "pixelIndex", "lcc")
}
\author{
Eliot McIntire
}
