% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{makeAndCleanInitialCohortData}
\alias{makeAndCleanInitialCohortData}
\title{Generate initial \code{cohortData} table}
\usage{
makeAndCleanInitialCohortData(inputDataTable, sppColumns,
  pixelGroupBiomassClass, doAssertion = getOption("LandR.assertions",
  TRUE))
}
\arguments{
\item{inputDataTable}{A \code{data.table} with columns described above.}

\item{sppColumns}{A vector of the names of the columns in \code{inputDataTable} that
represent percent cover by species}

\item{pixelGroupBiomassClass}{Round B to the nearest \code{pixelGroupBiomassClass}
to establish unique pixelGroups}

\item{doAssertion}{Turns on/off assertion. Defaults to \code{getOption("LandR.assertions")}}
}
\description{
Takes a single \code{data.table} input, which has the following columns in addition to
others that will be labelled with species name, and contain percent cover of each:
\itemize{
  \item age
  \item logAge
  \item initialEcoregionCode
  \item totalBiomass
  \item pixelIndex
  \item lcc
}
}
\author{
Eliot McIntire
}
