% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{.createCohortData}
\alias{.createCohortData}
\alias{makeAndCleanInitialCohortData}
\title{Generate template \code{cohortData} table}
\usage{
.createCohortData(inputDataTable, pixelGroupBiomassClass,
  doAssertion = getOption("LandR.assertions", TRUE), rescale = TRUE)

makeAndCleanInitialCohortData(inputDataTable, sppColumns,
  pixelGroupBiomassClass, doAssertion = getOption("LandR.assertions",
  TRUE), doSubset = TRUE)
}
\arguments{
\item{inputDataTable}{A \code{data.table} with columns described above.}

\item{pixelGroupBiomassClass}{Round B to the nearest \code{pixelGroupBiomassClass}
to establish unique pixelGroups}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{rescale}{Logical. If \code{TRUE}, the default, cover for each species will be rescaled
so all cover in pixelGroup or pixel sums to 100.}

\item{sppColumns}{A vector of the names of the columns in \code{inputDataTable} that
represent percent cover by species}

\item{doSubset}{Turns on/off subsetting. Defaults to \code{TRUE}.}
}
\description{
Takes a single \code{data.table} input, which has the following columns in addition to
others that will be labelled with species name, and contain percent cover of each:
\itemize{
  \item \code{pixelIndex} (integer)
  \item \code{age} (integer)
  \item \code{logAge} (numeric)
  \item \code{initialEcoregionCode} (factor)
  \item \code{totalBiomass} (integer)
  \item \code{lcc} (integer)
  \item \code{rasterToMatch} (integer)
  \item \code{speciesCode} (factor)
  \item \code{cover} (integer)
  \item \code{coverOrig} (integer)
  \item \code{B} (integer)
}
}
\author{
Eliot McIntire
}
\keyword{internal}
