% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{makeCohortDataFiles}
\alias{makeCohortDataFiles}
\title{Make the \code{cohortData} table, while modifying the temporary
\code{pixelCohortData} that will be used to prepare other files.}
\usage{
makeCohortDataFiles(
  pixelCohortData,
  columnsForPixelGroups,
  speciesEcoregion,
  pixelGroupBiomassClass,
  pixelGroupAgeClass,
  minAgeForGrouping = 0,
  rmImputedPix = FALSE,
  imputedPixID,
  pixelFateDT
)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{columnsForPixelGroups}{Default columns that define pixel groups}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-{ecoregion}-specific species
trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar
biophysical characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.}

\item{pixelGroupBiomassClass}{Integer. When assigning \code{pixelGroup} membership, this defines
the resolution of biomass that will be considered 'the same \code{pixelGroup}', e.g., if it is
100, then 5160 and 5240 will be the same}

\item{pixelGroupAgeClass}{Integer. When assigning \code{pixelGroup} membership, this defines the
resolution of ages that will be considered 'the same \code{pixelGroup}', e.g., if it is 10,
then 6 and 14 will be the same.}

\item{minAgeForGrouping}{Minimum age for regrouping. This may be because there is a source of
ages for young stands/trees that is very reliable, such as a fire database.
Ages below this will not be grouped together. Defaults to -1, meaning treat all ages equally.
If this is related to known ages from a high quality database, then use age of the oldest
trees in that database.}

\item{rmImputedPix}{Should imputed pixels be removed}

\item{imputedPixID}{a vector of IDs of pixels that suffered data imputation.}

\item{pixelFateDT}{A \code{data.table} of \code{pixelFateDT}; if none provided, will make an empty one.}
}
\value{
A list with a modified \code{pixelCohortData}, \code{cohortData}, and \code{pixelFateDT}
\code{data.table}s.
}
\description{
Takes a \code{pixelCohortData} table (see \code{makeAndCleanInitialCohortData}),
the \code{speciesEcoregion} list and returns a modified \code{pixelCohortData} and
the \code{cohortData} tables to be used in the simulation.
This function mainly removes unnecessary columns from \code{pixelCohortData},
subsets pixels with \code{biomass > 0}, generates \code{pixelGroups},
and adds \code{ecoregionGroup} and \code{totalBiomass} columns to \code{pixelCohortData}.
\code{cohortData} is then created by subsetting unique combinations of \code{pixelGroup} and
whatever columns are listed in \code{columnsForPixelGroups}.
The resulting \code{cohortData} table has the following columns:
\itemize{
\item \code{speciesCode} (factor)
\item \code{ecoregionGroup} (factor)
\item \code{pixelGroup} (integer)
\item \code{age} (integer)
\item \code{B} (integer)
}
}
