% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{makeEcoregionDT}
\alias{makeEcoregionDT}
\title{Make the \code{ecoregion} table}
\usage{
makeEcoregionDT(pixelCohortData, speciesEcoregion)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values.}
}
\value{
A \code{data.table} with ecoregion codes and their active status per \code{pixelID}.
}
\description{
This function creates a table containing pixel-wise ecoregion codes and whether they
  are "active" (have biomass > 0) or not for simulation. Unlike \code{ecoregionProducer},
  this function creates the \code{ecoregion} table from pixel information contained in
  \code{pixelCohortData}
}
