% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{makePickellStack}
\alias{makePickellStack}
\title{makePickellStack}
\usage{
makePickellStack(PickellRaster, sppEquiv, sppEquivCol, destinationPath)
}
\arguments{
\item{PickellRaster}{TODO: description needed}

\item{sppEquiv}{table with species name equivalencies between the
kNN format and the final naming format.
See \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{destinationPath}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
TODO: description and title needed
}
