% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{makePickellStack}
\alias{makePickellStack}
\title{\code{makePickellStack}}
\usage{
makePickellStack(PickellRaster, sppEquiv, sppEquivCol, destinationPath)
}
\arguments{
\item{PickellRaster}{TODO: description needed}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}
}
\value{
TODO: description needed
}
\description{
TODO: description and title needed
}
