% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{makePickellStack}
\alias{makePickellStack}
\title{makePickellStack}
\usage{
makePickellStack(PickellRaster, sppEquiv, sppEquivCol, destinationPath)
}
\arguments{
\item{PickellRaster}{TODO: description needed}

\item{sppEquiv}{TODO: description needed}

\item{sppEquivCol}{TODO: description needed}

\item{destinationPath}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
TODO: description and title needed
}
