% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makePixelGroupMap}
\alias{makePixelGroupMap}
\title{Create \code{makePixelGroupMap}}
\usage{
makePixelGroupMap(pixelCohortData, rasterToMatch)
}
\arguments{
\item{pixelCohortData}{The full \code{cohortData} \code{data.table}}

\item{rasterToMatch}{a \code{rasterToMatch} (e.g. the one used throughout the simulation)}
}
\value{
a raster with pixel groups
}
\description{
Create the \code{makePixelGroupMap} raster containing \code{pixelGroups} in \code{pixelCohortData}.
}
