% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makePixelTable}
\alias{makePixelTable}
\title{Make \code{pixelTable} from biomass, age, land-cover and species cover data}
\usage{
makePixelTable(
  speciesLayers,
  species,
  standAgeMap,
  ecoregionFiles,
  biomassMap,
  rasterToMatch,
  rstLCC,
  pixelGroupAgeClass = 1,
  printSummary = TRUE,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{speciesLayers}{stack of species layers rasters}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.}

\item{standAgeMap}{raster of stand age}

\item{ecoregionFiles}{A list with two objects: the \code{ecoregionMap} and a table summarizing
its information per \code{pixelID.} See \code{ecoregionProducer}.}

\item{biomassMap}{raster of total stand biomass}

\item{rasterToMatch}{a \code{rasterToMatch} (e.g., the one used throughout the simulation)}

\item{rstLCC}{raster of land-cover class}

\item{pixelGroupAgeClass}{When assigning \code{pixelGroup} membership, this defines the resolution
of ages that will be considered 'the same \code{pixelGroup}', e.g., if it is 10, then 6 and 14
will be the same}

\item{printSummary}{Logical. If \code{TRUE}, the default, a print out of the
\code{summary(pixelTable)} will occur.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\value{
A \code{data.table} as many rows as non-NA pixels in \code{rasterToMath} and
 the columns containing pixel data from the input raster layers.
}
\description{
Make \code{pixelTable} from biomass, age, land-cover and species cover data
}
