% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makeSpeciesEcoregion}
\alias{makeSpeciesEcoregion}
\title{Make speciesEcoregion using statistically estimated maxB, maxANPP and establishment probabilities}
\usage{
makeSpeciesEcoregion(cohortDataNoBiomass, cohortDataShort,
  cohortDataShortNoCover, species, modelCover, modelBiomass,
  successionTimestep, currentYear)
}
\arguments{
\item{cohortDataNoBiomass}{a subset of cohortData}

\item{cohortDataShort}{a subset of cohortData}

\item{cohortDataShortNoCover}{a subset of cohortData}

\item{species}{a \code{data.table} that has species traits such as longevity, shade tolerance, etc.}

\item{modelCover}{statistical model of species presence/absence}

\item{modelBiomass}{statistical model of species biomass}

\item{successionTimestep}{The time between successive seed dispersal events.}

\item{currentYear}{\code{time(sim)}

#' @return
A \code{speciesEcoregion} \code{data.table} with added columns for parameters
  \code{maxB}, \code{maxANPP} and \code{establishprob}}
}
\description{
Make speciesEcoregion using statistically estimated maxB, maxANPP and establishment probabilities
}
