% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{makeSpeciesEcoregion}
\alias{makeSpeciesEcoregion}
\title{Create \code{speciesEcoregion}}
\usage{
makeSpeciesEcoregion(
  cohortDataBiomass,
  cohortDataShort,
  cohortDataShortNoCover,
  species,
  modelCover,
  modelBiomass,
  successionTimestep,
  currentYear
)
}
\arguments{
\item{cohortDataBiomass}{a subset of \code{cohortData}}

\item{cohortDataShort}{a subset of \code{cohortData}}

\item{cohortDataShortNoCover}{a subset of \code{cohortData}}

\item{species}{a \code{data.table} with species traits such as longevity, shade tolerance, etc.}

\item{modelCover}{statistical model of species presence/absence}

\item{modelBiomass}{statistical model of species biomass}

\item{successionTimestep}{The time between successive seed dispersal events.}

\item{currentYear}{\code{time(sim)}}
}
\description{
Use statistically estimated \code{maxB}, \code{maxANPP} and establishment probabilities
to generate \code{specieEcoregion} table.
}
\details{
See Details.
}
\section{\code{establishprob}}{

This section takes the cover as estimated from the mature tree cover and
partitions it between resprouting and seeds Unfortunately, establishment by
seed is not independent of resprouting, i.e., some pixels would have both
Since we don't know the level of independence, we can't correctly assess how
much to discount the two. If there is resprouting > 0, then this is the
partitioning:
\code{establishprob = f(establishprob + resproutprob + jointEstablishProbResproutProb)}
If \code{jointEstablishProbResproutProb} is 0, then these are independent events
and the total cover probability can be partitioned easily between seeds and
resprout. This is unlikely ever to be the case. We are picking 50\% overlap as
a number that is better than 0 (totally independent probabilities, meaning no
pixel has both seeds and resprout potential) and  100\% overlap (totally
dependent probabilities, i.e., every pixel where there is seeds will also be
a pixel with resprouting) This is expressed with the "* 0.5" in the code.

#' @return
A \code{speciesEcoregion} \code{data.table} with added columns for parameters
\code{maxB}, \code{maxANPP} and \code{establishprob}
}

