% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{makeVegTypeMap}
\alias{makeVegTypeMap}
\title{Make a vegetation type map from a stack of species abundances}
\usage{
makeVegTypeMap(speciesStack, vegLeadingProportion, mixed = TRUE)
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of species abundances.
This must be one \code{RasterLayer} per species.}

\item{vegLeadingProportion}{The threshold as a proportion of the total abundance
that a species must have to be considered a "pure" stand of that type.
If no species reaches this proportion, then the pixel will be 'Mixed'.}

\item{mixed}{Logical. If \code{TRUE}, then a mixed pixel value will be identified and given
(see \code{vegLeadingProportion} argument)}
}
\value{
A factor raster
}
\description{
Make a vegetation type map from a stack of species abundances
}
