% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mergeSppRaster}
\alias{mergeSppRaster}
\title{Merge species percent-cover rasters}
\usage{
mergeSppRaster(sppMerge, speciesLayers, sppEquiv, column, suffix, dPath, ...)
}
\arguments{
\item{sppMerge}{TODO}

\item{speciesLayers}{stack of species layers rasters}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{column}{TODO}

\item{suffix}{TODO}

\item{dPath}{destination path TODO}

\item{...}{Additional arguments TODO}
}
\description{
Used internally in \code{overlayStacks}.
}
\keyword{internal}
