% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{mergeSppRaster}
\alias{mergeSppRaster}
\title{Merge species pecent-cover rasters}
\usage{
mergeSppRaster(sppMerge, speciesLayers, sppEquiv, column, suffix, dPath,
  ...)
}
\arguments{
\item{sppMerge}{TODO}

\item{speciesLayers}{TODO}

\item{sppEquiv}{TODO}

\item{column}{TODO}

\item{suffix}{TODO}

\item{dPath}{destination path TODO}

\item{...}{Additonal arguments TODO}
}
\description{
Used internally in \code{overlayStacks}.
}
\keyword{internal}
