% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{middlePixel}
\alias{middlePixel}
\title{Return the (approximate) middle pixel on a raster}
\usage{
middlePixel(ras)
}
\description{
This calculation is different depending on whether
the \code{nrow(ras)} and \code{ncol(ras)} are even or odd.
}
