% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{NAcover2zero}
\alias{NAcover2zero}
\title{Convert NA's in speciesLayers to zeros}
\usage{
NAcover2zero(speciesLayers, rasterToMatch)
}
\arguments{
\item{speciesLayers}{A \code{RasterStack} or \code{RasterLayer} that
should contain species cover data in the study area}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}
}
\value{
the \code{speciesLayers} with 0s in pixels that had NAs
}
\description{
Pixels that have NAs but are inside \code{rasterToMatch}
may need to be converted to 0s as they can could still potentially
be forested
}
