% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayLCCs.R
\name{overlayLCCs}
\alias{overlayLCCs}
\title{Overlay different LCC data sources}
\usage{
overlayLCCs(LCCs, forestedList, outputLayer, NAcombos,
  forestEquivalencies = NULL)
}
\arguments{
\item{LCCs}{A named list or named \code{RasterStack} of rasters whose content is Land Cover Class}

\item{forestedList}{A named list of same length and names as \code{LCCs} indicating
which classes in each LCC raster are 'forested', either permanent or transient}

\item{outputLayer}{A character string that matches one of the named elements in \code{LCCs}.
This will be the classification system returned.}

\item{NAcombos}{A character string that will be parsed within the "forestEquivalencies" table.
It should be a set of conditions that == 0, e.g., \code{"LCC2005 == 0"} or \code{"CC == 0 | LCC2005 == 0"},
where \code{0} is the non-forested pixels based on converting LCCs and forestedList to 1s and 0s.}

\item{forestEquivalencies}{A data.frame or NULL. If \code{NULL}, this function will
derive this table automatically from the other arguments. Otherwise, the user must
provide a data.frame with \code{length(LCCs) + 1} columns, and \code{2 ^ length(LCCs)}
rows. Currently not used.}
}
\description{
Overlay different LCC data sources
}
\author{
Eliot McIntire
}
