% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayLCCs.R
\name{overlayLCCs}
\alias{overlayLCCs}
\title{Overlay different LCC data sources}
\usage{
overlayLCCs(LCCs, forestedList, outputLayer, NAcondition, NNcondition,
  remapCondition, classesToReplace, availableERC_by_Sp,
  forestEquivalencies = NULL)
}
\arguments{
\item{LCCs}{A named list or named \code{RasterStack} of layers whose content is Land Cover Class}

\item{forestedList}{A named list of same length and names as \code{LCCs} indicating
which classes in each LCC raster are 'forested', either permanent or transient}

\item{outputLayer}{A character string that matches one of the named elements in \code{LCCs}.
This will be the classification system returned.}

\item{NAcondition}{A character string of a vectorized logical statement that
will be parsed within the \code{forestEquivalencies} table.
It should be a set of conditions with \code{== 0}, i.e., non-forested.
Examples:, e.g., \code{"LCC2005 == 0"} or \code{"CC == 0 | LCC2005 == 0"},
where \code{0} is the non-forested pixels based on converting LCCs and
\code{forestedList} to 1s and 0s.}

\item{NNcondition}{A character string of a vectorized logical statement that
will be parsed within the "forestEquivalencies" table.
It should be a set of conditions with \code{== 0}, i.e., non-forested.
Examples:, e.g., \code{"LCC2005 == 0"} or \code{"CC == 0 | LCC2005 == 0"},
where \code{0} is the non-forested pixels based on converting LCCs and
\code{forestedList} to 1s and 0s.}

\item{remapCondition}{Not yet implemented. This would be for a situation where
2 LCC layers are provided, one has information in a pixel, but not the one
which is \code{outputLayer}, so this needs a reclassify or remap.}

\item{classesToReplace}{Passed to \code{convertUnwantedLCC}, for the pixels where
\code{NNcondition} is \code{TRUE}}

\item{availableERC_by_Sp}{Passed to \code{convertUnwantedLCC}, for the pixels where
\code{NNcondition} is \code{TRUE}. If this is \code{NULL}, then it will be
created internally with all pixels with:
\code{data.table(initialEcoregionCode = LCCs[[outputLayer]][])}}

\item{forestEquivalencies}{A \code{data.frame} or \code{NULL}.
If \code{NULL}, this function will derive this table automatically from the
other arguments. Otherwise, the user must provide a \code{data.frame} with
\code{length(LCCs) + 1} columns, and \code{2 ^ length(LCCs)} rows.
Currently not used.}
}
\description{
Overlay different LCC data sources
}
\author{
Eliot McIntire
}
