% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{pixelFate}
\alias{pixelFate}
\title{Create or amend data to a \code{pixelFateDT} object}
\usage{
pixelFate(
  pixelFateDT,
  fate = NA_character_,
  pixelsRemoved = 0,
  runningPixelTotal = NA_integer_
)
}
\arguments{
\item{pixelFateDT}{A \code{pixelFateDT} \code{data.table} with 3 columns: \code{fate},
\code{pixelsRemoted}, and \code{runningPixelTotal}.}

\item{fate}{A character string (length 1) describing in words the change}

\item{pixelsRemoved}{A numeric indicating how many pixels were removed due to the \code{fate}.}

\item{runningPixelTotal}{an optional numeric with new, running total. If not supplied,
it will be calculated from the last row of \code{pixelFateDT} \code{runningTotal} minus the
\code{pixelsRemoved}}
}
\value{
A \code{pixelFateDT} object, updated with one extra row.
}
\description{
Create or amend data to a \code{pixelFateDT} object
}
