% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{plantNewCohorts}
\alias{plantNewCohorts}
\title{Create new cohorts based on provenance table with unique \code{pixelGroup} and add to \code{cohortData}}
\usage{
plantNewCohorts(
  newPixelCohortData,
  cohortData,
  pixelGroupMap,
  time,
  provenanceTable,
  speciesEcoregion,
  successionTimestep
)
}
\arguments{
\item{newPixelCohortData}{the cohorts that were harvested}

\item{cohortData}{A \code{data.table} with columns:
\code{pixelGroup}, \code{ecoregionGroup}, \code{speciesCode}, \code{age},
\code{B}, \code{mortality}, \code{aNPPAct}, ond \code{sumB}.}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group number
that corresponds exactly to ]\code{pixelGroup} column in \code{cohortData}.}

\item{time}{time of sim}

\item{provenanceTable}{A data.table with three columns: Location, Provenance, and speciesCode.
new cohorts are initiated at the Location speciesEcoregion from the corresponding
speciesEcoregion listed in the Provenance column}

\item{successionTimestep}{successionTimestep used in sim}
}
\value{
A \code{data.table} with a new \code{rbindlist}ed \code{cohortData}
}
\description{
Create new cohorts based on provenance table with unique \code{pixelGroup} and add to \code{cohortData}
}
