% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{plantNewCohorts}
\alias{plantNewCohorts}
\title{Create new cohorts based on provenance table with unique \code{pixelGroup} and add to \code{cohortData}}
\usage{
plantNewCohorts(
  newPixelCohortData,
  cohortData,
  pixelGroupMap,
  time,
  provenanceTable,
  speciesEcoregion,
  successionTimestep
)
}
\arguments{
\item{newPixelCohortData}{the cohorts that were harvested}

\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{provenanceTable}{A \code{data.table} with three columns:
\code{Location}, \code{Provenance}, and \code{speciesCode}.
New cohorts are initiated at the \code{Location} \code{speciesEcoregion} from the corresponding
\code{speciesEcoregion} listed in the \code{Provenance} column}

\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values.}

\item{successionTimestep}{succession timestep used in the simulation}

\item{currentTime}{The current simulation time e.g., \code{time(sim)}.}
}
\value{
A \code{data.table} with a new \code{cohortData}
}
\description{
Create new cohorts based on provenance table with unique \code{pixelGroup} and add to \code{cohortData}
}
