% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot raster objects using ggplot}
\usage{
plot_raster(x, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{A \code{SpatRaster} object}

\item{title}{character, the plot title}

\item{subtitle}{character, the plot subtitle}
}
\value{
ggplot object
}
\description{
Plot raster objects using ggplot
}
