% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVTM}
\alias{plotVTM}
\title{Summary plots of leading vegetation types}
\usage{
plotVTM(
  speciesStack = NULL,
  vtm = NULL,
  vegLeadingProportion = 0.8,
  sppEquiv,
  sppEquivCol,
  colors,
  title = "Leading vegetation types"
)
}
\arguments{
\item{speciesStack}{A \code{SpatRaster}, \code{RasterStack} or \code{RasterBrick}
of percent-cover-by-species layers.}

\item{vtm}{An optional vegetation type map (\code{RasterLayer} or \code{SpatRaster}).
If not supplied, will be produced internally by \code{makeVegTypeMap}.}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{colors}{Named vector of colour codes, named using species names.
NOTE: plot order will follow this order.}

\item{title}{The title to use for the generated plots.}
}
\description{
Create raster of leading vegetation types and \code{Plot} a bar chart summary
and a vegetation type map. NOTE: plot order will follow \code{colors} order.
}
\author{
Eliot McIntire
}
