% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVTM}
\alias{plotVTM}
\title{Summary plots of leading vegetation types}
\usage{
plotVTM(speciesStack = NULL, vtm = NULL, vegLeadingProportion = 0.8,
  sppEquiv, sppEquivCol, colors, title = "Leading vegetation types")
}
\arguments{
\item{speciesStack}{A \code{RasterStack} of percent-cover-by-species layers.}

\item{vtm}{An optional vegetation type map (\code{RasterLayer}).
If not supplied, will be produced internally by \code{makeVegTypeMap}.}

\item{vegLeadingProportion}{The minimum proportion cover required to consider
a species to be the "leading" one. Default 0.8.}

\item{sppEquiv}{a species equivalency table TODO: description needed}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{colors}{Named vector of colour codes, named using species names. NOTE:
plot order will follow this order.}

\item{title}{The title to use for the generated plots.}
}
\description{
Create raster of leading vegetation types and \code{Plot} a bar chart summary
and a vegetation type map. NOTE: plot order will follow \code{colors} order.
}
\author{
Eilot McIntire
}
