% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepEcoregions.R
\name{prepEcoregions}
\alias{prepEcoregions}
\title{Prepare egoregions objects}
\usage{
prepEcoregions(
  ecoregionRst = NULL,
  ecoregionLayer,
  ecoregionLayerField = NULL,
  rasterToMatchLarge,
  rstLCCAdj,
  pixelsToRm,
  cacheTags
)
}
\arguments{
\item{ecoregionRst}{an optional raster object that could be passed to sim, representing ecoregions}

\item{ecoregionLayer}{a spatial polygons object representing ecoregions}

\item{ecoregionLayerField}{the field in ecoregionLayer that represents ecoregions - optional}

\item{rasterToMatchLarge}{the rasterToMatchLarge object from sim}

\item{rstLCCAdj}{raster layer representing landcover adjusted for non-forest classes}

\item{pixelsToRm}{a vector of pixels to remove}

\item{cacheTags}{UserTags to pass to cache}
}
\description{
DESCRIPTION NEEDED
}
