% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputs_NTEMS.R
\name{prepInputs_NTEMS_LCC_FAO}
\alias{prepInputs_NTEMS_LCC_FAO}
\title{Obtain an LCC layer for a given year from NTEMS, with forest matching the FAO definition}
\usage{
prepInputs_NTEMS_LCC_FAO(
  year = 2010,
  disturbedCode = 240,
  resampleMethod = "near",
  ...
)
}
\arguments{
\item{year}{stack of species layers rasters}

\item{disturbedCode}{value assigned to pixels that are forest per FAO definition but not in LCC year}

\item{resampleMethod}{method used when resampling LCC layers to match \code{rasterToMatch}}

\item{...}{passed to \code{prepInputs}}
}
\value{
a \code{SpatRaster} with corrected forest pixels
}
\description{
Obtain an LCC layer for a given year from NTEMS, with forest matching the FAO definition
}
