% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputs_SCANFI_LCC_FAO}
\alias{prepInputs_SCANFI_LCC_FAO}
\title{Obtain an LCC layer for a given year from SCANFI, with forest matching the FAO definition}
\usage{
prepInputs_SCANFI_LCC_FAO(
  year = 2010,
  disturbedCode = 240,
  resampleMethod = "near",
  ...
)
}
\arguments{
\item{year}{data year for LCC data. 2000, 2010, and 2020 possible.}

\item{disturbedCode}{value assigned to pixels that are forest per FAO definition but not in LCC year}

\item{resampleMethod}{method used when resampling LCC layers to match \code{rasterToMatch}}

\item{...}{passed to \code{prepInputs}}
}
\value{
a \code{SpatRaster} with corrected forest pixels
}
\description{
Obtain an LCC layer for a given year from SCANFI, with forest matching the FAO definition
}
