% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputs_SCANFI_LCC_FAO}
\alias{prepInputs_SCANFI_LCC_FAO}
\title{Obtain an LCC layer for a given year from SCANFI, with forest matching the FAO definition}
\usage{
prepInputs_SCANFI_LCC_FAO(
  year = 2020,
  dataVersion = "V2",
  disturbedCode = 240,
  resampleMethod = "near",
  ...
)
}
\arguments{
\item{year}{data year for SCANFI landcover data. 2000, 2010, 2020 possible for V1.
1985, 1990, 1995, 2000, 2005, 2010, 2015, 2020, 2025 possible for V2.}

\item{dataVersion}{Character. SCANFI product version for data. Default is currently "V2".
"V1" also available.}

\item{disturbedCode}{value assigned to pixels that are forest per FAO definition but not in LCC year}

\item{resampleMethod}{method used when resampling LCC layers to match \code{rasterToMatch}}

\item{...}{passed to \code{prepInputs}}
}
\value{
a \code{SpatRaster} with corrected forest pixels
}
\description{
Obtain an LCC layer for a given year from SCANFI, with forest matching the FAO definition
}
