% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{prepInputsCanDEM}
\alias{prepInputsCanDEM}
\title{Get digital elevation map of Canada}
\usage{
prepInputsCanDEM(studyArea, rasterToMatch, destinationPath)
}
\arguments{
\item{studyArea}{A \code{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible]{prepInputs}}.}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}
}
\value{
\code{RasterLayer}
}
\description{
Defaults to using 3 arcmin DEM of Canada in lonlat.
}
