% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a raster of fire perimeters}
\usage{
prepInputsFireYear(..., rasterToMatch, fireField = "YEAR", earliestYear = 1950)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{fireField}{field used to rasterize fire polys}

\item{earliestYear}{the earliest fire date to allow}
}
\value{
a \code{SpatRaster} layer of fire perimeters with fire year values.
}
\description{
Create a raster of fire perimeters
}
\examples{
withr::local_options(list(
  reproducible.useTerra = TRUE,
  reproducible.rasterRead = "terra::rast"
))

randomPoly <- LandR::randomStudyArea(
  size = 1e+8, seed = 5
)

ras2match <- terra::rast(
  randomPoly,
  vals = 1,
  res = 100,
)
ras2match <- terra::mask(ras2match, randomPoly)

firePerimeters <- prepInputsFireYear(
  url = paste0(
    "https://cwfis.cfs.nrcan.gc.ca/downloads/",
    "nfdb/fire_poly/current_version/NFDB_poly.zip"
  ),
  destinationPath = file.path(tempdir(), "ex_prepInputsFireYear"),
  rasterToMatch = ras2match,
  earliestYear = 1950
)

if (interactive()) {
  terra::plot(firePerimeters)
  terra::plot(randomPoly, add = TRUE)
}

withr::deferred_run()

}
