% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a raster of fire perimeters}
\usage{
prepInputsFireYear(..., rasterToMatch, fireField = "YEAR", earliestYear = 1950)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}}

\item{rasterToMatch}{A \code{SpatRaster} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).
\code{rasterToMatch} should correspond to \code{studyArea} if used.}

\item{fireField}{field used to rasterize fire polys}

\item{earliestYear}{the earliest fire date to allow}
}
\value{
a \code{SpatRaster} layer of fire perimeters with fire year values.
}
\description{
Create a raster of fire perimeters
}
\examples{
withr::local_options(list(
  reproducible.useTerra = TRUE,
  reproducible.rasterRead = "terra::rast"
))

targetCRS <- terra::crs(SpaDES.tools::randomStudyArea())
randomPoly <- SpaDES.tools::randomStudyArea(
  center = terra::vect(cbind(-115, 50), crs = targetCRS),
  size = 1e+7,
)

buffExt <- terra::buffer(randomPoly, 1e+3) |> terra::ext()
ras2match <- terra::rast(
  resolution = 10,
  extent = terra::ext(randomPoly),
  crs = terra::crs(randomPoly)
)
ras2match <- terra::rasterize(randomPoly, ras2match)

firePerimeters <- prepInputsFireYear(
  url = paste0(
    "https://cwfis.cfs.nrcan.gc.ca/downloads/",
    "nfdb/fire_poly/current_version/NFDB_poly.zip"
  ),
  destinationPath = file.path(tempdir(), "ex_prepInputsFireYear"),
  rasterToMatch = ras2match,
  earliestYear = 1930
)

if (interactive()) {
  plot(firePerimeters)
  plot(randomPoly, add = TRUE)
}

withr::deferred_run()

}
