% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a rasterof fire polygons}
\usage{
prepInputsFireYear(..., rasterToMatch, field = "YEAR")
}
\arguments{
\item{field}{field used to rasterize fire polys}

\item{rastertoMatch}{template used for fasterize}
}
\description{
wrapper on prepInputs that will rasterize fire polygons.
}
