% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a raster of fire polygons}
\usage{
prepInputsFireYear(..., rasterToMatch, field = "YEAR", earliestYear = 1950)
}
\arguments{
\item{...}{Additional arguments passed to \code{prepInputs}}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{field}{field used to rasterize fire polys}

\item{earliestYear}{the earliest fire date to allow}
}
\description{
Wrapper on \code{prepInputs} that will rasterize fire polygons.
}
