% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a raster of fire perimeters}
\usage{
prepInputsFireYear(..., rasterToMatch, fireField = "YEAR", earliestYear = 1950)
}
\arguments{
\item{...}{Additional arguments passed to \code{prepInputs}}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{fireField}{field used to rasterize fire polys}

\item{earliestYear}{the earliest fire date to allow}
}
\value{
a \code{SpatRaster} layer of fire perimeters with fire year values.
}
\description{
Create a raster of fire perimeters
}
\examples{
library(SpaDES.tools)
library(terra)
library(reproducible)

options("reproducible.useTerra" = TRUE, "reproducible.rasterRead" = "terra::rast")

targetCRS <- crs(randomStudyArea())
randomPoly <- randomStudyArea(
  center = vect(cbind(-115, 50), crs = targetCRS),
  size = 1e+7,
)
buffExt <- buffer(randomPoly, 1e+3) |> ext()
ras2match <- rast(res = 10, ext = ext(randomPoly), crs = crs(randomPoly))
ras2match <- rasterize(randomPoly, ras2match)

firePerimeters <- prepInputsFireYear(
  url = paste0("https://cwfis.cfs.nrcan.gc.ca/downloads",
               "/nfdb/fire_poly/current_version/NFDB_poly.zip"),
  destinationPath = tempdir(),
  rasterToMatch = ras2match,
  earliestYear = 1930
)

if (interactive()) {
  plot(firePerimeters)
  plot(randomPoly, add = TRUE)
}
}
