% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsFireYear}
\alias{prepInputsFireYear}
\title{Create a raster of fire polygons}
\usage{
prepInputsFireYear(..., rasterToMatch, fireField = "YEAR", earliestYear = 1950)
}
\arguments{
\item{...}{Additional arguments passed to \code{prepInputs}}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{fireField}{field used to rasterize fire polys}

\item{earliestYear}{the earliest fire date to allow}
}
\value{
a raster layer of fire perimeters with fire year values.
}
\description{
Wrapper on \code{prepInputs} that will rasterize fire polygons.
}
\examples{
library(SpaDES.tools)
library(raster)
library(reproducible)
randomPoly <- randomStudyArea()
randomPoly
ras2match <- raster(res = 10, ext = extent(randomPoly), crs = crs(randomPoly))
ras2match <- rasterize(randomPoly, ras2match)
tempDir <- tempdir()
cacheRepo <- file.path(tempDir, "cache")

## ideaally, get the firePerimenters layer first
firePerimeters <- Cache(prepInputsFireYear,
                        url = "https://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/NFDB_poly.zip",
                        fun = "sf::st_read",
                        destinationPath = tempDir,
                        rasterToMatch = ras2match)

}
