% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputsLCC}
\alias{prepInputsLCC}
\title{Simple \code{prepInputs} for Canadian LCC data}
\usage{
prepInputsLCC(
  year = 2010,
  destinationPath = asPath("."),
  method = c("ngb", "near"),
  writeTo = NULL,
  ...
)
}
\arguments{
\item{year}{Numeric, either 2010 or 2015. See note re: backwards compatibility for 2005.}

\item{destinationPath}{Character string of a directory in which to download
and save the file that comes from \code{url} and is also where the function
will look for \code{archive} or \code{targetFile}. NOTE (still experimental):
To prevent repeated downloads in different locations, the user can also set
\code{options("reproducible.destinationPathShared")} to one or more local file paths to
search for the file before attempting to download. Default for that option is
\code{NULL} meaning do not search locally. The previous name
\code{options("reproducible.inputPaths")} is still accepted as a backwards-compatible
alias.}

\item{method}{passed to \link[terra:intersect]{terra::intersect} or \link[raster:intersect]{raster::intersect},
and \link[reproducible:prepInputs]{reproducible::prepInputs}}

\item{writeTo}{passed to \link[reproducible:prepInputs]{reproducible::prepInputs}}

\item{...}{Arguments passed to \code{terra::mask} (for \code{maskTo}), \code{terra::project} (for \code{projectTo})
or \code{terra::writeRaster} (for \code{writeTo}) and not used for \code{cropTo}, as well \code{postProcess}'s
\code{rasterToMatch} and \code{studyArea} arguments (see below). Commonly used arguments might be
\code{method}, \code{touches}, and \code{datatype}. If \code{filename} is passed, it will be ignored; use
\verb{writeTo = }.}
}
\description{
A wrapper around \code{prepInputs} for the Canadian Land Cover Classification product(s).
}
\note{
As of May 2021, NRCAN no longer provides/hosts the LCC2005 data.
A privately hosted version of the data is available to maintain backwards compatibility,
but new users/projects should use the 2010 (or newer) data.
}
