% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputsLCC}
\alias{prepInputsLCC}
\title{Simple \code{prepInputs} for Canadian LCC data}
\usage{
prepInputsLCC(
  year = 2010,
  destinationPath = asPath("."),
  studyArea = NULL,
  rasterToMatch = NULL,
  filename2 = NULL,
  ...
)
}
\arguments{
\item{year}{Numeric, either 2010 or 2015. See note re: backwards compatibility for 2005.}

\item{destinationPath}{Character string of a directory in which to download
and save the file that comes from \code{url} and is also where the function
will look for \code{archive} or \code{targetFile}. NOTE (still experimental):
To prevent repeated downloads in different locations, the user can also set
\code{options("reproducible.inputPaths")} to one or more local file paths to
search for the file before attempting to download. Default for that option is
\code{NULL} meaning do not search locally.}

\item{studyArea}{\verb{SpatialPolygons*} object used for masking and possibly cropping
if no \code{rasterToMatch} is provided.
If not in same CRS, then it will be \code{spTransform}ed to
CRS of \code{x} before masking. Currently, this function will not reproject the
\code{x}. Optional in \code{postProcess}.}

\item{rasterToMatch}{Template \verb{Raster*} object passed to the \code{to} argument of
\code{\link[reproducible:projectTo]{projectTo()}}, thus will changing the
resolution and projection of \code{x}.
See details in \code{\link[reproducible:postProcessTo]{postProcessTo()}}.}

\item{filename2}{File name passed to \code{\link[reproducible:writeTo]{writeTo()}}.}

\item{...}{Passed into \code{\link[reproducible:writeTo]{writeTo()}}}
}
\description{
A wrapper around \code{prepInputs} for the Canadian Land Cover Classification product(s).
}
\note{
As of May 2021, NRCAN no longer provides/hosts the LCC2005 data.
A privately hosted version of the data is available to maintain backwards compatibility,
but new users/projects should use the 2010 (or newer) data.
}
