% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputsLCC}
\alias{prepInputsLCC}
\title{Simple prepInputs for LCC2005 or LCC2010}
\usage{
prepInputsLCC(year = 2005, destinationPath = asPath("."),
  studyArea = NULL, rasterToMatch = NULL, filename2 = NULL, ...)
}
\arguments{
\item{year}{Numeric, either 2005 or 2010}
}
\description{
A wrapper around prepInputs for the Canadian Land Cover Classification product(s)
}
