% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{prepInputsLCC}
\alias{prepInputsLCC}
\title{Simple prepInputs for LCC2005 or LCC2010}
\usage{
prepInputsLCC(year = 2005, destinationPath = asPath("."),
  studyArea = NULL, rasterToMatch = NULL, filename2 = NULL, ...)
}
\arguments{
\item{year}{Numeric, either 2005 or 2010 (not yet implemented)}

\item{destinationPath}{Character string of a directory in which to download
and save the file that comes from \code{url} and is also where the function
will look for \code{archive} or \code{targetFile}. NOTE (still experimental):
To prevent repeated downloads in different locations, the user can also set
\code{options("reproducible.inputPaths")} to one or more local file paths to
search for the file before attempting to download. Default for that option is
\code{NULL} meaning do not search locally.}

\item{studyArea}{\code{SpatialPolygons*} object used for masking and possibly cropping
if no \code{rasterToMatch} is provided.
If not in same CRS, then it will be \code{spTransform}ed to
CRS of \code{x} before masking. Currently, this function will not reproject the
\code{x}. Optional in \code{postProcess}.}

\item{rasterToMatch}{Template \code{Raster*} object used for cropping (so extent should be
the extent of desired outcome) and reprojecting (including changing the
resolution and projection).
See details in \code{\link{postProcess}}.}

\item{filename2}{\code{filename2} is optional, and is either
NULL (no writing of outputs to disk), or several options
for writing the object to disk. If
\code{TRUE} (the default), it will give it a file name determined by
\code{.prefix(basename(filename1), prefix)}. If
a character string, it will use this as its file name. See
\code{\link{determineFilename}}.}

\item{...}{Passed to raster::crop}
}
\description{
A wrapper around prepInputs for the Canadian Land Cover Classification product(s)
}
