% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\name{prepInputsSpecies}
\alias{prepInputsSpecies}
\alias{prepInputsMainInput}
\title{Download and prepare a species traits table for use with \code{LandR_BiomassGMOrig} module}
\usage{
prepInputsSpecies(url = NULL, dPath, cacheTags = NULL)

prepInputsMainInput(url = NULL, dPath, cacheTags)
}
\arguments{
\item{url}{If NULL (the default), uses one from the LANDIS-II project:
\url{https://github.com/LANDIS-II-Foundation/Extensions-Succession/master/biomass-succession-archive/trunk/tests/v6.0-2.0/biomass-succession_test.txt"}).}

\item{dPath}{The destination path.}

\item{cacheTags}{User tags to pass to \code{Cache}.}
}
\value{
A \code{data.table} with columns ... TODO
}
\description{
TODO: add detailed description
}
\note{
This one is tailored to Canadian forests (?)
}
