% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsStandAgeMap}
\alias{prepInputsStandAgeMap}
\title{Create \code{standAgeMap}}
\usage{
prepInputsStandAgeMap(
  ...,
  dataSource = "KNN",
  dataYear = 2001,
  ageURL = NULL,
  ageFun = "terra::rast",
  maskWithRTM = TRUE,
  method = "bilinear",
  datatype = "INT2U",
  destinationPath = NULL,
  writeTo = NULL,
  firePerimeters = NULL,
  fireURL = paste0("https://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/",
    "fire_poly/current_version/NFDB_poly.zip"),
  fireFun = "terra::vect",
  fireField = "YEAR",
  rasterToMatch = NULL,
  startTime
)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}}

\item{dataSource}{Character. One of KNN, NTEMS, or SCANFI.
Defaults to KNN for \code{dataYear} 2001.
Also available:
\itemize{
\item KNN for \code{dataYear} 2011;
\item SCANFI for \code{dataYear} 2020.
}}

\item{dataYear}{Numeric. Year for which data is obtained. Can be 2001 or 2011 for KNN or 2020 for SCANFI.}

\item{ageURL}{url where age map is downloaded}

\item{ageFun}{passed to 'fun' arg of \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of stand age map}

\item{maskWithRTM}{passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of stand age map}

\item{method}{passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of stand age map}

\item{datatype}{passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of stand age map}

\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{writeTo}{passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of stand age map}

\item{firePerimeters}{fire raster layer fire year values.}

\item{fireURL}{url to download fire polygons used to update age map. If NULL or NA age
imputation is bypassed. Requires passing \code{rasterToMatch}. Only used if \code{firePerimeters}
is missing.}

\item{fireFun}{passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} of fire data. Only used if \code{firePerimeters}
is missing.}

\item{fireField}{field used to rasterize fire polys. Only used if \code{firePerimeters}
is missing.}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{startTime}{date of the last fire year to be considered (e.g., start of fire
period counting backwards). If missing, last fire year available will be used.}
}
\value{
a raster layer stand age map corrected for fires, with an attribute vector of pixel IDs
for which ages were corrected. If no corrections were applied the attribute vector is \code{integer(0)}.
}
\description{
Create the \code{standAgeMap} raster containing age estimates for \code{pixelCohortData}.
A separate \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} call will source Canadian National Fire Data Base
data to update ages of recently burned pixels. To suppress this, pass NULL/NA \code{fireURL}
}
\examples{
\dontrun{
library(SpaDES.tools)
library(terra)
library(reproducible)
randomPoly <- vect(randomStudyArea(size = 1e7))
randomPoly
ras2match <- rast(res = 250, ext = ext(randomPoly), crs = crs(randomPoly))
ras2match <- rasterize(randomPoly, ras2match)
tempDir <- file.path(tempdir(), "ex_prepInputsStandAgeMap")

## NOT USING FIRE PERIMETERS TO CORRECT STAND AGE
## rasterToMatch does not need to be provided, but can be for masking/cropping.
standAge <- prepInputsStandAgeMap(
  destinationPath = tempDir,
  rasterToMatch = ras2match,
  fireURL = NA
) ## or NULL
attr(standAge, "imputedPixID")

## USING FIRE PERIMETERS TO CORRECT STAND AGE
## ideally, get the firePerimenters layer first
firePerimeters <- Cache(prepInputsFireYear,
  url = paste0(
    "https://cwfis.cfs.nrcan.gc.ca/downloads",
    "/nfdb/fire_poly/current_version/NFDB_poly.zip"
  ),
  destinationPath = tempDir,
  rasterToMatch = ras2match
)

standAge <- prepInputsStandAgeMap(
  destinationPath = tempDir,
  firePerimeters = firePerimeters,
  rasterToMatch = ras2match
)
attr(standAge, "imputedPixID")

## not providing firePerimeters is still possible, but will be deprecated
## in this case 'rasterToMatch' MUST be provided
standAge <- prepInputsStandAgeMap(
  destinationPath = tempDir,
  rasterToMatch = ras2match
)
attr(standAge, "imputedPixID")
}
}
