% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepInputsStandAgeMap}
\alias{prepInputsStandAgeMap}
\title{Create \code{standAgeMap}}
\usage{
prepInputsStandAgeMap(
  ...,
  ageURL = NULL,
  ageFun = "raster::raster",
  maskWithRTM = TRUE,
  method = "bilinear",
  datatype = "INT2U",
  destinationPath = NULL,
  filename2 = NULL,
 
    fireURL = "https://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/NFDB_poly.zip",
  fireFun = "sf::st_read",
  rasterToMatch = NULL,
  fireField = "YEAR",
  startTime
)
}
\arguments{
\item{...}{additional arguments passed to \code{prepInputs}}

\item{ageURL}{url where age map is downloaded}

\item{ageFun}{passed to 'fun' arg of \code{prepInputs} of stand age map}

\item{maskWithRTM}{passed to \code{prepInputs} of stand age map}

\item{method}{passed to \code{prepInputs} of stand age map}

\item{datatype}{passed to \code{prepInputs} of stand age map}

\item{destinationPath}{directory where  age and fire data will be downloaded}

\item{filename2}{passed to \code{prepInputs} of stand age map}

\item{fireURL}{url to download fire polygons used to update age map. If NULL or NA age imputation is bypassed.}

\item{fireFun}{passed to \code{prepInputs} of fire data}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{fireField}{field used to rasterize fire polys}

\item{startTime}{date of first fire year.}
}
\value{
a stand age map corrected for fires
}
\description{
Create the \code{standAgeMap} raster containing age estimates for \code{pixelCohortData}.
A separate \code{prepInputs} call will source NDFB data used to update ages of recently burned
pixels.
}
