% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepRawBiomassMap}
\alias{prepRawBiomassMap}
\title{Create \code{rawBiomassMap}}
\usage{
prepRawBiomassMap(
  dataSource = "SCANFI",
  dataYear = "2020",
  dataVersion = "V2",
  ...
)
}
\arguments{
\item{dataSource}{Character. One of KNN, NTEMS, or SCANFI.
Defaults to KNN for \code{dataYear} 2001.
Also available:
\itemize{
\item KNN for \code{dataYear} 2011;
\item NTEMS for \code{dataYear} 2015;
\item SCANFI for \code{dataYear} 2000, 2010, or 2020.
}}

\item{dataYear}{Numeric. Year for which data is obtained. Can be 2001 or 2011 for KNN, 2015 for NTEMS,
2000, 2010, or 2020 for SCANFI V1, or 1985, 1990, 1995, 2000, 2005, 2010, 2015, 2020 (default), 2025 possible for V2.}

\item{dataVersion}{Character. SCANFI product version for data. Default is currently V2. V1 also available.}

\item{...}{arguments passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} and \code{\link[reproducible:Cache]{reproducible::Cache()}}.
If the following arguments are not provided, the following values will be used:
\itemize{
\item{\code{url}: by default, the 2020 SCANFI stand biomass map is downloaded from
a private google drive }
\item{\code{useSAcrs} and \code{projectTo}: \code{FALSE} and \code{NA}}
\item{\code{method}: \code{"bilinear"}}
\item{\code{datatype}: \code{"INT2U"}}
\item{\code{overwrite}: \code{TRUE}}
\item{\code{omitArgs}: \code{c("destinationPath", "targetFile", "userTags", "stable")}}
}}
}
\value{
a \code{rawBiomassMap} raster
}
\description{
Create the \code{rawBiomassMap} raster containing biomass estimates for \code{pixelCohortData}.
}
