% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepRawBiomassMap}
\alias{prepRawBiomassMap}
\title{Create \code{rawBiomassMap}}
\usage{
prepRawBiomassMap(dataSource = "KNN", dataYear = "2011", ...)
}
\arguments{
\item{dataSource}{Character. One of KNN, NTEMS, or SCANFI.
Defaults to KNN for \code{dataYear} 2001.
Also available:
\itemize{
\item KNN for \code{dataYear} 2011;
\item NTEMS for \code{dataYear} 2015;
\item SCANFI for \code{dataYear} 2000, 2010, or 2020.
}}

\item{dataYear}{Numeric. Year for which data is obtained. Can be 2001 or 2011 for KNN, 2015 for NTEMS, or 2000, 2010, or 2020 for SCANFI.}

\item{...}{arguments passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}} and \code{\link[reproducible:Cache]{reproducible::Cache()}}.
If the following arguments are not provided, the following values will be used:
\itemize{
\item{\code{url}: by default, the 2001 kNN stand biomass map is downloaded from
the NRCan National Forest Inventory}
\item{\code{useSAcrs} and \code{projectTo}: \code{FALSE} and \code{NA}}
\item{\code{method}: \code{"bilinear"}}
\item{\code{datatype}: \code{"INT2U"}}
\item{\code{overwrite}: \code{TRUE}}
\item{\code{omitArgs}: \code{c("destinationPath", "targetFile", "userTags", "stable")}}
}}
}
\value{
a \code{rawBiomassMap} raster
}
\description{
Create the \code{rawBiomassMap} raster containing biomass estimates for \code{pixelCohortData}.
}
