% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepRawBiomassMap}
\alias{prepRawBiomassMap}
\title{Create \code{rawBiomassMap}}
\usage{
prepRawBiomassMap(studyAreaName, cacheTags, ...)
}
\arguments{
\item{studyAreaName}{character string giving the study area name.}

\item{cacheTags}{a character vector of strings to pass to \code{Cache(userTags)}}

\item{...}{arguments passed to \code{prepInputs} and \code{Cache}. If the following arguments
are not provided, the following values will be used:
\itemize{
\item{url: by default, the 2001 kNN stand biomass map is downloaded from
the NRCan National Forest Inventory}
\item{useSAcrs: FALSE}
\item{method: "bilinear"}
\item{datatype: "INT2U"}
\item{filename2: \code{suffix("rawBiomassMap.tif", paste0("_", studyAreaName))}}
\item{overwrite: TRUE}
\item{userTags: \code{c(cacheTags, "rawBiomassMap")}}
\item{omitArgs: \code{c("destinationPath", "targetFile", "userTags", "stable")}}
}}
}
\value{
a rawBiomassMap raster
}
\description{
Create the \code{rawBiomassMap} raster containing biomass estimates for
\code{pixelCohortData}.
Wrapper on \code{prepInputs} that will rasterize fire polygons.
}
