% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{prepRawBiomassMap}
\alias{prepRawBiomassMap}
\title{Create \code{rawBiomassMap}}
\usage{
prepRawBiomassMap(studyAreaName, cacheTags, ...)
}
\arguments{
\item{studyAreaName}{character string giving the study area name.}

\item{cacheTags}{a character vector of strings to pass to \code{Cache(userTags)}}

\item{...}{arguments passed to \code{prepInputs} and \code{Cache}. If the following arguments
are not provided, the following values will be used:
\itemize{
\item{\code{url}: by default, the 2001 kNN stand biomass map is downloaded from
the NRCan National Forest Inventory}
\item{\code{useSAcrs}: \code{FALSE}}
\item{\code{method}: \code{"bilinear"}}
\item{\code{datatype}: \code{"INT2U"}}
\item{\code{filename2}: \code{suffix("rawBiomassMap.tif", paste0("_", studyAreaName))}}
\item{\code{overwrite}: \code{TRUE}}
\item{\code{userTags}: \code{c(cacheTags, "rawBiomassMap")}}
\item{\code{omitArgs}: \code{c("destinationPath", "targetFile", "userTags", "stable")}}
}}
}
\value{
a \code{rawBiomassMap} raster
}
\description{
Create the \code{rawBiomassMap} raster containing biomass estimates for
\code{pixelCohortData}.
Wrapper on \code{prepInputs} that will rasterize fire polygons.
}
