% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{prepSpeciesLayers_CASFRI}
\alias{prepSpeciesLayers_CASFRI}
\alias{prepSpeciesLayers_KNN}
\alias{prepSpeciesLayers_Pickell}
\alias{prepSpeciesLayers_ForestInventory}
\title{Prepare species layers}
\usage{
prepSpeciesLayers_CASFRI(destinationPath, outputPath, url = NULL,
  studyArea, rasterToMatch, sppEquiv, sppEquivCol)

prepSpeciesLayers_KNN(destinationPath, outputPath, url = NULL, studyArea,
  rasterToMatch, sppEquiv, sppEquivCol)

prepSpeciesLayers_Pickell(destinationPath, outputPath, url = NULL,
  studyArea, rasterToMatch, sppEquiv, sppEquivCol)

prepSpeciesLayers_ForestInventory(destinationPath, outputPath,
  url = NULL, studyArea, rasterToMatch, sppEquiv, sppEquivCol)
}
\arguments{
\item{destinationPath}{TODO: description needed}

\item{outputPath}{TODO: description needed}

\item{url}{TODO: description needed; if \code{NULL}, the default, use the default source url}

\item{studyArea}{TODO: description needed}

\item{rasterToMatch}{TODO: description needed}

\item{sppEquiv}{table with species name equivalencies between the
kNN format and the final naming format.
See \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}
}
\value{
TODO: description needed
}
\description{
TODO: description needed
}
