% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSpeciesLayers.R
\name{prepSpeciesLayers_KNN}
\alias{prepSpeciesLayers_KNN}
\alias{prepSpeciesLayers_CASFRI}
\alias{prepSpeciesLayers_Pickell}
\alias{prepSpeciesLayers_ForestInventory}
\title{Prepare species layers}
\usage{
prepSpeciesLayers_KNN(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  thresh = 10,
  ...
)

prepSpeciesLayers_CASFRI(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  ...
)

prepSpeciesLayers_Pickell(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  ...
)

prepSpeciesLayers_ForestInventory(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  ...
)
}
\arguments{
\item{destinationPath}{TODO: description needed}

\item{outputPath}{TODO: description needed}

\item{url}{TODO: description needed; if \code{NULL}, the default, use the default source url}

\item{studyArea}{TODO: description needed}

\item{rasterToMatch}{A \code{RasterLayer} object to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
If not supplied, it will attempt to use "Boreal", if it exists,
otherwise it will use the first column in \code{sppEquiv}}

\item{thresh}{threshold \% cover used to defined the species as "present" in the study area.
If at least one pixel has \code{cover >= thresh} , the species is considered "present".
Otherwise the raster is excluded from the output. Defaults to 10.}

\item{...}{other arguments, used for compatibility with other \code{prepSpeciesLayers} functions.}
}
\value{
TODO: description needed
}
\description{
TODO: description needed
}
