% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeCanadianForestRasters.R
\name{calc_raster_counts}
\alias{calc_raster_counts}
\alias{calc_raster_stats}
\alias{plot_raster_stats}
\title{Raster summary statistics and maps by polygon}
\usage{
calc_raster_counts(
  raster,
  polygons = NULL,
  polygon_id = NULL,
  filter_ids = NULL
)

calc_raster_stats(
  raster,
  polygons = NULL,
  polygon_id = NULL,
  filter_ids = NULL
)

plot_raster_stats(
  raster,
  polygons = NULL,
  polygon_id = NULL,
  filter_ids = NULL,
  counts_df = NULL,
  stats_df = NULL,
  aggregate_factor = 1,
  remove_zeros = FALSE,
  raster_label = NULL,
  inset_canada = TRUE,
  bin_width = 10,
  output_dir = ".",
  csv_file = NULL
)
}
\arguments{
\item{raster}{A \code{SpatRaster} object.}

\item{polygons}{A \code{SpatVector} or \code{sf} object of polygons.}

\item{polygon_id}{Name of the column in \code{polygons} to use as polygon ID.}

\item{filter_ids}{Optional. Polygon IDs to filter.}

\item{counts_df}{optional \code{data.frame} of frequency counts output by \code{calc_raster_counts}.
If not provided, that function will be called internally.}

\item{stats_df}{optional \code{data.frame} of raster statistics output by \code{calc_raster_stats}.
If not provided, that function will be called internally.}

\item{aggregate_factor}{Optional. Integer factor to aggregate raster for plotting. Default is \code{1} (no aggregation).}

\item{remove_zeros}{Logical. Removes zeroes from raster and counts for plotting.
Useful to eliminate non-forested areas from forest rasters.}

\item{raster_label}{Optional. Character label for raster used in plots.
Defaults to raster layer name.}

\item{inset_canada}{Logical. Generates a small inset map of plotted polygons within Canada.}

\item{bin_width}{Numeric. Bin size for histogram plots.}

\item{output_dir}{Directory to save figures and optional csv file.}

\item{csv_file}{Optional. Base filename to save the output CSV file.
If provided, a suffix will be added to this filename to denote the 'counts' and 'stats' tables
(e.g., if \code{csv_file = "ecozones.csv"}, the output files saved to \code{output_dir} will be
\code{ecozones_counts.csv} and \code{ecozones_stats.csv}).}
}
\value{
A \code{data.frame} with columns: \code{ID}, \code{value}, \code{count}.

A \code{data.frame} of computed statistics (via \code{calc_raster_stats()}), with side effects
of saving summary plots (and optionally a \file{.csv}) to disk.
}
\description{
\code{calc_raster_stats} iteratively calculates the frequency of raster values within each polygon.
Supports user-provided polygons or automatic retrieval of ecoregions/ecozones.
}
\details{
\code{plot_raster_stats} calculates and returns statistics (min, max, median, mean,
25th percentile, 75th percentile, and proportion of zeroes) for a raster,
and creates and saves figures for each polygon in a vector layer, including
a histogram of raster values, a map of raster values within the polygon, and
a summary of statistics . Optionally, zero values can be removed from the
analysis, and the raster can be aggregated to a lower resolution to speed up plotting.
}
\note{
Large rasters should be processed on disk, where possible, e.g. by setting
\code{terraOptions(memfrac = 0.0)}.
}
\examples{
if (interactive()) {
  if (requireNamespace("dplyr", quietly = TRUE) &&
      requireNamespace("geodata", quietly = TRUE) &&
      requireNamespace("purrr", quietly = TRUE) &&
      requireNamespace("withr", quietly = TRUE) &&
      requireNamespace("zonal", quietly = TRUE)) {
    tmp_pth <- withr::local_tempdir()

    ## work with raster on disk instead of in memory
    terraOptions(memfrac = 0.0)

    age <- prepInputsStandAgeMap(
      dataSource = "KNN",
      dataYear = 2011,
      destinationPath = tmp_pth
    )

    ecozones <- reproducible::prepInputs(
      url = "https://sis.agr.gc.ca/cansis/nsdb/ecostrat/zone/ecozone_shp.zip",
      destinationPath = tmp_pth
    ) |>
     sf::st_make_valid()

    id_col <- "ZONE_NAME"

    ## WARNING: don't use Cache() below -- excessive RAM usage

    ecozone_counts <- calc_raster_counts(
      raster = age,
      polygons = ecozones,
      polygon_id = id_col
    )

    ecozone_stats <- calc_raster_stats(
     raster = age,
     polygons = ecozones,
      polygon_id = id_col
    )

    gc()

    ## below will call calc_raster_counts and calc_raster_stats internally,
    ## so we shortcut this by passing the previously-calculated tables
    stats_ecozone <- plot_raster_stats(
      raster = age,
      polygons = ecozones,
      polygon_id = id_col,
      filter_ids = NULL,
      counts_df = ecozone_counts,
      stats_df = ecozone_stats,
      aggregate_factor = 1,
      remove_zeros = FALSE,
      raster_label = "age",
      inset_canada = TRUE,
      bin_width = 10,
      output_dir = tmp_pth,
      csv_file = NULL
    )

    gc()

    ## review the plots produced in tmp_pth
    list.files(tmp_pth, pattern = "[.]png", full.names = TRUE)

    identical(ecozone_stats, stats_ecozone) ## TRUE
  }
}

}
