% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{rasterRead}
\alias{rasterRead}
\title{Determine the function with which to read a raster}
\usage{
rasterRead(...)
}
\arguments{
\item{...}{passed to \code{\link[terra:rast]{terra::rast()}} or \code{\link[raster:raster]{raster::raster()}}}
}
\value{
the function
}
\description{
TODO: Move to \code{reproducible}
}
