% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{rasterRead}
\alias{rasterRead}
\title{Wrapper to read a raster using a specific package function}
\usage{
rasterRead(...)
}
\arguments{
\item{...}{passed to the function}
}
\value{
a raster of the type returned by the function
}
\description{
Evaluates the function given by the option \code{reproducible.rasterRead}.
E.g., \code{terra::rast} or \code{raster::raster}.
}
\details{
TODO: Move to \code{reproducible}
}
