% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{asInt}
\alias{asInt}
\alias{isInt}
\alias{reclass}
\title{Helpers for transition to \code{terra}}
\usage{
asInt(ras)

isInt(ras)

reclass(ras, tab)
}
\value{
\code{asInt} returns a \verb{*Raster} with values converted to \code{integer}, if they weren't already.

\code{isInt} returns a logical as per \code{is.integer}.

\code{reclass} returns a \verb{*Raster} with values reclassified as per \code{terra::classify}
and \code{raster::reclassify}.
}
\description{
These all create a single function that can be used for either \code{Raster} or \code{SpatRaster}
objects.
}
