% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-raster-utils.R
\name{.NAvalueFlag}
\alias{.NAvalueFlag}
\alias{.stack}
\alias{.projectExtent}
\alias{asInt}
\alias{isInt}
\alias{reclass}
\title{Set NA values in \code{Raster}/\code{SpatRaster}}
\usage{
.NAvalueFlag(ras, NAval)

.stack(rasList)

.projectExtent(ras, crs)

asInt(ras)

isInt(ras)

reclass(ras, tab)
}
\arguments{
\item{ras}{a \code{Raster}, or \code{SpatRaster} object}

\item{NAval}{the value to use as \code{NA}}

\item{rasList}{a list of \code{Raster} or \code{SpatRaster} objects}

\item{crs}{passed to \code{raster::projectRaster(..., crs = crs)}
and \code{terra::project(..., y = crs)}}

\item{tab}{matrix of values to reclassify. See \code{terra::classify}
and \code{raster::reclassify}.}
}
\value{
a raster with attributed NA values

a stacked raster

the projected extent

\code{asInt} returns a \verb{*Raster} with values converted to \code{integer}, if they weren't already.

\code{isInt} returns a logical as per \code{is.integer}.

\code{reclass} returns a \verb{*Raster} with values reclassified as per \code{terra::classify}
and \code{raster::reclassify}.
}
\description{
These all create a single function that can be used for either \code{Raster} or \code{SpatRaster}
objects.
}
