% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputObjects.R
\name{replaceAgeInFires}
\alias{replaceAgeInFires}
\title{Replace stand age with time since last fire}
\usage{
replaceAgeInFires(standAgeMap, firePerimeters, startTime)
}
\arguments{
\item{standAgeMap}{a raster layer stand age map}

\item{firePerimeters}{the earliest fire date to allow}

\item{startTime}{date of the last fire year to be considered (e.g., start of fire
period counting backwards). If missing, last fire year available will be used.}
}
\value{
a raster layer stand age map corrected for fires, with an attribute vector of pixel IDs
for which ages were corrected. If no corrections were applied the attribute vector is \code{integer(0)}.
}
\description{
Replace stand age with time since last fire
}
\examples{
\dontrun{
randomPoly <- terra::vect(SpaDES.tools::randomStudyArea(size = 1e7))
randomPoly
ras2match <- terra::rast(resolution = 250,
                         ext = terra::ext(randomPoly),
                         crs = terra::crs(randomPoly))
ras2match <- terra::rasterize(randomPoly, ras2match)
tempDir <- file.path(tempdir(), "ex_replaceAgeInFires")

standAge <- reproducible::prepInputsStandAgeMap(
  destinationPath = tempDir,
  rasterToMatch = ras2match,
  fireURL = NA
) ## or NULL
attr(standAge, "imputedPixID")

firePerimeters <- reproducible::Cache(prepInputsFireYear,
  url = paste0(
    "https://cwfis.cfs.nrcan.gc.ca/downloads",
    "/nfdb/fire_poly/current_version/NFDB_poly.zip"
  ),
  destinationPath = tempDir,
  rasterToMatch = ras2match
)
standAge <- replaceAgeInFires(standAge, firePerimeters)
attr(standAge, "imputedPixID")
}
}
