% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{rmMissingCohorts}
\alias{rmMissingCohorts}
\title{Remove missing cohorts from cohortData based on pixelGroupMap}
\usage{
rmMissingCohorts(
  cohortData,
  pixelGroupMap,
  doAssertion = getOption("LandR.assertions", TRUE)
)
}
\arguments{
\item{cohortData}{A \code{data.table} with columns: \code{pixelGroup}, \code{ecoregionGroup},
\code{speciesCode}, and optionally \code{age}, \code{B}, \code{mortality}, \code{aNPPAct},
and \code{sumB}.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that correspondsd exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}
}
\value{
A \code{list} with 2 \code{data.table} objects, \code{cohortData} and \code{pixelGroupMap},
each updated based on missing pixelGroups in the other.
}
\description{
Remove missing cohorts from cohortData based on pixelGroupMap
}
