% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandR.R
\name{rmMissingCohorts}
\alias{rmMissingCohorts}
\title{Remove missing cohorts from cohortData based on pixelGroupMap}
\usage{
rmMissingCohorts(cohortData, pixelGroupMap, firePixelTable)
}
\arguments{
\item{cohortData}{a \code{data.table} with columns:
"pixelGroup", "ecoregionGroup", "speciesCode", "age", "B", "mortality", "aNPPAct", "sumB"}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group number that
correspondsd exactly to ]\code{pixelGroup} column in \code{cohortData}}

\item{firePixelTable}{A data.table with 2 columns, \code{pixelIndex} and \code{pixelGroup}.
This will be used in conjunction
with cohortData and pixelGroupMap to ensure that everything matches correctly.}
}
\value{
A \code{data.table} with a new, cohortData
}
\description{
Remove missing cohorts from cohortData based on pixelGroupMap
}
