% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scheduleDisturbance}
\alias{scheduleDisturbance}
\title{Test whether disturbance should be scheduled}
\usage{
scheduleDisturbance(disturbanceLayer, currentYear)
}
\arguments{
\item{disturbanceLayer}{a rasterLayer object}

\item{currentYear}{time of simulation}
}
\value{
logical vector signifying whether to schedule event
}
\description{
Test whether disturbance should be scheduled
}
\examples{
doEvent <- scheduleDisturbance(sim$rstCurrentBurn, time(sim))
}
