% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R, R/misc.R
\name{speciesEcoregionLatestYear}
\alias{speciesEcoregionLatestYear}
\title{Pull out the values from speciesEcoregion table for current time}
\usage{
speciesEcoregionLatestYear(speciesEcoregion, currentTime)

speciesEcoregionLatestYear(speciesEcoregion, currentTime)
}
\arguments{
\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values}

\item{currentTime}{The current time e.g., \code{time(sim)}}

\item{speciesEcoregion}{A \code{data.table} with \code{speciesEcoregion} values}

\item{currentTime}{The current time e.g., \code{time(sim)}}
}
\description{
Pull out the values from speciesEcoregion table for current time

Pull out the values from speciesEcoregion table for current time
}
\note{
TODO

TODO
}
