% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{speciesEcoregionLatestYear}
\alias{speciesEcoregionLatestYear}
\title{Pull out the values from \code{speciesEcoregion} table for current time}
\usage{
speciesEcoregionLatestYear(speciesEcoregion, currentTime)
}
\arguments{
\item{speciesEcoregion}{A \code{data.table} with \code{species}-\code{ecoregion}-specific species trait values.
Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar biophysical
characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.
May contain columns \code{inflationFactor}  (used to adjust \code{maxB}) and \code{mANPPproportion}
(used to calculate \code{maxANPP}).}

\item{currentTime}{The current simulation time e.g., \code{time(sim)}.}
}
\value{
The \code{speciesEcoregion} input object, but with data from only one year, the year
that is less than or equal to the \code{currentTime}
}
\description{
Pull out the values from \code{speciesEcoregion} table for current time
}
