% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{speciesEcoregionStack}
\alias{speciesEcoregionStack}
\title{Create Stacks of the \code{speciesEcoregion} content}
\usage{
speciesEcoregionStack(
  ecoregionMap,
  speciesEcoregion,
  columns = c("establishprob", "maxB", "maxANPP")
)
}
\arguments{
\item{ecoregionMap}{The \code{ecoregionMap}, a raster of all the unique ecoregion groupings,
which group together pixels of similar biophysical characteristics.}

\item{speciesEcoregion}{A \code{data.table} with \code{species}-{ecoregion}-specific species
trait values. Ecoregion refers to "ecolocation", a categorical variable grouping sites with similar
biophysical characteristics. The table should have at least the following columns: \code{speciesCode} and
\code{ecoregionGroup}, character representation of species and ecoregion groups respectively,
\code{maxB} the maximum biomass for the species in a given 'ecoregion', \code{maxANPP} the maximum
aboveground net primary productivity and \code{SEP} the species establishment probability.}

\item{columns}{The columns to use in the \code{speciesEcoregion} data.table.
Default is \code{c("establishprob", "maxB", "maxANPP")}}
}
\description{
This will output a list of \code{RasterStack} objects. Each \code{RasterStack} show raster maps
of one of the columns listed in \code{columns} and each \code{RasterLayer} will be one species.
}
