% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{speciesEcoregionStack}
\alias{speciesEcoregionStack}
\title{Create Stacks of the speciesEcoregion content}
\usage{
speciesEcoregionStack(
  ecoregionMap,
  speciesEcoregion,
  columns = c("establishprob", "maxB", "maxANPP")
)
}
\arguments{
\item{ecoregionMap}{The Factor RasterLayer of the ecoregionMap}

\item{speciesEcoregion}{The data.table with the speciesEcoregion information}

\item{columns}{The columns to use in the \code{speciesEcoregion} data.table.
Default is \code{c("establishprob", "maxB", "maxANPP")}}
}
\description{
This will output a list of RasterStack objects. Each \code{RasterStack} show
raster maps of one of the columns listed in \code{columns} and each
\code{RasterLayer} will be one species.
}
