% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPresentFromKNN.R
\name{speciesInStudyArea}
\alias{speciesInStudyArea}
\title{Get species list in a given study area for a forest in Canada}
\usage{
speciesInStudyArea(
  studyArea,
  url = NULL,
  speciesPresentRas = NULL,
  sppEquivCol = NULL,
  dPath = getOption("reproducible.destinationPath")
)
}
\arguments{
\item{studyArea}{A \verb{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{url}{A url to get a \code{speciesPresence} raster e.g., from \code{peciesPresentFromKNN}}

\item{speciesPresentRas}{A factor raster where the character string is a string of
species names, separated by 2 underscores, sorted alphabetically. Can be produced
with \code{speciesPresentFromKNN}}

\item{sppEquivCol}{An optional column from \code{LandR::sppEquivalencies_CA}.
If passed the KNN species will be returned according to this naming convention.}

\item{dPath}{Passed to \code{destinationPath} in \code{preProcess}.}
}
\value{
A named list of length 2: \code{speciesRas} is a factor \code{RasterLayer}
and \code{speciesList} is a character string containing the unique, sorted
species on the \code{speciesRas}, for convenience.
}
\description{
\code{speciesInStudyArea} defaults to use a url of a dataset uploaded to Google Drive that is
from Canadian Forests, but a different factor raster can be passed e.g.,
from \code{speciesPresentFromKNN}.
}
