% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPresentFromKNN.R
\name{speciesPresentFromKNN}
\alias{speciesPresentFromKNN}
\alias{speciesInStudyArea}
\title{Make a Species factor raster}
\usage{
speciesPresentFromKNN(
  year = 2011,
  dPath = asPath("."),
  res = 2000,
  minPctCover = 10
)

speciesInStudyArea(studyArea, url = NULL, speciesPresentRas = NULL)
}
\arguments{
\item{year}{Default (and only implemented) is 2011. This will download the 2011 KNN data layers}

\item{dPath}{A character string indicating where to download all the KNN layers}

\item{res}{The resolution (one dimension, in m) for the resulting raster}

\item{minPctCover}{An integer indicating what percent cover a species must have
in a pixel to be considered present in that pixel.}

\item{studyArea}{a vector map (e.g., SpatialPolygonsDataFrame)}

\item{url}{A url to get a speciesPresence raster e.g., from \code{peciesPresentFromKNN}}

\item{speciesPresentRas}{A factor raster where the character string is a string of
species names, separated by 2 underscores, sorted alphabetically. Can be produced
with \code{speciesPresentFromKNN}}
}
\value{
A terra rast object with 2 layers, "speciesPresent" is a factor rast, with
a legend (i.e., it is numbers on a map, that correspond to a legend) and
"numberSpecies" which represents the number of species in each pixel.

A named list of length 2: \code{speciesRas} is a factor \code{RasterLayer}
and \code{speciesList} is a character string containing the unique, sorted
species on the \code{speciesRas}, for convenience.
}
\description{
This will download all KNN layers in (Boreal) Forest of Canada, and make
a factor raster at resolution provided by \code{res} (larger is faster).

\code{speciesInStudyArea} defaults to use a url of a dataset uploaded
to googledrive that is from
Canadian Boreal Forests, but a different factor raster can be passed e.g.,
from \code{speciesPresentFromKNN}.
}
\examples{
# Make the dataset
speciesPresent <- speciesPresentFromKNN("~/data/KNN")

\dontrun{
  # To upload this:
  speciesPresentRas <- raster::stack(speciesPresent)[[1]]
  fn <- "SpeciesPresentInCanadianForests.tif"
  writeRaster(speciesPresentRas, file = fn)
  out <- googledrive::drive_put(fn)
  driveID <- "1GrV5LjXS_N4iMwGYfBjUZ4elT2DkuKyJ"
}

# Get species list
sa <- LandR::randomStudyArea(size = 1e11)
species <- LandR::speciesInStudyArea(sa)

}
