% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPresentFromKNN.R
\name{speciesPresentFromKNN}
\alias{speciesPresentFromKNN}
\title{Make a species factor raster from KNN}
\usage{
speciesPresentFromKNN(
  year = 2011,
  dPath = asPath("."),
  res = 2000,
  minPctCover = 10
)
}
\arguments{
\item{year}{Default (and only implemented) is 2011. This will download the 2011 KNN data layers}

\item{dPath}{A character string indicating where to download all the KNN layers}

\item{res}{The resolution (one dimension, in m) for the resulting raster}

\item{minPctCover}{An integer indicating what percent cover a species must have
in a pixel to be considered present in that pixel.}
}
\value{
A \code{SpatRaster} object with 2 layers: \code{"speciesPresent"} is a factor, with
a legend (i.e., it is numbers on a map, that correspond to a legend) and
\code{"numberSpecies"} which represents the number of species in each pixel.
}
\description{
This will download all KNN layers in Forests of Canada, and make
a factor raster at resolution provided by \code{res} (larger is faster).
}
