% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\alias{prepSpeciesTable}
\title{Download and prepare a species traits table for use with L\code{Biomass_core} module}
\usage{
getSpeciesTable(url = NULL, dPath = tempdir(), cacheTags = NULL)

prepSpeciesTable(
  speciesTable,
  speciesLayers,
  sppEquiv = NULL,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{url}{If NULL (the default), uses one from D. Cyr's LANDIS-II files:
\url{https://github.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv}).}

\item{dPath}{The destination path.}

\item{cacheTags}{User tags to pass to \code{Cache}.}

\item{speciesTable}{A raw species traits table}

\item{speciesLayers}{stack of species layers rasters}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}
}
\value{
A \code{data.table} with columns ... TODO
}
\description{
TODO: add detailed description
}
\note{
This one is tailored to Canadian forests (?)
}
