% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\alias{prepSpeciesTable}
\title{Download and prepare a species traits table for use with \code{Biomass_core} module}
\usage{
getSpeciesTable(url = NULL, dPath = tempdir(), cacheTags = NULL)

prepSpeciesTable(
  speciesTable,
  speciesLayers = NULL,
  sppEquiv = NULL,
  sppEquivCol = "LandR",
  areas = c("BSW", "BP", "MC")
)
}
\arguments{
\item{url}{If NULL (the default), uses one from D. Cyr's LANDIS-II files:
\url{https://github.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv}).}

\item{dPath}{The destination path.}

\item{cacheTags}{User tags to pass to \code{Cache}.}

\item{speciesTable}{The raw species traits table, with \strong{at least} the following columns:
\itemize{
\item \code{speciesCode} an character representation of species;
\item \code{Area} a character of geographical area within a species range for which trait values are relevant;
\item \code{longevity} species longevity in years, \code{sexualmature} age in years at sexual maturity;
\item \code{shadetolerance} numeric value between 1-5 of relative shade tolerance (with respect to other species);
\item \code{seeddistance_eff} the "effective" seed dispersal distance;
\item \code{seeddistance_max} a numeric with the maximum seed dispersal distance;
\item \code{mortalityshape} and \code{growthcurve}: growth curve shape parameters.
Other columns (e.g. fire-related traits) can also be included depending on LandR modules in use.
Please see the LANDIS-II Biomass Succession Extension v3.2.1 manual (Scheller and Miranda 2015)
for further detail.
}}

\item{speciesLayers}{Deprecated.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{areas}{A character vector of areas to use. Can be one or more of
\code{c("Acadian", "AM", "NorthShore", "BP", "BSE", "BSW", "LSJ", "MC", "PM", "WestON")}.
If it is more than one, this function will take the minimum value, within a species.
These are short versions of the Canada Ecoprovinces.
Currently defaults to \code{c("BSW", "BP", "MC")} for historical reasons.}
}
\value{
A \code{data.table} with columns ... TODO
}
\description{
\code{prepSpeciesTable}
}
\note{
This one is tailored to Canadian forests (?)
}
