% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDIS-II-tables.R
\name{speciesTableUpdate}
\alias{speciesTableUpdate}
\title{Change species table of parameters/traits}
\usage{
speciesTableUpdate(species, speciesTable, sppEquiv, sppEquivCol)
}
\arguments{
\item{species}{a \code{data.table} that has species traits such as longevity, shade tolerance, etc.}

\item{speciesTable}{TODO: DESCRIPTION NEEDED}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}
}
\value{
An updated species \code{data.table}
}
\description{
Changes longevity values in the species table according to Burton & Cumming (1995).
Only the species and values present in Boreal Shield West (BSW), Boreal Plains (BP) and
Montane Cordillera (MC) \code{speciesTable$Area} are being changed.
All others follow Dominic Cyr and Yan Boulanger's trait values
(\url{https://raw.githubusercontent.com/dcyr/LANDIS-II_IA_generalUseFiles/master/speciesTraits.csv}).
Note that BSW and BP areas correspond more closely to the region considered in Table 2 of
Burton & Cumming (1995), while MC will correspond to both tables.
}
