% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spiralSeedDispersal}
\alias{spiralSeedDispersal}
\title{Multiplies two doubles}
\usage{
spiralSeedDispersal(
  cellCoords,
  speciesVectorsList,
  rcvSpeciesByIndex,
  speciesTable,
  numCols,
  numCells,
  cellSize,
  xmin,
  ymin,
  k,
  b,
  successionTimestep
)
}
\arguments{
\item{cellCoords}{Matrix, 2 columns, of x-y coordinates of the Receive cells}

\item{overallMaxDist}{The maximum distance to test for seed dispersal}
}
\value{
A list.
}
\description{
Multiplies two doubles
}
