% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spiralSeedDispersal}
\alias{spiralSeedDispersal}
\title{Ward seed dispersal using Rcpp}
\usage{
spiralSeedDispersal(
  cellCoords,
  speciesVectorsList,
  rcvSpeciesByIndex,
  speciesTable,
  numCols,
  numCells,
  cellSize,
  xmin,
  ymin,
  k,
  b,
  successionTimestep,
  verbose = 0
)
}
\arguments{
\item{cellCoords}{Matrix, 2 columns, of x-y coordinates of the Receive cells}

\item{speciesVectorsList}{A list, where each element is a vector of NA and the speciesCode
value of the list. The length of each vector MUST be the number of cells in the
raster whose \code{cellCoords} are provided.}

\item{rcvSpeciesByIndex}{A list of length \code{NROW(cellCoords)} where each element
is the vector of speciesCodes that are capable of being received in the
corresponding \code{cellCoords}}

\item{speciesTable}{A numeric matrix with species traits. Must have column 3 be
\code{seeddistance_max}, column 2 be \code{seeddistance_eff}, and sorted in
increasing order on the first column, speciesCode. The speciesCode values must
be \code{seq(1, NROW(speciesTable))}. The names of these columns is not important,
only the position in the matrix}

\item{numCols}{Integer, number of columns in the raster whose \code{cellCoords}
were provided}

\item{numCells}{Integer, number of cells in the raster whose \code{cellCoords}
were provided}

\item{cellSize}{Integer, the \code{res(ras)[1]} of the raster whose \code{cellCoords}
were provided}

\item{xmin}{Integer, the \code{xmin(ras)} of the raster whose \code{cellCoords}
were provided}

\item{ymin}{Integer, the \code{ymin(ras)} of the raster whose \code{cellCoords}
were provided}

\item{k}{Numeric, parameter passed to Ward dispersal kernel}

\item{b}{Numeric, parameter passed to Ward dispersal kernel}

\item{successionTimestep}{Integer, Same as Biomass_core.}

\item{verbose}{Numeric, length 1. Currently \code{0} (no messaging),
\code{1} (some messaging) and \code{2} or greater (more messaging) are active. Default is
\code{getOption("LandR.verbose", TRUE)}.}
}
\value{
A logical matrix with ncols = \code{length(speciesVectorsList)} and nrows =
  \code{NROW(cellCoords)}, indicating whether that cellCoords successfully
  received seeds from each species.
}
\description{
This uses a spiral pattern outwards from the \code{cellCoords} cells on
a raster with the dimensions \code{numCols}, \code{numCells}, \code{xmin},
\code{ymin}, and \code{cellSize}. For each cell in \code{cellCoords},
it evaluates whether there is a successful "dispersal" \code{to} that cell
for the species that can disperse there as identified by \code{rcvSpeciesByIndex}.
It will search outwards testing each and every cell in the spiral until
the maximum distance is reached as specified in the 3rd column (named or unnamed)
of \code{speciesTable}.
}
