% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sppColors}
\alias{sppColors}
\title{Create species color vector from a sppEquiv table}
\usage{
sppColors(sppEquiv, sppEquivCol, newVals = NULL, palette)
}
\arguments{
\item{sppEquiv}{table with species name equivalencies between the
kNN format and the final naming format.
See \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.}

\item{newVals}{An optional character vector of extra names to use, e.g., "Mixed".}

\item{palette}{An RColorBrewer palette, e.g., "Accent".
Can get RColorBrewer palette names from
\code{rownames(RColorBrewer::brewer.pal.info)}.}
}
\value{
A named vector of color codes, where the names are the species names
plus any extra names passed with \code{newVals}.
}
\description{
Create species color vector from a sppEquiv table
}
