% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sppColors}
\alias{sppColors}
\title{Create species color vector from a sppEquiv table}
\usage{
sppColors(sppEquiv, sppEquivCol, newVals, palette)
}
\arguments{
\item{sppEquiv}{A species equivalency table, e.g., \code{data("sppEquivalencies_CA")}.}

\item{sppEquivCol}{The name of the column to get names from.}

\item{newVals}{An optional character vector of extra names to use, e.g., "Mixed".}

\item{palette}{An RColorBrewer palette, e.g., "Accent".
Can get RColorBrewer palette names from
\code{rownames(RColorBrewer::brewer.pal.info)}.}
}
\value{
A named vector of color codes, where the names are the species names
plus any extra names passed with \code{newVals}.
}
\description{
Create species color vector from a sppEquiv table
}
