% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseName.R
\name{sppEquivCheck}
\alias{sppEquivCheck}
\title{Check and expand sppEquiv}
\usage{
sppEquivCheck(sppEquiv, ensureColumns = NULL)
}
\arguments{
\item{sppEquiv}{A character vector or data.table with named column(s). If this
data.table does not have columns named \code{ensureColumns}, then it will
attempt to merge this data.table with LandR::sppEquivalencies_CA to get
\code{ensureColumns}.}

\item{ensureColumns}{A character vector of column names that must be in \code{sppEquiv}.
If these are not present, then the function will attempt to merge with
LandR::sppEquivalencies_CA, so the column name(s) of \code{sppEquiv} must match
column names in LandR::sppEquivalencies_CA.}
}
\value{
A data.table with potentially all columns in LandR::sppEquivalencies_CA.
}
\description{
This will expand a \code{sppEquiv} object that is only a vector or only a one-column
data.table into a many column data.table, if the columns that are present do not
contain \code{ensureColumns}.
}
