% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppHarmonize.R
\name{sppHarmonize}
\alias{sppHarmonize}
\title{Harmonize the 3 components that bring species into Biomass_**}
\usage{
sppHarmonize(sppEquiv, sppNameVector, sppEquivCol)
}
\arguments{
\item{sppEquiv}{A data.table or NULL representing the sppEquiv table (see manual).
If NULL, then it will use the `LandR::sppEquivalencies_CA`}

\item{sppNameVector}{A character vector of species to use. These species must all
be from one naming convention, i.e., from one column in the sppEquiv.}

\item{sppEquivCol}{A character string normally provided from the P(sim)$sppEquivCol
(see manual). If `NA`, the default, then this will try to determine which
column the `sppNameVector` used and use that. If `sppNameVector` is NULL, then
it will default to `"Boreal"`.}
}
\value{
Returns a named list with the same names as the arguments. These should
  likely be assigned to the `sim` object in the module following this function call.
}
\description{
This function will attempt to harmonize many potential issues/conflicts that
may arise under different combinations of supplied objects to the three
arguments. See manual for details.
}
\examples{
sppOuts <- sppHarmonize(sim$sppEquiv, sim$sppNameVector, P(sim)$sppEquivCol)
sim$sppEquiv <- sppOuts$sppEquiv
sim$sppNameVector <- sppOuts$sppNameVector
P(sim)$sppEquivCol <- sppOuts$sppEquivCol

}
