% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppHarmonize.R
\name{sppHarmonize}
\alias{sppHarmonize}
\title{Harmonize the 3 components that bring species into Biomass_**}
\usage{
sppHarmonize(
  sppEquiv,
  sppNameVector,
  sppEquivCol,
  sppColorVect,
  vegLeadingProportion = 0,
  studyArea
)
}
\arguments{
\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppNameVector}{A character vector of species to use. These species must all
be from one naming convention, i.e., from one column in the \code{sppEquiv}.}

\item{sppEquivCol}{A character string normally provided from the \code{P(sim)$sppEquivCol}
(see manual). If \code{NA}, the default, then this will try to determine which
column the \code{sppNameVector} used and use that. If \code{sppNameVector} is NULL, then
it will default to \code{"Boreal"}.}

\item{sppColorVect}{A named vector of colours to use for plotting.
The names must conform with species name convention used
(see LandR::sppEquivalencies_CA for possible naming
conventions) and should also contain a colour for 'Mixed',
when a mixed forest type is supported  (see vegLeadingProportion
parameter in LandR::vegTypeGenerator for details on mixed
fores types).}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{studyArea}{A \code{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible]{prepInputs}}.}
}
\value{
Returns a named list with the same names as the arguments. These should
likely be assigned to the \code{sim} object in the module following this function call.
}
\description{
This function will attempt to harmonize many potential issues/conflicts that
may arise under different combinations of supplied objects to the three
arguments. See manual for details.
}
\examples{
## not run. usage example within module
# sppOuts <- sppHarmonize(sim$sppEquiv, sim$sppNameVector, P(sim)$sppEquivCol)
# sim$sppEquiv <- sppOuts$sppEquiv
# sim$sppNameVector <- sppOuts$sppNameVector
# P(sim)$sppEquivCol <- sppOuts$sppEquivCol

}
