% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{statsModel}
\alias{statsModel}
\title{The generic statistical model -- to run lmer or glmer}
\usage{
statsModel(form, .specialData, ...)
}
\arguments{
\item{form}{A quoted formula to test}

\item{.specialData}{The custom dataset required for the model}

\item{...}{Anything passed to args for the model}
}
\description{
This does a few things including R squared, gets the fitted values.
It appears that running the models "asis" without this wrapper
does not work with \code{Cache}. The return of the model in
a list solves this problem.
}
