% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\docType{data}
\name{uniqueCohortDefinition}
\alias{uniqueCohortDefinition}
\alias{uniqueSpeciesEcoregionDefinition}
\title{The columns in a \code{cohortData} that define "unique"}
\format{An object of class \code{character} of length 4.}
\usage{
uniqueCohortDefinition

uniqueSpeciesEcoregionDefinition
}
\description{
If two pixels have identical values in all of these columns, they are the same \code{pixelGroup}.
}
\keyword{datasets}
