% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{updateCohortData}
\alias{updateCohortData}
\alias{.initiateNewCohorts}
\title{Add cohorts to \code{cohortData} and \code{pixelGroupMap}}
\usage{
updateCohortData(newCohortData, cohortData, pixelGroupMap, time,
  speciesEcoregion, firePixelTable = NULL, successionTimestep)

.initiateNewCohorts(newCohortData, cohortData, pixelGroupMap, time,
  speciesEcoregion)
}
\arguments{
\item{newCohortData}{must be a complete cohortData object with newly
created cohorts. They do not have to have \code{pixelGroup} values yet; they
can be overlapping with \code{cohortData}, (i.e., they can be regenerated on
empty pixels or on alreayd occupied pixels). Columns it must have:
\code{pixelGroup}, \code{speciesCode}, \code{age}, \code{ecoregionGroup}. The
remaining 4 (see \code{cohortData}) will be created with \code{0}s}

\item{cohortData}{a \code{data.table} with columns: \code{pixelGroup},
\code{ecoregionGroup}, \code{speciesCode}, \code{age}, \code{B},
\code{mortality}, \code{aNPPAct}, and \code{sumB}.}

\item{pixelGroupMap}{Raster layer with pixel values equal to a pixel group
number that correspondsd exactly to \code{pixelGroup} column in
\code{cohortData}}

\item{time}{Current time e.g., time(sim). This is used to extract the correct
parameters in \code{speciesEcoregion} table if there are different values
over time}

\item{speciesEcoregion}{A \code{speciesEcoregion} table.}

\item{firePixelTable}{A data.table with 2 columns, \code{pixelIndex} and
\code{pixelGroup}}

\item{successionTimestep}{The time between successive seed dispersal events.
In LANDIS-II, this is called "Succession Timestep". This is used here}
}
\value{
A list of length 2, \code{cohortData} and \code{pixelGroupMap}, with
\code{newCohortData} inserted.

\code{.initiateNewCohorts} returns A \code{data.table} with a new,
\code{rbindlist}ed cohortData
}
\description{
This is a wrapper for  \code{generatePixelGroups}, \code{initiateNewCohort} and
updates to \code{pixelGroupMap} via assignment to new \code{pixelIndex}
values in \code{newCohortData}. By running these all together,
there is less chance that they will diverge. There are some checks
internally for consistency.
}
\details{
Does the following:
\enumerate{
  \item add new cohort data into \code{cohortdata};
  \item assign initial biomass and age for new cohort;
  \item assign the new pixelgroup to the pixels that have new cohort;
  \item update the pixelgroup map.
}
}
