% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{vtm2conifdecid}
\alias{vtm2conifdecid}
\alias{vegTransitions}
\alias{plotVegTransitions}
\title{Create a summaries of vegetation type transitions}
\usage{
vtm2conifdecid(vtm, sppEquiv = NULL, sppEquivCol = "LandR", zones)

vegTransitions(vtm, zones, field, times, na.rm = FALSE, dest = ".")

plotVegTransitions(transitions_df)
}
\arguments{
\item{vtm}{character vector of file paths to vegetation type maps (see \code{\link[=vegTypeMapGenerator]{vegTypeMapGenerator()}}.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{zones}{\code{sf} polygons object delineating the reporting area (\code{zone}) boundaries.}

\item{field}{character string of the column name in \code{zones} to use for grouping.}

\item{times}{numeric vector of years corresponding to \code{vtm}.}

\item{na.rm}{logical. If \code{TRUE}, remove rows with \code{NA} values in \code{vegType}
(they won't appear as a stratum in the alluvial diagram).
If \code{FALSE}, these \code{NA} values will be replaced with \code{"_NA_"} so transitions
between vegetated and non-vegetated pixels can be visualized.}

\item{dest}{character, specifying a destination directory}

\item{transitions_df}{A data frame with columns \code{pixelID}, \code{zone}, \code{vegType}, and \code{time}.
(i.e., output of \code{vegTransitions()}).}
}
\value{
\itemize{
\item \code{vtm2conifdecid()} returns a character vector of file paths to the conifer-deciduous maps.
}

\itemize{
\item \code{vegTransitions()} returns a \code{arrow} dataset with columns \code{pixelID}, \code{zone}, \code{vegType},
and \code{time}.
}

\itemize{
\item \code{plotVegTransitions()} returns a list of \code{ggplot} objects, one for each zone.
}
}
\description{
Create a summaries of vegetation type transitions

Plot vegetation type transitions
}
\note{
creating these plots for large landscapes can be computationally intensive
(time and memory use).
}
