% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transitions.R
\name{vegTransitions}
\alias{vegTransitions}
\alias{plotVegTransitions}
\title{Create a vegetation transitions data frame}
\usage{
vegTransitions(fvtm, ecoregion, studyArea, times)

plotVegTransitions(transitions_df)
}
\arguments{
\item{fvtm}{character vector of filepaths to vegetation type maps.}

\item{ecoregion}{\code{SpatRaster} of ecoregion (or other) codes by which to group produce plots.}

\item{studyArea}{\code{sf} polygons object delineating the area to use for cropping and masking
of \code{ecoregion} (e.g., \code{studyAreaReporting}).}

\item{times}{numeric vector of years corresponding to the \code{fvtm} files.}

\item{transition_df}{A data frame with columns \code{pixelID}, \code{ecoregion}, \code{vegType}, and \code{time}.
(i.e., output of \code{vegTransitions()}).}
}
\value{
\code{data.frame} with columns \code{pixelID}, \code{ecoregion}, \code{vegType}, and \code{time}.

A list of ggplot objects, one for each ecoregion.
}
\description{
Create a vegetation transitions data frame

Plot vegetation type transitions
}
\note{
creating these plots for large landscapes can be computationally intensive
(time and memory use).
}
