% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{vegTypeGenerator}
\alias{vegTypeGenerator}
\title{Generate and add vegetation type column to \code{cohortData}}
\usage{
vegTypeGenerator(
  x,
  vegLeadingProportion = 0.8,
  mixedType = 2,
  sppEquiv = NULL,
  sppEquivCol,
  pixelGroupColName = "pixelGroup",
  doAssertion = getOption("LandR.assertions", TRUE),
  ...
)
}
\arguments{
\item{x}{A \code{cohortData} object}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{mixedType}{An integer defining whether mixed stands are of any kind of species
admixture (1), or only when deciduous mixed with conifer (2).
Defaults to 2.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{pixelGroupColName}{Name of the column in \code{pixelGroup} to use.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{...}{Additional arguments.}
}
\value{
\code{x} with a new column, 'leading', coding the vegetation type
   of each group defined by \code{pixelGroupColName}
}
\description{
This function is a simplification of \code{vegTypeMapGenerator}
and instead of generating a map, it adds the vegetation type column
to the \code{cohortData} table.
}
\examples{
library(data.table)
x <- data.table(pixelGroup = rep(1:2, each = 2), B = c(100, 200, 20, 400),
                speciesCode = rep(c("Pice_Gla", "Popu_Tre"), 2))
vegTypeGenerator(x)
}
\author{
Eliot McIntire, Ceres Barros, Alex Chubaty
}
