% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{vegTypeMapGenerator}
\alias{vegTypeMapGenerator}
\alias{vegTypeMapGenerator.RasterStack}
\alias{vegTypeMapGenerator.data.table}
\title{Generate vegetation type map}
\usage{
vegTypeMapGenerator(x, ...)

\method{vegTypeMapGenerator}{RasterStack}(x, ..., doAssertion = getOption("LandR.doAssertion", TRUE))

\method{vegTypeMapGenerator}{data.table}(
  x,
  pixelGroupMap,
  vegLeadingProportion = 0.8,
  mixedType = 2,
  sppEquiv = NULL,
  sppEquivCol,
  colors,
  pixelGroupColName = "pixelGroup",
  doAssertion = getOption("LandR.assertions", TRUE),
  ...
)
}
\arguments{
\item{x}{Either a \code{cohortData} object or a \code{speciesCover} \code{RasterStack}}

\item{...}{Additional arguments.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{pixelGroupMap}{A \code{raster}}

\item{vegLeadingProportion}{Numeric between 0-1, determing the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{mixedType}{An integer defining whether mixed stands are of any kind of species
admixture (1), or only when deciduous mixed with conifer (2).
Defaults to 2.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{colors}{A named vector of color codes. The names MUST match the names of species
in \code{cohortData$speciesCode}, plus an optional "Mixed" color.}

\item{pixelGroupColName}{Name of the column in \code{pixelGroup} to use}
}
\description{
Generate vegetation type map
}
\author{
Eliot McIntire, Ceres Barros, Alex Chubaty
}
