% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{vegTypeMapGenerator}
\alias{vegTypeMapGenerator}
\alias{vegTypeMapGenerator.RasterStack}
\alias{vegTypeMapGenerator.data.table}
\title{Generate vegetation type map}
\usage{
vegTypeMapGenerator(x, ...)

\method{vegTypeMapGenerator}{RasterStack}(x, ..., doAssertion = getOption("LandR.doAssertion", TRUE))

\method{vegTypeMapGenerator}{data.table}(
  x,
  pixelGroupMap,
  vegLeadingProportion = 0.8,
  mixedType = 2,
  sppEquiv = NULL,
  sppEquivCol,
  colors,
  pixelGroupColName = "pixelGroup",
  doAssertion = getOption("LandR.assertions", TRUE),
  ...
)
}
\arguments{
\item{x}{Either a \code{cohortData} object or a \code{speciesCover} \code{RasterStack}}

\item{...}{Additional arguments.}

\item{doAssertion}{A logical indicating whether some internal tests should be run to
ensure the function is running correctly.
Default: \code{getOption("LandR.assertions", TRUE)}.}

\item{pixelGroupMap}{A \code{RasterLayer} with pixel values equal to a pixel group
number that corresponds exactly to \code{pixelGroup} column in \code{cohortData}.}

\item{vegLeadingProportion}{Numeric between 0-1, determining the relative biomass
threshold a species needs to pass to be considered "leading".}

\item{mixedType}{An integer defining whether mixed stands are of any kind of species
admixture (1), or only when deciduous mixed with conifer (2).
Defaults to 2.}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{colors}{A named vector of colour codes. The names MUST match the names of species
in \code{cohortData$speciesCode}, plus an optional "Mixed" colour.}

\item{pixelGroupColName}{Name of the column in \code{pixelGroup} to use.}
}
\description{
Generate vegetation type map
}
\examples{
library(data.table)
library(raster)
x <- data.table(pixelGroup = rep(1:2, each = 2), B = c(100, 200, 20, 400),
                speciesCode = rep(c("Pice_Gla", "Popu_Tre"), 2))
pixelGroupMap <- raster(extent(0,3, 0, 3), res = 1)
pixelGroupMap[] <- sample(1:2, size = 9, replace = TRUE)
vtm <- vegTypeMapGenerator(x, pixelGroupMap = pixelGroupMap)
}
\author{
Eliot McIntire, Ceres Barros, Alex Chubaty
}
