% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedDispersalLANDIS.R
\docType{methods}
\name{WardFast}
\alias{WardFast}
\title{WardFast Seed Dispersal kernel}
\format{
An object of class \code{expression} of length 1.
}
\usage{
WardFast
}
\description{
A probability distribution used in LANDIS-II.
}
\author{
Eliot McIntire
}
\keyword{datasets}
